% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internals.R
\name{ac_filter}
\alias{ac_filter}
\alias{.acs_filter_by_land}
\alias{.acs_filter_by_container}
\title{AC* helper: AC* filters}
\usage{
.acs_filter_by_land(.particles, .bathy)

.acs_filter_by_container(.particles, .moorings, .receivers, .threshold)
}
\arguments{
\item{.particles}{A \code{\link{data.table}} that defines particle locations. This should contain two columns named \code{x_now} and \code{y_now}.}

\item{.bathy}{The bathymetry \code{\link{SpatRaster}}. \code{NAs} are taken to define inhospitable habitats (e.g., on land).}

\item{.moorings}{A \code{\link{data.table}} of receiver IDs and locations.}

\item{.receivers}{An \code{integer} vector of receiver(s) at which the individual was next detected, used to filter \code{.moorings}.}

\item{.threshold}{A \code{numeric} value that defines the distance from \code{.receivers} within which valid particles are located.}
}
\value{
The functions return the \code{.particles} \code{\link{data.table}} with the subset of particles that passed the filtering criteria.
}
\description{
These functions filter particles according to AC* criteria in \code{\link[=pf_forward_2]{pf_forward_2()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
