% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac-classes.R
\docType{package}
\name{ac_record-class}
\alias{ac_record-class}
\title{AC outputs: \code{\linkS4class{ac_record}} objects}
\description{
An S3 class that defines the named \code{list} returned by a AC-branch algorithm.
}
\author{
Edward Lavender
}
