% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-classes.R
\docType{package}
\name{acs-class}
\alias{acs-class}
\title{AC outputs: \code{\linkS4class{acs}} objects}
\description{
An S3 class that defines the named \code{list} returned by a AC-branch algorithm.
}
\author{
Edward Lavender
}
