% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internal.R
\name{acs_container}
\alias{acs_container}
\alias{.acs_container}
\alias{.acs_container_init}
\title{AC* helper: AC* container(s)}
\usage{
.acs_container(.cinfo)

.acs_container_init(.obs, .dlist)
}
\value{
The functions return a \code{\link{SpatVector}}.
}
\description{
These functions define acoustic containers for ACPF algorithm(s).
}
\details{
At the first time step, the acoustic container is the intersection between the container(s) around the receivers that recorded the first detection (if applicable) and the container(s) around the receivers that recorded the next detection.
}
\author{
Edward Lavender
}
\keyword{internal}
