% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_detection_containers}
\alias{acs_setup_detection_containers}
\title{Set up detection containers}
\usage{
acs_setup_detection_containers(.grid, .moorings)
}
\arguments{
\item{.grid}{A \code{\link{SpatRaster}}.}

\item{.moorings}{A \code{\link{data.table}}}
}
\value{
The function returns a named list.
}
\description{
This function defines receiver detection containers.
}
\details{
Details
}
\examples{
# Define grid
grid <- dat_gebco()
terra::plot(grid)
# Define receiver detection ranges
dat_moorings$receiver_range <- 500
# Define detection containers
containers <- acs_setup_detection_containers(grid, dat_moorings)
# Visualise an example container
terra::plot(containers[[dat_moorings$receiver_id[1]]])
points(dat_moorings$receiver_easting[1], dat_moorings$receiver_northing[1])

}
\author{
Edward Lavender
}
