% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_detection_kernel}
\alias{acs_setup_detection_kernel}
\title{AC* set up: detection probability}
\usage{
acs_setup_detection_kernel(
  .mooring,
  .bathy,
  .mask = TRUE,
  .pdetx = pdetlogistic,
  ...
)
}
\arguments{
\item{.mooring}{A one-row \code{\link{data.table}} that defines the location of an acoustic receiver and associated information required to calculate detection probability via \code{.pdetx}. In this function, receiver coordinate columns (\code{receiver_x} and \code{receiver_y}) and the detection range (\code{receiver_range}) is required.}

\item{.bathy}{A \code{\link{SpatRaster}} that defines the grid on which detection probability is calculated.}

\item{.mask}{A \code{logical} variable that defines whether or not to mask the detection kernel by \code{.bathy}. Use \code{.mask = TRUE} if \code{.bathy} contains \code{NA}s.}

\item{.pdetx}{A function that calculates detection probability, such as \code{\link[=pdetlogistic]{pdetlogistic()}}. In this implementation, the function is used to translate a \code{\link{SpatRaster}} of distances (m) (from each grid cell to the receiver in \code{.mooring}) via \code{\link[terra:app]{terra::app()}}. The function must accept a .\code{gamma} argument (even if this is ignored, see below).}

\item{...}{Additional arguments passed to \code{.pdetx}. These arguments are passed to \code{\link[=pdetlogistic]{pdetlogistic()}}  by default. \code{.gamma} is set internally to \code{.mooring$receiver_range}.}
}
\value{
The function returns a \code{\link{SpatRaster}} that defines the detection kernel around a specific receiver.
}
\description{
This function is part of a set of \verb{acs_setup_*()} functions that prepare the layers required to evaluate the likelihood of acoustic observations in an AC*PF algorithm using the inbuilt functions. The function is an example detection probability function, of the kind required by \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}.
}
\details{
An AC*PF algorithm is a particle filtering algorithm that incorporates acoustic observations to reconstruct the possible movements of an individual (see \code{\link[=pf_forward]{pf_forward()}}). At each time step in such an algorithm, we evaluate the likelihood of acoustic observations (the presence or absence of detections at each operational receiver, accounting for receiver placement) given particle samples. The likelihood of the acoustic observations depends upon how detection probability declines away from receiver(s) in space; i.e., the shape of a 'detection kernel' (see \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}). For any one receiver, the form of the kernel depends on the input to the \code{.pdetkernel} in \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}. This function exemplifies one possible input to this argument, which is a model in which detection probability declines logistically with distance from a receiver.
}
\section{Warning}{
\itemize{
\item The default settings for this function are used to streamline examples and do not represent a generically suitable model.
\item The function does not check user inputs.
}
}

\examples{
# Set up data list(see `?pat_setup_data()`)
dlist <- dat_dlist()

# Define function inputs
m <- dlist$data$moorings[1, ]
b <- dlist$spatial$bathy

# Generate a detection kernel
k <- acs_setup_detection_kernel(m, b)
terra::plot(k)
points(m$receiver_x, m$receiver_y, pch = ".")

}
\seealso{
To implement an AC*PF algorithm, see:
\enumerate{
\item \code{\link[=pat_setup_data]{pat_setup_data()}} to set up datasets;
\item \verb{acs_setup_*()} functions to prepare layers required for likelihood calculations, i.e.:
\itemize{
\item \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}}, which identifies detection overlaps;
\item \code{\link[=acs_setup_detection_kernel]{acs_setup_detection_kernel()}}, which prepares a detection kernel;
\item \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}, which prepares detection kernels;
}
\item \code{\link[=pf_setup_obs]{pf_setup_obs()}} to define a timeline of observations;
\item \code{\link[=pf_forward]{pf_forward()}} to run the simulation (using \code{\link[=pf_lik_ac]{pf_lik_ac()}} to evaluate the likelihood of acoustic observations);
}
}
\author{
Edward Lavender
}
