% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internal.R
\name{acs_setup_detection_kernels-internal}
\alias{acs_setup_detection_kernels-internal}
\alias{.acs_setup_detection_kernels_pk1}
\alias{.acs_setup_detection_kernels_pk0}
\alias{.acs_setup_detection_kernels_ll}
\title{AC* helper: AC* kernels}
\usage{
.acs_setup_detection_kernels_pk1(.dlist, .pdetkernel, ...)

.acs_setup_detection_kernels_pk0(.pk1)

.acs_setup_detection_kernels_ll(.dlist, .pk0)
}
\description{
These functions support \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}.
}
\keyword{internal}
