% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_detection_overlaps}
\alias{acs_setup_detection_overlaps}
\title{Set up detection container overlaps}
\usage{
acs_setup_detection_overlaps(.containers, .moorings, .services = NULL)
}
\arguments{
\item{.containers}{A named list}

\item{.moorings}{A \code{\link{data.table}}}

\item{.services}{A \code{\link{data.table}}}
}
\value{
The function returns a named list.
}
\description{
Set up detection container overlaps
}
\details{
Details.
}
\author{
Edward Lavender
}
