% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_detection_pr}
\alias{acs_setup_detection_pr}
\title{AC* set up: calculate detection probability around a receiver}
\usage{
acs_setup_detection_pr(
  .data,
  .bathy,
  .coef = c(alpha = 2.5, beta = -0.02),
  ...
)
}
\arguments{
\item{.data}{A one-row \code{\link{data.table}} that defines the location of the receiver and associated information used by the model of detection probability.}

\item{.bathy}{A \code{\link{SpatRaster}} that defines the grid over which detection probability is calculated.}

\item{.coef}{A named vector that defines the intercept (\code{alpha}) and gradient (\code{beta}) parameters of a logistic model for detection probability (on the logit scale) (see Details).}

\item{...}{Additional arguments (none implemented).}
}
\value{
The function returns a \code{\link{SpatRaster}} that defines the probability of detection in each cell, according to a pre-defined model.
}
\description{
This function is an example detection probability function, of the kind required by \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}.
}
\details{
In the AC* algorithms, a model of the detection process informs the set of possible locations for an individual. The information provided by this model is represented in the form of kernels, which are created via \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}. For any one receiver, the form of the kernel depends on the input to \code{.calc_detection_pr}. This function exemplifies one possible input to this argument, which is a model in which detection probability declines logistically with distance from a receiver, according to the values specified in \code{.coef}.
}
\section{Warning}{
\itemize{
\item This function is used to streamline examples and does not represent a generically suitable detection probability model.
\item The function does not check user inputs.
}
}

\examples{
m <- dat_moorings[1, , drop = FALSE]
p <- acs_setup_detection_pr(m, dat_gebco())
terra::plot(p)
points(m$receiver_easting, m$receiver_northing, pch = ".")

}
\author{
Edward Lavender
}
