% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_obs}
\alias{acs_setup_obs}
\title{Set up movement datasets}
\usage{
acs_setup_obs(.acoustics, .archival = NULL, .step, .mobility)
}
\arguments{
\item{.acoustics}{A \code{\link{data.table}}}

\item{.archival}{A \code{\link{data.table}}}

\item{.step}{An character}

\item{.mobility}{A number that defines ...}
}
\value{
The function returns a \code{\link{data.table}}.
}
\description{
TO DO
}
\author{
Edward Lavender
}
