% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers.R
\name{as.im.SpatRaster}
\alias{as.im.SpatRaster}
\title{Convert a \code{SpatRaster} to a \code{spatstat} \code{im}}
\source{
This function is based on \code{maptools::as.im.RasterLayer}.
}
\usage{
as.im.SpatRaster(from)
}
\description{
Convert a \code{SpatRaster} to a \code{spatstat} \code{im}
}
\keyword{internal}
