% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-internal.R
\name{cang_planar}
\alias{cang_planar}
\title{Simulation: internal distribution functions}
\usage{
cang_planar(.xy0, .xy1, .convention = c("180", "360"))
}
\description{
These are internal distribution function helpers.
}
\author{
Edward Lavender
}
\keyword{internal}
