% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_dir_exists}
\alias{check_dir_exists}
\title{Check a directory exists}
\usage{
check_dir_exists(x, type = c("abort", "warn"))
}
\arguments{
\item{x}{A string that defines a directory.}

\item{type}{A character that defines whether or not to throw a warning or an error.}
}
\value{
If the directory \code{x} does not exist, the function returns a warning or error; otherwise \code{x} is invisibly returned.
}
\description{
This function checks that a directory exists.
}
\examples{
\dontrun{
check_dir_exists(tempdir()) # works
check_dir_exists(file.path(tempdir(), "blah")) # error
}
}
\author{
Edward Lavender
}
\keyword{internal}
