% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_inherits}
\alias{check_inherits}
\title{Check inherited class(es)}
\usage{
check_inherits(x, class)
}
\arguments{
\item{x}{An object.}

\item{class}{The required class(es).}
}
\value{
The function checks whether \code{x} inherits \code{class} and, if not, returns an error.
}
\description{
This function checks inherited class(es).
}
\keyword{internal}
