% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal-check.R
\name{check_utils}
\alias{check_utils}
\alias{check_dir_exists}
\alias{check_dir_empty}
\alias{check_dir_contents_ext}
\alias{check_inherits}
\alias{check_names}
\alias{check_named_list}
\alias{check_tz}
\alias{check_new_colnames}
\alias{check_not_null}
\title{Utilities: check functions}
\usage{
check_dir_exists(input)

check_dir_empty(input, action = abort)

check_dir_contents_ext(input, ext, ...)

check_inherits(input, class)

check_names(
  input,
  req,
  extract_names = names,
  type = all,
  action = abort,
  arg = deparse(substitute(input))
)

check_named_list(input, ignore_empty = TRUE, arg = deparse(substitute(input)))

check_tz(input, arg = deparse(substitute(input)))

check_new_colnames(.data, .new)

check_not_null(input, req)
}
\description{
These internal functions perform generic checks on user inputs.
}
\author{
Edward Lavender
}
\keyword{internal}
