% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{cleanup}
\alias{cleanup}
\title{Utilities: clean up}
\usage{
cleanup(x)
}
\arguments{
\item{x}{A \code{character} vector (see \code{\link[=unlink]{unlink()}}).}
}
\description{
This function deletes temporary files and or directories recursively.
}
\details{
This function is a simple wrapper for \code{\link[=unlink]{unlink()}} that is used at the end of examples.
}
\author{
Edward Lavender
}
