% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coa.R
\name{coa}
\alias{coa}
\title{COA: centres of activity}
\usage{
coa(
  .map,
  .detections,
  .moorings = NULL,
  .delta_t,
  .split = NULL,
  .plot_weights = TRUE,
  ...,
  .one_page = TRUE
)
}
\arguments{
\item{.map}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} that defines the study area (see \code{\link{glossary}}). Here, \code{.map} is used to:
\itemize{
\item Extract \code{map_value} at centres of activity, for consistency with other routines (such as \code{\link[=pf_filter]{pf_filter()}});
}}

\item{.detections, .moorings}{Acoustic detection \href{s}{\code{data.table::data.table}}.
\itemize{
\item \code{.detections} is a \code{\link[data.table:data.table]{data.table::data.table}} of acoustic detections, with the following columns: \code{receiver_id} (or \code{sensor_id}), \code{timestamp} and (optionally) \code{receiver_x} and \code{receiver_y} columns;
\item (optional) \code{.moorings} is a \code{\link[data.table:data.table]{data.table::data.table}} of receiver coordinates, which should be provided if unavailable in \code{.detections}, with \code{receiver_id}, \code{receiver_x} and \code{receiver_y} columns;
}

Receiver coordinates \strong{must be planar}.}

\item{.delta_t}{The time interval over which to calculate COAs. This can be specified in any way understood by \code{\link[lubridate:round_date]{lubridate::floor_date()}} (see the \code{unit} argument).}

\item{.split}{(optional) A \code{character} that defines the name of the grouping factor in \code{.detections} (e.g., \code{individual_id} for \code{\link{dat_detections}}).}

\item{.plot_weights, ..., .one_page}{Plot arguments.
\itemize{
\item \code{.plot_weights} is a \code{logical} variable that defines whether or not to plot the frequency distribution of weights for each \code{.split} value (i.e., the frequency distribution of the number of detections at each receiver in each time interval, excluding time intervals without detections).
\item \code{...} is a placeholder for arguments passed to \code{\link[graphics:hist]{graphics::hist()}}, excluding \code{main}.
\item \code{.one_page} A \code{logical} variable that defines whether or not to plot all histograms on one page.
}}
}
\value{
The function returns a \code{\link[data.table:data.table]{data.table::data.table}} with the following columns:
\itemize{
\item \code{{.split}}---a \code{character} vector that distinguishes groups, if applicable;
\item \code{timestep}---an \code{integer} vector of time steps;
\item \code{timestamp}---a \code{POSIXt} vector of time stamps;
\item \code{map_value}, \code{x}, \code{y}---the value of \code{.map} at COAs and their coordinates;
}

Data are arranged by \code{.split} and \code{timestamp}.
}
\description{
This function calculates centres of activity (COAs) from detections at acoustic receivers.
}
\details{
COAs are calculated as a weighted mean of the locations of receivers at which individuals are detected over consecutive time intervals, weighted by the frequency of detections at each of those receivers. COAs are calculated via \code{\link[stats:weighted.mean]{stats::weighted.mean()}}, which assumes planar coordinates, for consistency with other \code{\link{patter}} routines. To handle longitude/latitude coordinates, use \code{geosphere::geomean()}.

This function replaces \href{https://edwardlavender.github.io/flapper/reference/coa.html}{\code{flapper::coa()}}. See  \href{https://edwardlavender.github.io/flapper/reference/coa_setup_delta_t.html}{\code{flapper::coa_setup_delta_t()}} to evaluate alternative time internals.

On Linux with \code{JuliaCall}, this function generally cannot be used within a \code{Julia} session.
}
\examples{
if (patter_run(.julia = FALSE, .geospatial = TRUE)) {

  library(data.table)

  #### Set up example
  map        <- dat_gebco()
  detections <- dat_detections
  moorings   <- dat_moorings

  #### Example (1): Calculate COAs for an example individual
  id     <- dat_detections$individual_id[1]
  det    <- dat_detections[individual_id == id, ]
  coa(.map = map, .detections = det, .moorings = moorings,
      .delta_t = "2 hours")

  #### Example (2): Customise the time interval via `.delta_t`
  coa(.map = map, .detections = det, .moorings = moorings,
      .delta_t = "4 hours")

  #### Example (3): Calculate COAs for multiple individuals via `.split`
  # Calculate COAs
  coa(.map = map, .detections = det, .moorings = moorings,
      .delta_t = "6 hours", .split = "individual_id")
  # Use one page for plots via `.one_page = TRUE`
  coa(.map = map, .detections = det, .moorings = moorings,
      .delta_t = "6 hours", .split = "individual_id",
      .one_page = FALSE)
  # Suppress plots via `.plot = FALSE`
  coa(.map = map, .detections = det, .moorings = moorings,
      .delta_t = "6 hours", .split = "individual_id",
      .one_page = TRUE)

  #### Example (4): Specify `.detections` only
  # `.moorings` is not required if `.detections` contains receiver coordinates
  coa(.map = map,
      .detections =
        detections |>
        merge(moorings, by = "receiver_id"),
      .delta_t = "6 hours", .split = "individual_id")

}
}
\seealso{
\itemize{
\item To derive location samples from a particle filtering algorithm, see \code{\link[=pf_filter]{pf_filter()}} and associates;
\item For mapping utilisation distributions from coordinates, see \verb{map_*()} functions (i.e., \code{\link[=map_pou]{map_pou()}} and \code{\link[=map_dens]{map_dens()}});
}
}
\author{
Edward Lavender
}
