% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coa-internals.R
\name{coa_check}
\alias{coa_check}
\alias{.coa_check_acoustics}
\title{COA internals}
\usage{
.coa_check_acoustics(.acoustics, .split)
}
\description{
COA internals
}
\author{
Edward Lavender
}
\keyword{internal}
