% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colProds.matrix}
\alias{colProds.matrix}
\title{Calculate column products}
\usage{
colProds.matrix(.data, .cols = seq_len(ncol(.data)))
}
\arguments{
\item{.data}{A \code{\link{matrix}}.}

\item{.cols}{A vector that defines the columns in \code{.data} to multiply.}
}
\value{
The function returns a \code{numeric} vector.
}
\description{
This function calculates column products for each row in a \code{\link{matrix}}.
}
\author{
Edward Lavender
}
\keyword{internal}
