% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_log}
\alias{create_log}
\title{Create a log file}
\usage{
create_log(.file)
}
\arguments{
\item{.file}{A character path to a file. \code{NULL} is permitted.}
}
\value{
The function returns \code{invisible(TRUE)}.
}
\description{
This function creates a .txt file (\code{.file}) via \code{\link{file.create}}.
}
\author{
Edward Lavender
}
\keyword{internal}
