% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_acoustics}
\alias{dat_acoustics}
\title{Passive acoustic telemetry detections}
\format{
A dataframe with 39,242 observations and 9 variables:
\describe{
\itemize{
\item \code{individual_id}---a unique identifier of the individual that was detected.
\item \code{timestamp}---a \code{POSIXct} object that defines the time of each observation.
\item \code{receiver_id}---a unique identifier of the receiver at which the individual was detected (see \code{\link{dat_moorings}}).
}
}
}
\source{
Data were collected by, and belong to, Marine Scotland Science and NatureScot. Data were processed by Edward Lavender. If you wish to use these data, please contact Marine Scotland Science and NatureScot for further information.
}
\usage{
dat_acoustics
}
\description{
A dataset containing a sample of processed flapper skate (\emph{Dipturus intermedius}) detection time series. Data are arranged by \code{individual_id}, \code{timestamp} and then \code{receiver_id} (see below).
}
\references{
Data collection and processing are described in Lavender (2022). Modelling the movements of flapper skate (\emph{Dipturus intermedius}) in relation to a Scottish Marine Protected Area. University of St Andrews. https://www.doi.org/10.17630/sta/201
}
\keyword{datasets}
