% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_archival}
\alias{dat_archival}
\title{Dataset: archival time series}
\format{
A \code{\link{data.table}} with 75,000 observations and 3 variables:
\itemize{
\item \code{individual_id}---a number that defines each individual.
\item \code{timestamp}---a \code{POSIXct} object that defines the time of each observation.
\item \code{depth}---a number that defines the depth (m) of the individual at each time step.
}
}
\source{
Data were collected by, and belong to, Marine Scotland Science and NatureScot. Data were processed by Edward Lavender. If you wish to use these data, please contact Marine Scotland Science and NatureScot for further information.
}
\usage{
dat_archival
}
\description{
A dataset containing a sample of flapper skate (\emph{Dipturus intermedius}) depth (m) time series. Observations were sampled every 2 minutes using archival tags. Data are arranged by \code{individual_id} and then \code{timestamp}.
}
\references{
Data collection and processing are described in Lavender (2022). Modelling the movements of flapper skate (\emph{Dipturus intermedius}) in relation to a Scottish Marine Protected Area. University of St Andrews. https://www.doi.org/10.17630/sta/201
}
\keyword{datasets}
