% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_moorings}
\alias{dat_moorings}
\title{Dataset: Passive acoustic telemetry moorings}
\format{
A \code{\link{data.table}} with 40 observations and 7 variables:
\itemize{
\item \code{receiver_id}---an integer that distinguishes each unique receiver deployment.
\item \code{receiver_start}---a date that defines the start date of each receiver's deployment.
\item \code{receiver_end}---a date that defines the end date of each receiver's deployment.
\item \code{receiver_lon}---a number that defines the longitude (decimal degrees) of each receiver.
\item \code{receiver_lat}---a number that defines the latitude (decimal degrees) of each receiver.
\item \code{reciver_easting}---a number that defines the receiver easting (UTM 29N).
\item \code{receiver_northing}---a number that defines the receiver northing (UTM 29N).
\item \code{receiver_depth}---a number that defines the approximate depth of each receiver below the surface (m).
}
}
\source{
Data were collected by, and belong to, Marine Scotland Science and NatureScot. Data were processed by Edward Lavender. If you wish to use these data, please contact Marine Scotland Science and NatureScot for further information.
}
\usage{
dat_moorings
}
\description{
A dataset with receiver locations and associated information. Data are arranged by \code{receiver_id} (see below).
}
\references{
Data collection and processing are described in Lavender (2022). Modelling the movements of flapper skate (\emph{Dipturus intermedius}) in relation to a Scottish Marine Protected Area. University of St Andrews. https://www.doi.org/10.17630/sta/201
}
\keyword{datasets}
