% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats-patter.R
\name{datasets-algorithms}
\alias{datasets-algorithms}
\alias{dat_path}
\alias{dat_coa}
\alias{dat_pff}
\title{Datasets: example algorithm outputs}
\source{
For full details on the algorithm parameters used to generate these datasets, see \url{https://github.com/edwardlavender/patter/blob/main/data-raw/003-add-data-algorithms.R}.
}
\usage{
dat_path()

dat_coa()

dat_pff()
}
\value{
The functions return a dataset or a \code{character} string that defines the path to a dataset. See the corresponding function documentation for full details.
}
\description{
These functions load example outputs from key \code{\link{patter}} functions. They are included in the package to streamline function examples and tests.
}
\details{
\itemize{
\item \code{\link[=dat_path]{dat_path()}} reads an example output from \code{\link[=sim_path_walk]{sim_path_walk()}};
\item \code{\link[=dat_coa]{dat_coa()}} reads an example output from \code{\link[=coa]{coa()}};
\item \code{\link[=dat_pff]{dat_pff()}} reads an example output from \code{\link[=pf_filter]{pf_filter()}};
}
}
\examples{
# Load inbuilt datasets
dat_path() |> summary()
dat_coa() |> summary()
dat_pff() |> summary()

}
\seealso{
Particle filters and smoothers sample states (particles) that represent the possible locations of an individual through time, accounting for all data and the individual's movement.
\itemize{
\item To simulate artificial datasets, see \verb{sim_*()} functions (especially \code{\link[=sim_path_walk]{sim_path_walk()}}, \code{\link[=sim_array]{sim_array()}} and \code{\link[=sim_observations]{sim_observations()}}).
\item To assemble real-world datasets for the filter, see \code{\link{assemble}}\verb{_*()} functions.
\item \code{\link[=pf_filter]{pf_filter()}} runs the filter:
\itemize{
\item For state types, see \code{\link{State}};
\item For observation models, see \code{\link{ModelObs}};
\item For movement models, see \code{\link{ModelMove}};
}
\item To run particle smoothing, use \code{\link[=pf_smoother_two_filter]{pf_smoother_two_filter()}}.
\item To map emergent patterns of space use, use a \verb{map_*()} function (such as \code{\link[=map_pou]{map_pou()}}, \code{\link[=map_dens]{map_dens()}} and \code{\link[=map_hr]{map_hr()}}).
}
}
\author{
Edward Lavender
}
