% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats-mefs.R
\docType{data}
\name{datasets-mefs}
\alias{datasets-mefs}
\alias{dat_moorings}
\alias{dat_detections}
\alias{dat_archival}
\alias{dat_gebco}
\alias{dat_coast}
\alias{dat_mpa}
\title{Datasets: the MEFS project}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 40 rows and 8 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 39242 rows and 3 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 75000 rows and 3 columns.
}
\source{
Biologging and biotelemetry data were collected by, and belong to, Marine Scotland Science and NatureScot. Data were processed by Lavender (2022). If you wish to use these data, please contact Marine Scotland Science and NatureScot for further information.

Bathymetry data were sourced from GEBCO Compilation Group (2019) GEBCO 2019 Grid. \url{https://www.doi.org/10.5285/836f016a-33be-6ddc-e053-6c86abc0788e}

Coastline data were sourced from \href{https://edwardlavender.github.io/flapper/reference/dat_coast.html}{\code{flapper::dat_coast}}.

MPA boundary data were assembled by E. Lavender.
}
\usage{
dat_moorings

dat_detections

dat_archival

dat_gebco(.return = c("SpatRaster", "character"))

dat_coast(.return = c("SpatVector", "character"))

dat_mpa(.return = c("SpatVector", "character"))
}
\arguments{
\item{.return}{A \code{character} that defines the object returned by \code{\link[=dat_gebco]{dat_gebco()}}, \code{\link[=dat_coast]{dat_coast()}} or \code{\link[=dat_mpa]{dat_mpa()}}:
\itemize{
\item \code{SpatRaster} returns a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}};
\item \code{SpatVector} returns a \code{\link[terra:SpatVector-class]{terra::SpatVector}};
\item \code{character} returns a file path to the raster. This is the best option on Linux with \code{JuliaCall} if \code{JULIA_SESSION = "TRUE"};
}}
}
\description{
These are sample datasets collated by the Movement Ecology of Flapper Skate (MEFS) project (Lavender, 2022). The flapper skate (\emph{Dipturus intermedius}) is a Critically Endangered benthic elasmobranch. As part of the MEFS project, flapper skate were tagged with acoustic transmitters and archival (data storage) tags off the west coast of Scotland in 2016--17. Acoustic transmissions were recorded at an array of passive acoustic telemetry receivers. Concurrent depth time series were recorded by archival tags and recovered from recaptured individuals. For full details, see the references below. The datasets are defined within \code{\link{patter}} to illustrate package functions using real-world datasets.
}
\section{Moorings}{
\code{\link{dat_moorings}} defines passive acoustic telemetry receiver locations and associated information. This includes the following columns:
\itemize{
\item \code{receiver_id}---an \code{integer} vector that defines unique receiver deployments;
\item \code{receiver_start}, \code{receiver_end}---\code{Date} vectors that define receiver deployment start and end dates;
\item \code{receiver_x}, \code{receiver_y}---\code{numeric} vectors that define receiver coordinates (in UTM 29N);
\item \code{receiver_alpha}, \code{receiver_beta}, \code{receiver_gamma}---a \code{numeric} vectors that define detection probability parameters;
}

Data are arranged by \code{receiver_id}.
}

\section{Detections}{
\code{\link{dat_detections}} contains sample detection time series. This includes the following columns:
\itemize{
\item \code{individual_id}---an \code{integer} vector that identifies individuals;
\item \code{timestamp}---a \code{POSIXct} vector that defines the time of each observation;
\item \code{receiver_id}---the receiver ID (see \code{\link{dat_moorings}});
}

Data are arranged by \code{individual_id}, \code{timestamp} and then \code{receiver_id}.
}

\section{Archival}{
\code{\link{dat_archival}} contains sample depth time series. Observations were sampled every 2 minutes. The data includes the following columns:
\itemize{
\item \code{individual_id}---an \code{integer} vector that identifies individuals (as in \code{\link{dat_detections}});
\item \code{timestamp}---a \code{POSIXct} vector that defines the time of each observation;
\item \code{depth}---a \code{numeric} vector that defines the depth (m) of the individual at each time step;
}

Data are arranged by \code{individual_id} and then \code{timestamp}.
}

\section{Bathymetry}{
\code{\link[=dat_gebco]{dat_gebco()}} returns a bathymetry (m) dataset for the west coast of Scotland where MEFS data were collected. This dataset is a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} with the following properties:
\itemize{
\item \code{dimensions}---264, 190, 1 (\code{nrow}, \code{ncol}, \code{nlyr});
\item \code{resolution}---100, 100  (\code{x}, \code{y});
\item \code{extent}---695492.1, 714492.1, 6246657, 6273057  (\code{xmin}, \code{xmax}, \code{ymin}, \code{ymax});
\item \verb{coord. ref.}---WGS 84 / UTM zone 29N (EPSG:32629);
}
}

\section{Coastline}{
\code{\link[=dat_coast]{dat_coast()}} returns a \code{\link[terra:SpatVector-class]{terra::SpatVector}} of coastline for the same region.
}

\section{Marine Protected Area}{
\code{\link[=dat_mpa]{dat_mpa()}} returns a \code{\link[terra:SpatVector-class]{terra::SpatVector}} of the boundaries of the Loch Sunart to the Sound of Jura Marine Protected Area (MPA) with the following columns:
\itemize{
\item \code{id}---A \code{character} label for each polygon;
\item \code{open}---A \code{character} that defines whether or not each \code{id} is \code{open} or \code{closed} to fishing;
\item \code{col}---A \code{character} that defines a colour (for visualising);
}
}

\references{
Data collection and processing are described in Lavender (2022). Modelling the movements of flapper skate (\emph{Dipturus intermedius}) in relation to a Scottish Marine Protected Area. University of St Andrews. \url{https://www.doi.org/10.17630/sta/201}

For further information on the MEFS project, see:
\itemize{
\item Lavender, E. et al. (2021). Movement patterns of a Critically Endangered elasmobranch (\emph{Dipturus intermedius}) in a Marine Protected Area. Aquat. Conserv. 32: 348–65. \url{https://www.doi.org/10.1002/aqc.3753}
\item Lavender, E. et al. (2021). Individual variation and environmental cycles in the vertical movements of a benthic elasmobranch. Mar. Biol. 168: 164. \url{https://www.doi.org/10.1007/s00227-021-03973-1}
\item Lavender, E. et al. (2022). Behavioural responses of a large, benthic elasmobranch to catch-and-release angling. Front. Mar. Sci. 9: 864344. \url{https://www.doi.org/10.3389/fmars.2022.864344}
\item Thorburn, J. et al. (2021). Seasonal and ontogenetic variation in depth use by a Critically Endangered benthic elasmobranch and its implications for spatial management. Front. Mar. Sci. 8: 829.\url{https://www.doi.org/10.3389/fmars.2021.656368}
\item Lavender, E. et al. (2022). Benthic animal-borne sensors and citizen science combine to validate ocean modelling. Sci. Rep. 12: 16613. \url{https://www.doi.org/1038/s41598-022-20254-z}
\item Lavender, E. et al. (2023). An integrative modelling framework for passive acoustic telemetry. Meth. Ecol. Evol: 14, 2626–2638. \url{https://doi.org/10.1111/2041-210X.14193}
}
}
\keyword{datasets}
