% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc}
\alias{dc}
\title{DC: the depth-contour algorithm}
\usage{
dc(
  .obs,
  .bathy,
  .model,
  ...,
  .save_record = FALSE,
  .write_record = NULL,
  .cl = NULL,
  .varlist = NULL,
  .progress = TRUE,
  .verbose = FALSE,
  .con = ""
)
}
\arguments{
\item{.obs}{A \code{\link{data.table}} that defines (sorted) depth time series for a selected individual. At a minimum, this must contain the following columns:
\itemize{
\item \code{timestep}---an \code{integer} that defines the time step;
\item Any columns required by \code{.model}
}}

\item{.bathy}{A \code{\link{SpatRaster}} that defines the bathymetry.}

\item{.model, ...}{A \code{function} that defines the possible locations of the individual on \code{.bathy} at a given time step. This must accept three inputs in the following order:
\itemize{
\item \code{.obs}---the \code{.obs} \code{\link{data.table}};
\item \code{.t}---an \code{integer} that defines the time step (used to index \code{.obs});
\item \code{.bathy}---the bathymetry \code{\link{SpatRaster}};
\item \code{...}---additional arguments, if necessary;
}

\code{model} should return a \code{\link{SpatRaster}}. Cell values are assumed to represent probabilities.}

\item{.save_record}{A logical variable that defines whether or not to save the output of \code{.model} at each time step in memory.}

\item{.write_record}{A named list, passed to \code{\link[terra:writeRaster]{terra::writeRaster}}, to save the \code{record} \code{\link{SpatRaster}}s to file at each time step. The \code{filename} argument should define the directory in which to write files (see \code{\link[=acs]{acs()}}).}

\item{.cl, .varlist}{Parallelisation options (see \code{\link[=cl_lapply]{cl_lapply()}}).}

\item{.progress, .verbose, .con}{Options to monitor function progress (see \code{\link[=acs]{acs()}})}
}
\value{
The function returns an \code{\linkS4class{acs}} object.
}
\description{
This function implements the depth contour algorithm.
}
\author{
Edward Lavender
}
