% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc_setup_model}
\alias{dc_setup_model}
\title{DC: a template DC model}
\usage{
dc_setup_model(.obs, .t, .bathy, ...)
}
\arguments{
\item{.obs}{The \code{.obs} \code{\link{data.table}} from \code{\link[=dc]{dc()}}.}

\item{.t}{An \code{integer} that defines the time step (used to index \code{.obs}).}

\item{.bathy}{The \code{.bathy} \code{\link{SpatRaster}} from \code{\link[=dc]{dc()}}.}

\item{...}{Additional arguments passed from \code{\link[=dc]{dc()}} (unused here).}
}
\value{
The function returns a normalised \code{\link{SpatRaster}}.
}
\description{
This function is an example DC model, of the kind required by \code{.model} in \code{\link[=dc]{dc()}}.
}
\details{
This model assumes the individual can be located anywhere between a shallow and deep depth limit (as defined by \code{depth_shallow} and \code{depth_deep} columns in \code{.obs}) with equal probability.
}
\section{Warning}{
\itemize{
\item This function is used to streamline examples and does not represent a generically suitable model.
\item The function does not check user inputs.
}
}

\examples{
# For examples, see `?dc()`

}
\author{
Edward lavender
}
