% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers.R
\name{degrees}
\alias{degrees}
\title{Spatial helper: create circular angle in degrees}
\usage{
degrees(.x)
}
\arguments{
\item{.x}{A numeric vector of angles (in degrees).}
}
\value{
The function returns a \code{\link[circular:circular]{circular::circular}} object.
}
\description{
This function is a wrapper for \code{circular::circular(x, units = "degrees")}.
}
\examples{
degrees(10)
degrees(c(10, 20))
}
\author{
Edward Lavender
}
