% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dist_2d}
\alias{dist_2d}
\title{Distances: line transects}
\usage{
dist_2d(.x0, .y0, .x1, .y1)
}
\arguments{
\item{.x0, .y0, .x1, .y1}{Coordinates.}
}
\description{
This function calculates the distance between coordinate pairs. For planar coordinates, it is faster than \code{terra::distance(..., lonlat = FALSE, pairwise = TRUE)}.
}
\details{
Planar coordinates are assumed.
}
\keyword{internal}
