% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dist_along_path}
\alias{dist_along_path}
\title{Distances: along a path}
\usage{
dist_along_path(.xy, .lonlat = FALSE)
}
\arguments{
\item{.xy}{A two-column matrix of coordinates.}

\item{.lonlat}{A \code{logical} variable that defines whether or not the coordinates are in longitude/latitude format.}
}
\description{
This function calculates distances between sequential coordinates in a matrix.
}
\details{
This function is a simple wrapper for \code{\link[terra:distance]{terra::distance()}}.
}
\seealso{
\itemize{
\item \verb{dist_*} functions in \href{https://github.com/edwardlavender/flapper}{\code{flapper}};
\item \code{\link[=dist_along_path]{dist_along_path()}} calculates distances along a path;
\item \code{\link[=dist_btw_cells]{dist_btw_cells()}} calculates distances between \code{\link{SpatRaster}} cells;
}
}
