% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dist_btw_cells}
\alias{dist_btw_cells}
\title{Distances: between cells on a \code{\link{SpatRaster}}}
\usage{
dist_btw_cells(.bathy, .from, .to, .lonlat = FALSE, ...)
}
\arguments{
\item{.bathy}{A \code{\link{SpatRaster}}.}

\item{.from, .to}{Integer vectors that define cell IDs.}

\item{.lonlat}{A \code{logical} variable that defines whether or not the \code{\link{.bathy}} is in longitude/latitude format.}

\item{...}{Arguments passed to  \code{\link[terra:distance]{terra::distance()}}}
}
\description{
This function calculates distances between cells on a \code{\link{SpatRaster}}
}
