% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internals.R
\name{.acs_container_1}
\alias{.acs_container_1}
\title{AC* helper: define AC* container(s)}
\usage{
.acs_container_1(.obs, .detection_kernels)
}
\arguments{
\item{.obs}{The \code{.obs} \code{\link{data.table}}.}

\item{.detection_kernels}{A \code{\link{list}} of detection kernels.}
}
\value{
The function returns a \code{\link{SpatRaster}} (the acoustic container).
}
\description{
This function defines the first AC* container in \code{\link[=pf_forward_2]{pf_forward_2()}}.
}
\details{
In \code{\link[=acs]{acs()}}, acoustic containers are defined using \code{\link[terra:buffer]{terra::buffer()}}.

This functions are used in \code{\link[=pf_forward_2]{pf_forward_2()}}.
\itemize{
\item \code{\link[=.acs_container_1]{.acs_container_1()}} defines the acoustic container at the first time step, which is the intersection between the container(s) around the receivers that recorded the first detection (if applicable) and the container(s) around the receivers that recorded the next detection.
}
}
\author{
Edward Lavender
}
\keyword{internal}
