% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internals.R
\name{.acs_given_detection}
\alias{.acs_given_detection}
\title{AC* helper: Define the individual's location given the data}
\usage{
.acs_given_detection(.detections, .absences, .kernels)
}
\arguments{
\item{.detections}{An integer vector of the receiver(s) that recorded detections at a given time step.}

\item{.absences}{An integer vector of the remaining, overlapping receiver(s) that did not record a detection, from \code{\link[=.acs_absences]{.acs_absences()}}.}

\item{.kernels}{A list from \code{\link{acs_setup_detection_kernels}}.}
}
\description{
This function defines an (unnormalised) probability surface that defines the possible locations of an individual given one or more detections.
}
\details{
In the AC* algorithms, at the moment of detection, the probability kernels that describe the possible locations of an individual given the data depend on both the receivers that record detections and those that did not (eqn S5 in Lavender et al., 2023). This function solves eqn S5. For computational efficency, the equation is solved in a stepwise manor such that the number of necessary operations is kept to a minimum.
}
\section{Warning}{
For speed, this function performs no internal checks.
}

\author{
Edward Lavender
}
\keyword{internal}
