% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internal.R
\name{.acs_given_detection_particles}
\alias{.acs_given_detection_particles}
\title{AC* helper: define the individual's location given detection(s)}
\usage{
.acs_given_detection_particles(.detections, .absences, .kernels, .particles)
}
\arguments{
\item{.detections}{An \code{integer} vector of the receiver(s) that recorded detections at a given time step.}

\item{.absences}{An \code{integer} vector of the remaining, overlapping receiver(s) that did not record a detection, from \code{\link[=.acs_absences]{.acs_absences()}}.}

\item{.kernels}{A \code{list} from \code{\link{acs_setup_detection_kernels}}.}

\item{.particles}{A \code{\link{data.table}} with a \code{cell_now} column that defines particle locations on the grid.}
}
\description{
These function defines the relative plausibility possible locations of an individual given one or more detections or for selected particle positions (for \code{\link[=pf_forward]{pf_forward()}}).
}
\details{
In the AC* algorithms, at the moment of detection, likelihood of the acoustic data given a particle sample depends on depends on both the receivers that record detections and those that did not (eqn S5 in Lavender et al., 2023). This function solves eqn S5.
}
\section{Warning}{
For speed, these functions performs no internal checks.
}

\author{
Edward Lavender
}
\keyword{internal}
