% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internal.R
\name{.acs_setup_obs_receiver_id_next}
\alias{.acs_setup_obs_receiver_id_next}
\title{AC* helper: define the receiver(s) at which the next detection was recorded}
\usage{
.acs_setup_obs_receiver_id_next(.receiver_id)
}
\arguments{
\item{.receiver_id}{A \code{list} column.}
}
\value{
The function returns a \code{list} column that defines, for each time step, the receiver(s) that recorded the next detection.
}
\description{
This function, for each time step, the receiver(s) at which the next detection was recorded.
}
\seealso{
The \code{receiver_id_next} column is required by \code{\link[=pf_forward_2]{pf_forward_2()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
