% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers.R
\name{.is_lonlat}
\alias{.is_lonlat}
\title{Spatial helper: identify lon/lat moorings}
\usage{
.is_lonlat(.data)
}
\arguments{
\item{.data}{A \code{\link{data.table}} with \code{receiver_easting} and \code{receiver_northing} or \code{receiver_lon} and \code{receiver_lat} columns (i.e., \code{\link{dat_moorings}}).}
}
\value{
A logical variable that defines whether or not to use longitude/latitude coordinates.
}
\description{
This function identifies moorings data in lon/lat format.
}
\author{
Edward Lavender
}
\keyword{internal}
