% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-internal.R
\name{.map_mark}
\alias{.map_mark}
\title{Map: location marks (weights)}
\usage{
.map_mark(.coord)
}
\arguments{
\item{.coord}{A \code{\link{data.table}} that defines coordinates, with the following columns:
\itemize{
\item (required) \code{id}---a vector that distinguishes coordinate IDs (such as grid cells or unique coordinate pairs);
\item (optional) \code{timestep}---a vector that distinguishes time steps, if coordinates have been sampled through time;
\item (optional) \code{mark}---a \code{numeric} vector of coordinate weights;
}}
}
\value{
The function returns a \code{\link{data.table}}. Columns match \code{.coord}, excluding \code{timestep} if specified (since weights are summarised within and across time steps). There is one row for each unique coordinate ID and a \code{mark} column that defines the associated weights.
}
\description{
This function calculates coordinate weights for a dataset or time series.
}
\details{
Equal weights are assumed unless specified in the \code{mark} column. Default or supplied weights are normalised to sum to one at each time step (if necessary). The total weight of each coordinate ID within time steps is calculated and then these weights are aggregated by coordinate ID across the whole time series (if applicable) and renomalised to sum to one.
}
\seealso{
This function is used by \code{\link[=.map_coord]{.map_coord()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
