% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward-internals.R
\name{.pf_backward_killer_outputs}
\alias{.pf_backward_killer_outputs}
\title{PF: backward pass internals}
\usage{
.pf_backward_killer_outputs(.start, .history, .record)
}
\description{
Internal functions for the backward pass.
}
\author{
Edward Lavender
}
\keyword{internal}
