% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward-internal.R
\name{.pf_backward_output}
\alias{.pf_backward_output}
\title{PF: backward pass internals}
\usage{
.pf_backward_output(.start, .history, .path = NULL, .record)
}
\description{
Internal functions for the backward pass.
}
\author{
Edward Lavender
}
\keyword{internal}
