% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-internal.R
\name{.pf_history_dt}
\alias{.pf_history_dt}
\title{PF: collate particle histories}
\usage{
.pf_history_dt(.history, ..., .collect = TRUE)
}
\arguments{
\item{.history}{The particle samples, supplied as:
\itemize{
\item A \code{\linkS4class{pf_particles}} object;
\item The \code{history} element of a \code{\linkS4class{pf_particles}} object;
\item A \code{character} string that defines the directory containing parquet files;
\item A \code{list} of file paths (e.g., from \code{\link[=pf_setup_files]{pf_setup_files()}});
}}

\item{...}{If \code{.history} is a directory, \code{...} additional arguments can be passed to \code{\link[arrow:open_dataset]{arrow::open_dataset()}} via \code{...}.}

\item{.collect}{If \code{.history} is a directory, \code{.collect} is a \code{logical} variable that defines whether or not to collect the dataset in memory.}
}
\value{
The function returns a \code{\link{data.table}} or an \code{\link[arrow:Dataset]{arrow::FileSystemDataset}} (if \code{.history} is a directory and \code{.collect = FALSE}).
}
\description{
This function collates particle histories into a single \code{\link{data.table}}.
}
\details{
\enumerate{
\item If \code{.history} is a \code{\linkS4class{pf_particles}} object, or the history element of such an object, \code{data.table::rbindlist(..., fill = TRUE)} is used to combine particle samples.
\item If \code{.history} is a folder, \code{\link[arrow:open_dataset]{arrow::open_dataset()}} plus (optionally) \code{\link[dplyr:compute]{dplyr::collect()}} is used. Additional arguments, such as \code{schema} can be passed to \code{\link[arrow:open_dataset]{arrow::open_dataset()}} via \code{...}.
\item If \code{.history} is a \code{list} of parquet files, files are iteratively read into memory (via \code{\link[arrow:read_parquet]{arrow::read_parquet()}}) and combined (via \code{data.table::rbindlist(..., fill = TRUE)}. This option is much slower than option 2.
}
}
\examples{
\dontrun{
require(arrow)

# Use pf_particles-class object
.pf_history_dt(dat_pfbk())

# Use history element
.pf_history_dt(dat_pfbk()$history)

# Use directory (and optionally select a subset of columns)
pfbk_folder <- system.file("extdata", "acpf", "backward", "killer",
                           package = "patter", mustWork = TRUE)
sch <- schema(timestep = int32(),
              cell_now = int32(),
              x_now = double(),
              y_now = double())
.pf_history_dt(pfbk_folder, schema = sch)

# Use directory with `.collect = FALSE`
.pf_history_dt(pfbk_folder, schema = sch, .collect = FALSE)

# Use a file list
.pf_history_dt(pf_setup_files(pfbk_folder))
}
}
\author{
Edward Lavender
}
\keyword{internal}
