% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-likelihood-internals.R
\name{.pf_lik}
\alias{.pf_lik}
\title{PF: Likelihood function internals}
\usage{
.pf_lik(
  .particles,
  .obs,
  .t,
  .dlist,
  .stack,
  .diagnostics = list(),
  .trial = NA_integer_
)
}
\arguments{
\item{.particles, .obs, .t, .dlist}{Arguments passed to likelihood functions (see \code{\link{pf_lik}}).}

\item{.stack}{A named \code{list} of likelihood functions (see \code{\link{pf_lik}}).}

\item{.diagnostics}{An empty \code{list}, used to store likelihood diagnostics, or \code{NULL}.}

\item{.trial}{An \code{integer} that distinguishes trials.}
}
\value{
A \code{\link{data.table}} that defines valid proposal locations, likelihoods and weights. A \code{diagnostics} attribute contains proposal diagnostics.
}
\description{
This function evaluates the likelihood of the data given location proposals in \code{\link[=pf_forward]{pf_forward()}}, wrapping specified likelihood functions (see \code{\link{pf_lik}}).
}
\author{
Edward Lavender
}
\keyword{internal}
