% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map-internal.R
\name{.pf_map_weights}
\alias{.pf_map_weights}
\title{PF: Calculate location weights}
\usage{
.pf_map_weights(.pxy)
}
\arguments{
\item{.pxy}{A \code{\link{data.table}} that contains locations. This should include a \code{cell_id} column and may include a \code{timestep} column.}
}
\value{
The function returns a \code{\link{data.table}}.
}
\description{
This function calculates location weights in \code{\link[=pf_map_pou]{pf_map_pou()}} and \code{\link[=pf_map_dens]{pf_map_dens()}}.
}
\details{
This function can be used to calculate weights for any set of coordinates (grid cells), including but not exclusively from particle filtering.
}
\author{
Edward Lavender
}
\keyword{internal}
