% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map-internals.R
\name{.pf_path_pivot_checks}
\alias{.pf_path_pivot_checks}
\title{PF: mapping}
\usage{
.pf_path_pivot_checks(.obs, .cols)
}
\description{
These functions support the reconstruction of movement paths
}
\keyword{internal}
