% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_setup-internal.R
\name{.pf_setup_obs_receiver_id_next}
\alias{.pf_setup_obs_receiver_id_next}
\title{PF: set up internals}
\usage{
.pf_setup_obs_receiver_id_next(.receiver_id)
}
\arguments{
\item{.receiver_id}{A \code{list} column.}
}
\value{
The function returns a \code{list} column that defines, for each time step, the receiver(s) that recorded the next detection.
}
\description{
\code{\link[=.pf_setup_obs_receiver_id_next]{.pf_setup_obs_receiver_id_next()}} defines, for each time step, the receiver(s) at which the next detection was recorded.
}
\seealso{
The \code{receiver_id_next} column is required by \code{\link[=pf_forward]{pf_forward()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
