% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward_2-utils-internal.R
\name{.pf_util_write}
\alias{.pf_util_write}
\title{PF helper: internal utilities}
\usage{
.pf_util_write(.particles, .diagnostics, .psink, .dsink, .write)
}
\description{
PF helper: internal utilities
}
\keyword{internal}
