% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_setup.R
\name{example_setup}
\alias{example_setup}
\alias{example_setup.default}
\alias{example_setup.pf_smoother_two_filter}
\title{Examples: streamline set up}
\usage{
example_setup(.fun)

\method{example_setup}{default}(.fun)

\method{example_setup}{pf_smoother_two_filter}(.fun)
}
\arguments{
\item{.fun}{A \code{character} that defines the name of a \code{\link{patter}} function with a corresponding \code{\link[=example_setup]{example_setup()}} method.}
}
\description{
These functions are used to streamline package examples.
}
\author{
Edward Lavender
}
