% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers.R
\name{geomean}
\alias{geomean}
\title{Calculate the centre of mass of weighted coordinates}
\usage{
geomean(xy, w)
}
\arguments{
\item{xy, w}{Arguments passed to \code{\link[geosphere:geomean]{geosphere::geomean()}}.}
}
\description{
This is a wrapper for \code{\link[geosphere:geomean]{geosphere::geomean()}} that handles one-row matrices.
}
\author{
Edward Lavender
}
\keyword{internal}
