% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hr.R
\name{get_hr}
\alias{get_hr}
\alias{get_hr_prop}
\alias{get_hr_core}
\alias{get_hr_home}
\alias{get_hr_full}
\title{Get animal 'home ranges'}
\usage{
get_hr_prop(.x, .prop = 0.5, .add = FALSE, ...)

get_hr_core(.x, .add = FALSE, ...)

get_hr_home(.x, .add = FALSE, ...)

get_hr_full(.x, .add = FALSE, ...)
}
\arguments{
\item{.x}{A \code{\link{SpatRaster}} (utilisation distribution).}

\item{.prop}{For \code{\link[=get_hr_prop]{get_hr_prop()}}, \code{.prop} is a number that defines the range proportion.}

\item{.add}{A logical variable that defines whether or not to add a polygon of the range to an existing map.}

\item{...}{If \code{.add = TRUE}, \code{...} is a place holder for additional arguments passed to \code{\link[terra:plot]{terra::plot()}}.}
}
\value{
The functions return a \code{\link{SpatRaster}}. Cells with a value of one are inside the specified range boundaries; cells with a value of zero are beyond range boundaries. If \code{.add} is \code{TRUE}, the boundaries are added to an existing plot.
}
\description{
These functions extract 'home range' estimates from a \code{\link{SpatRaster}} that describes the intensity of movements within an area.
}
\details{
These functions are modelled on the \href{https://edwardlavender.github.io/flapper/reference/get_hr.html}{\verb{get_hr_*()}} functions in the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package, where full details are provided. The \code{\link{spatialEco}} package is required.
}
\author{
Edward Lavender
}
