% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{glossary}
\alias{glossary}
\title{Glossary}
\description{
This is a glossary of key arguments in \code{\link{patter}}.
}
\section{\code{.map}}{
\code{.map} is \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} (supported on Windows or MacOS or, if using \code{JuliaConnectoR}, Linux) or a \code{character} that defines the path to raster (supported on Windows, MacOS and Linux, via \code{JuliaCall} or \code{JuliaConnectoR}) that defines the area of interest. \code{NAs} define inhospitable regions (such as land). A planar (e.g., Universal Transverse Mercator) projection with coordinates in metres is currently required.
}

\section{\code{.model_obs}}{
\code{.model_obs} is a named \code{list} of \code{\link{ModelObs}} sub-types (named) and associated parameters (\code{\link[data.table:data.table]{data.table::data.table}} elements).
}

\section{\code{.model_move}}{
\code{.model_move} is a \code{character} string that defines a \code{\link{ModelMove}} instance.
}

\section{\code{.state}}{
\code{.state} is a \code{character} that defines a \code{\link{State}} sub-type.
}

\section{\code{.yobs}}{
\code{.yobs} is a named \code{list} of observational datasets, one for each data type. Element names must correspond \code{\link{ModelObs}} sub-types defined in \code{Julia}. Each element must be a \code{\link[data.table:data.table]{data.table::data.table}} with the following columns:
\itemize{
\item \code{timestamp}---a \code{POSIXct} vector of time stamps;
\item \code{sensor_id}---an \code{integer} vector of sensor IDs, such as receivers;
\item \code{obs}---a vector of observations;
\item Additional columns that define the parameters of the observation model, as defined by a \code{\link{ModelObs}} structure;
}

No other columns should be included.

For real-world analyses, see \code{\link{assemble}}\verb{_*()} functions to format datasets as required. Other types of datasets require manual preparation.
}

\seealso{
\code{\link{State}}, \code{\link{ModelMove}}, \code{\link{ModelObs}}, \code{\link[=sim_path_walk]{sim_path_walk()}}, \code{\link[=sim_observations]{sim_observations()}}, \code{\link[=pf_filter]{pf_filter()}}
}
\author{
Edward Lavender
}
