% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mktempdir}
\alias{mktempdir}
\title{Make a temporary directory}
\usage{
mktempdir(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{file.path}}.}
}
\description{
This function wipes and (re)-creates a temporary directory.
}
\author{
Edward Lavender
}
