% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-name.R
\name{name_particles}
\alias{name_particles}
\title{Julia: naming functions}
\usage{
name_particles(
  .fun = c("pf_filter", "pf_smoother_two_filter"),
  .direction = c("forward", "backward")
)
}
\description{
A set of internal naming functions.
}
\author{
Edward Lavender
}
\keyword{internal}
