% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise}
\alias{normalise}
\title{Normalise a \code{\link{SpatRaster}}}
\usage{
normalise(x)
}
\arguments{
\item{x}{A \code{\link{SpatRaster}}.}
}
\description{
This function normalises a \code{\link{SpatRaster}}.
}
\section{Warning}{
\itemize{
\item For speed in iterative applications (e.g., \code{\link[=acs]{acs()}}), this function does not implement any internal checks.
\item \code{NA}s are ignored.
}
}

\author{
Edward Lavender
}
