% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{par_mf}
\alias{par_mf}
\title{Plot helper: collate plots on one page}
\source{
\code{prettyGraphics::par_mf()}
}
\usage{
par_mf(.n)
}
\arguments{
\item{.n}{An \code{integer} that defines the number of plots.}
}
\value{
The function returns an integer vector of two numbers.
}
\description{
This function is imported from \code{prettyGraphics::par_mf()}.
}
\examples{
# par_mf(10)
}
\author{
Edward Lavender
}
\keyword{internal}
