% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter-deprecated.R
\name{patter-deprecated}
\alias{patter-deprecated}
\alias{pf_plot_xy}
\title{Deprecated functions}
\usage{
pf_plot_xy(...)
}
\arguments{
\item{...}{Function arguments.}
}
\description{
These functions are deprecated in \code{\link{patter}}.
}
\details{
\itemize{
\item \code{\link[=pf_plot_xy]{pf_plot_xy()}} has been renamed \code{\link[=plot_xyt]{plot_xyt()}};
}
}
