% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter-package.R
\docType{package}
\name{patter-package}
\alias{patter}
\alias{patter-package}
\title{patter: Movement Modelling For Passive Acoustic Telemetry}
\description{
A new implementation of the 'flapper' family of algorithms for passive acoustic telemetry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/patter/}
  \item Report bugs at \url{https://github.com/edwardlavender/patter/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
\keyword{internal}
