% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter-progress.R
\name{patter-progress}
\alias{patter-progress}
\title{\code{\link{patter}} options: progress}
\description{
\code{\link{patter}} functions enable progress monitoring via function arguments and global options. At the time of writing, the main tools to monitor and enhance function progress are as follows:
\itemize{
\item User output messages (via the \code{patter.verbose} option and the \code{.verbose} argument);
}

Only selected \code{\link{patter}} functions support these options but this may expand in the future, depending on user feedback.
}
\section{User outputs}{
User output messages are controlled via the \code{.verbose} argument. There is a global option \code{patter.verbose} that can be set to suppress user output messages. See the internal \code{\link{cat_}} documentation for permitted inputs to \code{.verbose}.
}

\author{
Edward Lavender
}
