% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter-package.R
\docType{package}
\name{patter}
\alias{patter-package}
\alias{patter}
\title{\code{\link{patter}} for passive acoustic telemetry}
\description{
\code{\link{patter}} is a \code{R} implementation of particle filtering, smoothing and sampling algorithms for animal movement modelling in passive acoustic telemetry systems. This methodology enables the reconstruction of movement paths and patterns of space use \code{\link{patter}} unifies a suite of methods formerly known as the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} algorithms (Lavender et al., 2023) and supersedes the experimental \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package.
}
\references{
Lavender, E. et al. (2023). An integrative modelling framework for passive acoustic telemetry. Methods in Ecology and Evolution. \url{https://doi.org/10.1111/2041-210X.14193}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/patter/}
  \item Report bugs at \url{https://github.com/edwardlavender/patter/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
