% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{pb}
\alias{pb}
\alias{pb_init}
\alias{pb_tick}
\title{Progress bar wrappers}
\usage{
pb_init(.n, .progress)

pb_tick(.pb, .progress)
}
\description{
Progress bar wrappers
}
\examples{
\dontrun{
# Define loop
loop <- function(.progress) {
  n <- 10L
  pb <- pb_init(n, .progress)
  for (i in seq_len(n)){
    Sys.sleep(0.1)
    pb_tick(pb, .progress)
  }
}

# Show progress bar
loop(.progress = TRUE)

# Hide no progress bar
loop(.progress = FALSE)
}
}
\keyword{internal}
