% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-classes.R
\docType{package}
\name{pf-class}
\alias{pf-class}
\title{PF outputs: \code{\linkS4class{pf}} objects}
\description{
An S3 class that defines the named \code{list} returned by a PF-branch algorithm.
}
\author{
Edward Lavender
}
