% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward.R
\name{pf_backward_*}
\alias{pf_backward_*}
\title{PF: run the backward pass}
\description{
\code{\link{patter}} provides two options for the backward simulation:
\itemize{
\item \code{\link[=pf_backward_killer]{pf_backward_killer()}}
\item \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}
}
}
