% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward_killer.R
\name{pf_backward_killer_diagnostics}
\alias{pf_backward_killer_diagnostics}
\title{PF: backward killer diagnostics}
\usage{
pf_backward_killer_diagnostics(
  .sink,
  .cl = NULL,
  .cl_varlist = NULL,
  .cl_chunks = TRUE
)
}
\arguments{
\item{.sink}{A character string that defines the directory containing outputs.}

\item{.cl, .cl_varlist, .cl_chunks}{(optional) Cluster options, passed to \code{\link[=cl_lapply]{cl_lapply()}}.}
}
\value{
The function returns a \code{\link{data.table}} with the following columns:
\itemize{
\item \code{timestep}---an \code{integer} that defines the timestep;
\item \code{n}---an \code{integer} that defines the number of particles;
\item \code{n_u}---an \code{integer} that defines the number of unique location samples;
\item \code{ess}---a \code{double} that defines the effective sample size (see \code{\link[=.pf_diag_ess]{.pf_diag_ess()}});
}
}
\description{
This function collates diagnostics from \code{\link[=pf_backward_killer]{pf_backward_killer()}} outputs.
}
\author{
Edward Lavender
}
