% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map.R
\name{pf_coord}
\alias{pf_coord}
\title{PF: particle coordinates}
\usage{
pf_coord(.history, .bathy, .obs = NULL, .cols = NULL)
}
\arguments{
\item{.history}{Particle samples, provided in any format accepted by \code{\link[=.pf_history_dt]{.pf_history_dt()}}.}

\item{.bathy}{The bathymetry \code{\link{SpatRaster}}.}

\item{.obs, .cols}{(optional) A \code{\link{data.table}} and a character vector of column names in \code{.obs} to match onto the output. \code{.obs} must contain a \code{timestep} column for matching.}
}
\value{
The function returns a \code{\link{data.table}} that defines time steps, sampled locations and, optionally, information from \code{.obs}, with the following columns:
\itemize{
\item \code{timestep}---an \code{integer} vector that defines the time step;
\item \code{cell_id}---an \code{integer} vector that defines the cell ID on \code{.bathy};
\item \code{cell_x}, \code{cell_y}, \code{cell_z}---\code{double}s that define cell coordinates;
}

If \code{.obs} is supplied, the output also contains any columns specified in \code{.cols}.
}
\description{
This function collects particle samples and extracts coordinates.
}
\details{
This function is not memory safe.
}
\examples{
p <- dat_pfbk()
pf_coord(p$history, dat_gebco())

}
\seealso{
TO DO
}
\author{
Edward Lavender
}
