% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map.R
\name{pf_coords}
\alias{pf_coords}
\title{PF: particle coordinates}
\usage{
pf_coords(.history, .bathy, .obs = NULL, .cols = NULL)
}
\arguments{
\item{.history}{Particle samples from the particle filter, provided either as:
\itemize{
\item A \code{list} of \code{\link{data.table}}s that define cell samples; i.e., the \code{history} element of a \code{\linkS4class{pf}} object. This must contain a column that defines accepted cell samples at each time step (\code{cell_now}).
\item An ordered list of file paths (from \code{\link[=pf_setup_files]{pf_setup_files()}}) that define the directories in which particle samples were written from the forward simulation (as parquet files).
}}

\item{.bathy}{The bathymetry \code{\link{SpatRaster}}.}

\item{.obs, .cols}{(optional) A \code{\link{data.table}} and a character vector of column names in \code{.obs} to match onto the output. \code{.obs} must contain a \code{timestep} column for matching.}
}
\value{
The function returns a \code{\link{data.table}} that defines timesteps, sampled locations and, optionally, information from \code{.obs}, with the following columns:
\itemize{
\item \code{timestep}---an \code{integer} vector that defines the time step;
\item \code{cell_id}---an \code{integer} vector that defines the cell ID on \code{.bathy};
\item \code{cell_x}, \code{cell_y}, \code{cell_z}---\code{double}s that define cell coordinates;
}

If \code{.obs} is supplied, the output also contains any columns specified in \code{.cols}.
}
\description{
This function collects particle samples and extracts coordinates.
}
\details{
This function is not memory safe.
}
\examples{
p <- dat_pfb()
pf_coords(p$history, dat_gebco())

}
\seealso{
\itemize{
\item The PF (forward simulation) is implemented by \code{\link[=pf_forward_*]{pf_forward_*()}}:
\itemize{
\item \code{\link[=pf_forward_1]{pf_forward_1()}} refines AC-branch algorithm (\code{\link[=acs]{acs()}} and \code{\link[=dc]{dc()}}) outputs using PF;
\item \code{\link[=pf_forward_2]{pf_forward_2()}} is an integrated implementation that couples AC- and PF-branch algorithms internally;
}
\item PF is supported by:
\itemize{
\item Setup helpers, namely \code{\link[=pf_setup_files]{pf_setup_files()}};
\item Template movement models, namely \code{\link[=pf_kick]{pf_kick()}};
}
\item The backward pass is implemented by \code{\link[=pf_backward]{pf_backward()}};
\item Movement paths are built from PF outputs via \code{pf_path()} functions:
\itemize{
\item \code{\link[=pf_path]{pf_path()}} reconstructs paths;
\item \code{\link[=pf_path_pivot]{pf_path_pivot()}} supports path reconstruction;
}
\item To reconstruct maps of space use, see:
\itemize{
\item \code{\link[=pf_coords]{pf_coords()}} to extract particle coordinates;
\item \code{\link[=pf_map_pou]{pf_map_pou()}} for probability-of-use maps;
\item \code{\link[=pf_map_dens]{pf_map_dens()}} for smooth utilisation distributions;
\item \code{\link[=get_hr]{get_hr()}} for home range estimates;
}
}
}
\author{
Edward Lavender
}
