% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_diag.R
\name{pf_diag-internal}
\alias{pf_diag-internal}
\alias{.pf_diag_any}
\alias{.pf_diag_nu}
\alias{.pf_diag_ess}
\title{PF: particle diagnostics}
\usage{
.pf_diag_any(.particles)

.pf_diag_nu(.particles)

.pf_diag_ess(.logwt)
}
\description{
These are internal functions that calculate diagnostic statistics from selected particle samples.
}
\section{Particle diagnostics}{
Particle diagnostics measure how the diversity of particle samples evolves through time. Current diagnostic metrics include the number of unique particle samples at each time step and the effective sample size. Particle diagnostics have multiple uses:
\itemize{
\item \strong{Convergence.} We formerly tracked particle diagnostics in \code{\link[=pf_forward]{pf_forward()}} to facilitate identification of the causes of convergence failures, but this is not currently implemented.
\item \strong{Sampling.} Diagnostics indicate sampling sufficiency. In general, a low ratio of the number of unique particles to the total number of particles indicates effective coverage of the possible locations of an individual.
\item \strong{Particle degeneracy.} Diagnostics measure particle degeneracy, i.e., the decay in the number of unique particles through time as trajectories are rendered invalid by observations.
}
}

\section{Internal routines}{
\itemize{
\item \code{\link[=.pf_diag_any]{.pf_diag_any()}} identifies whether or not any particle samples remain;
\item \code{\link[=.pf_diag_nu]{.pf_diag_nu()}} counts the number of unique particle samples (grid cells);
\item \code{\link[=.pf_diag_ess]{.pf_diag_ess()}} calculates effective sample size from normalised particle (log) weights;
}
}

\seealso{
\itemize{
\item \code{\link[=pf_forward]{pf_forward()}} and associates implement particle filtering algorithms;
\item \code{\link[=pf_diag_summary]{pf_diag_summary()}} summarises diagnostics;
\item \code{\link{.pf_diag}}\verb{_()} functions are internal routines that calculate diagnostic statistics;
}
}
\author{
Edward Lavender
}
\keyword{internal}
