% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward_sampler.R
\name{pf_distinct}
\alias{pf_distinct}
\title{PF: isolate distinct samples}
\usage{
pf_distinct(
  .history,
  .cl = NULL,
  .cl_varlist = NULL,
  .cl_chunks = TRUE,
  .save_opts = FALSE,
  .write_opts = list()
)
}
\arguments{
\item{.history}{Particle samples from the backward pass, provided either as:
\itemize{
\item A \code{list} of \code{\link{data.table}}s that define cell samples; i.e., the \code{history} element of a \code{\linkS4class{pf}} object.
\item An ordered list of file paths (from \code{\link[=pf_setup_files]{pf_setup_files()}}) that define the directories in which particle samples were written from the forward simulation (as parquet files).
}}

\item{.cl, .cl_varlist, .cl_chunks}{Parallelisation options, passed to \code{\link[=cl_lapply]{cl_lapply()}}.}

\item{.save_opts, .write_opts}{Arguments to save particle histories in memory or write them to file (see \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}).}
}
\value{
The function returns a \code{list} of distinct particle samples (if \code{.save_opts = TRUE}) or \code{NULL}. If \code{.write_opts} is supplied, distinct particle samples are written to file.
}
\description{
This function identifies distinct particle (cell) samples from the forward filter for \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}.
}
\details{
This function iterates over time steps, identifies distinct cells and saves these in memory and/or writes them to file for \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}. This is designed to improve speed in \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}, which only requires distinct samples.
}
\author{
Edward Lavender
}
