% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-initialise.R
\name{pf_filter_init}
\alias{pf_filter_init}
\title{Initialise the particle filter
This is an internal function that initialises the particle filter:
\itemize{
\item Initial states are set via \code{\link[=set_states_init]{set_states_init()}};
\item Observations are set via \code{\link[=set_yobs_dict]{set_yobs_dict()}};
}}
\usage{
pf_filter_init(
  .timeline,
  .state = "StateXY",
  .xinit = NULL,
  .model_move,
  .yobs,
  .n_particle,
  .direction,
  .verbose = getOption("patter.verbose")
)
}
\description{
Initialise the particle filter
This is an internal function that initialises the particle filter:
\itemize{
\item Initial states are set via \code{\link[=set_states_init]{set_states_init()}};
\item Observations are set via \code{\link[=set_yobs_dict]{set_yobs_dict()}};
}
}
\author{
Edward Lavender
}
\keyword{internal}
