% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward_2-utils-internal.R
\name{pf_forward_2_utils}
\alias{pf_forward_2_utils}
\alias{.pf_dirs}
\alias{.pf_write}
\alias{.pf_startup}
\alias{pf_start_t}
\alias{.pf_next}
\alias{pf_ppast}
\title{PF helper: internal utilities}
\usage{
.pf_dirs(.write_opts)

.pf_write(.particles, .diagnostics, .psink, .dsink, .write)

.pf_startup(
  .rerun,
  .obs,
  .lonlat,
  .bathy,
  .moorings,
  .detection_overlaps,
  .detection_kernels,
  .update_ac,
  .write_opts
)

pf_start_t(.rerun, .rerun_from)

.pf_next(.particles)

pf_ppast(.particles, .history, .t)
}
\description{
PF helper: internal utilities
}
\keyword{internal}
