% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward.R
\name{pf_forward_diagnostics}
\alias{pf_forward_diagnostics}
\title{PF: forward run diagnostics}
\usage{
pf_forward_diagnostics(.sink)
}
\arguments{
\item{.sink}{A character string that defines the directory in which \code{\link[=pf_forward]{pf_forward()}} outputs are saved. \code{.sink} should contain the \verb{diagnostics/} directory that is created by \code{\link[=pf_forward]{pf_forward()}}.}
}
\value{
The function returns the diagnostics \code{\link{data.table}}.
}
\description{
This function collates diagnostics from the forward run.
}
\details{
This function is not memory safe.
}
\author{
Edward Lavender
}
