% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-likelihood.R
\name{pf_lik}
\alias{pf_lik}
\alias{acs_filter_land}
\alias{acs_filter_container}
\alias{pf_lik_ac}
\alias{pf_lik_update}
\title{PF: calculate particle likelihoods}
\usage{
acs_filter_land(.particles, .bathy)

acs_filter_container(.particles, .moorings, .receivers, .threshold)

pf_lik_ac(.particles, .obs, .t, .detection_overlaps, .detection_kernels)

pf_lik_update(.particles, .obs, .t, .bathy, .update_ac)

pf_lik(
  .particles,
  .obs,
  .t,
  .bathy,
  .is_land,
  .moorings,
  .detection_overlaps,
  .detection_kernels,
  .update_ac,
  .trial = NA_integer_
)
}
\arguments{
\item{.moorings}{For \code{\link{acs_filter_container}},}

\item{.detection_overlaps, .detection_kernels, }{For \code{\link{pf_lik_ac}}}

\item{.update_ac}{For \code{\link{pf_lik_update}}}

\item{.is_land}{For \code{\link{acs_filter_land}},}
}
\description{
PF: calculate particle likelihoods
}
\details{
TO DO
provide likelihood so you can stack them how you like e.g. filters first then DC or AC
each function should accept particles, return filtered particles & likelihood column
}
\author{
Edward Lavender
}
\keyword{internal}
