% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-particles.R
\name{pf_particle}
\alias{pf_particle}
\alias{.pf_sample_origin}
\alias{.pf_particles_origin}
\alias{.pf_particles_kick}
\alias{.pf_particles_sampler}
\title{PF: particle helpers}
\usage{
.pf_sample_origin(.particles, .sample, .n, .trial_crit, .trial_count)

.pf_particles_origin(
  .obs,
  .dlist,
  .origin,
  .grid = FALSE,
  .likelihood,
  .sample,
  .n,
  .trial_crit,
  .trial_count
)

.pf_particles_kick(
  .particles,
  .rpropose,
  .obs,
  .t,
  .bathy,
  .pf_lik,
  .sample,
  .n,
  .trial_crit,
  .trial_count
)

.pf_particles_sampler(
  .particles,
  .obs,
  .t,
  .dlist,
  .likelihood,
  .dpropose,
  .sample,
  .n,
  .trial_crit,
  .trial_count,
  .control
)
}
\description{
PF: particle helpers
}
\details{
Hierarchy
pf_propose_
pf_sample_
pf_particles

.pf_rpropose_origin proposes starting locations
.pf_sample_origin samples particles
.pf_particles_origin integrates .pf_propose_origin, likelihood calculations and .pf_particles_origin

.pf_propose_kick and .pf_propose_reachable are subsequent proposal functions (see XXX)
.pf_particles_kick and .pf_particles_sampler are the integration functions which implement proposals and sampling
}
\keyword{internal}
