% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-classes.R
\name{pf_particles-class}
\alias{pf_particles-class}
\title{PF: \code{\linkS4class{pf_particles}} objects}
\description{
An S3 class that defines the named \code{list} returned by \code{\link[=pf_forward]{pf_forward()}} and \code{\link[=pf_backward_*]{pf_backward_*()}}.
}
\section{Structure}{
\code{\linkS4class{pf_particles}} is a label used to denote outputs from selected functions in \code{\link{patter}}. The structure of this class is not strictly defined and primarily exists to streamline documentation. At the time of writing, \code{\linkS4class{pf_particles}} objects may comprise the following elements:
\itemize{
\item \code{history}---a \code{list} of particle samples;
\item \code{convergence}---a \code{logical} value that defines convergence success;
\item \code{time}---a \code{data.table} of timings;
}
}

\section{\code{history}}{
\code{history} is a \code{list} with one element for each time step. Each element is a \code{\link{data.table}} that contains location (particle) samples. Each row is a specific sample.

In \code{\link[=pf_forward]{pf_forward()}}, the columns in this table depend on \code{\link[=pf_forward]{pf_forward()}}'s \code{.likelihood} function(s) and the \code{.record$cols} argument (see \code{\link[=pf_opt_record]{pf_opt_record()}}). By default, the following columns are computed:
\itemize{
\item \code{timestep}---an \code{integer} vector that defines the time step;
\item \code{cell_past}---an \code{integer} vector that identifies the grid cells of previous particle samples;
\item \code{x_past}, \code{y_past}---\code{numeric} vectors that define the coordinates of previous particle samples;
\item \code{cell_now}---an \code{integer} vector that identifies the grid cells of current particle samples;
\item \code{x_now}, \code{y_now}---\code{numeric} vectors that define the coordinates of current particle samples;
\item \code{lik}---a \code{numeric} vector of likelihoods;
\item \code{weight}---a \code{numeric} vectors of weights;
}

Coordinates are defined in continuous space by \code{\link[=pf_rpropose_kick]{pf_rpropose_kick()}} and on the grid by \code{\link[=pf_rpropose_reachable]{pf_rpropose_reachable()}}.

The likelihood column (\code{lik}) is the product of the likelihood scores from each likelihood function. (Successive likelihood functions should \emph{update} this column.)

The weight column (\code{weight}) contains particle weights. Weights are proportional to the product of weights from the previous time step, the likelihood at the current time step (and the movement density, in the case of directed sampling). After resampling, we obtain \code{.n} equally weighted particles. Weights are normalised to sum to one at each time step.

Additional columns may be included in \code{history} if computed by inputted likelihood functions. At the time of writing, most default likelihood functions calculate likelihoods without defining additional columns. However, \code{\link[=pf_lik_dc]{pf_lik_dc()}} adds \code{bathy} column, with values extracted from \code{.dlist$spatial$bathy}.
If directed sampling is used, \code{bathy} and \code{dens} columns are included where required (after the first time step). \code{dens} is \code{numeric} vector that defines the probability density of moving from \verb{(x_past, y_past}) to \verb{(x_now, y_now)}.

Since columns are recorded only when required, not all columns are recorded at each time step. To combine \code{\link{data.table}}s, use \code{\link[=.pf_history_dt]{.pf_history_dt()}}.

If \code{.record$save = FALSE}, \code{.history} is an empty \code{list} and the individual \code{\link{data.table}}s are written to file in \verb{\{.record$sink\}/history/} (see \code{\link[=pf_opt_record]{pf_opt_record()}}).
}

\section{\code{convergence}}{
For \code{\link[=pf_forward]{pf_forward()}}, \code{convergence} is a \code{logical} variable that defines whether or not the algorithm successfully converged (i.e., reached the end of the time series).
}

\section{\code{time}}{
A \code{data.table} that records the \code{start}, \code{end} and \code{duration} of algorithm runs (as \code{POSIXct} and \code{\link{difftime}} objects).
}

\seealso{
The forward filtering--backward sampling algorithm samples locations (particles) that represent the possible locations of an individual through time, accounting for all data and the individual's movement.
\itemize{
\item To set up data, use \code{\link[=pat_setup_data]{pat_setup_data()}}.
\item \code{\link[=pf_forward]{pf_forward()}} implements the forward filter:
\itemize{
\item To set up an observations timeline, use \code{\link[=pf_setup_obs]{pf_setup_obs()}}.
\item For proposal (movement) models, see \code{\link{pf_propose}}.
\item For likelihood functions, to evaluate the likelihood of the data at proposal locations, see \code{\link{pf_lik}}.
\item For sampling functions, to (re)sample plausible proposal locations, see \code{\link{pf_sample}}.
\item For tuning parameters, see \code{\link{pf_opt}}.
}
\item \code{\link[=pf_backward_*]{pf_backward_*()}} refines outputs from the forward filter:
\itemize{
\item \code{\link[=pf_backward_killer]{pf_backward_killer()}} removes dead-ends;
\item \code{\link{pf_backward_sampler}}\verb{_*()} implements the backward sampler;
}
\item To reconstruct movement paths from particle samples, use \code{\link[=pf_path]{pf_path()}}.
\item To map emergent patterns of space use, use \code{\link[=pf_coord]{pf_coord()}} plus a \verb{map_*()} function, such as \code{\link[=map_pou]{map_pou()}}, \code{\link[=map_dens]{map_dens()}} and/or \code{\link{map_hr}}\verb{_()}.
\item For additional utilities, see supporting \verb{pf_*()} functions, such as \code{\link[=pf_files]{pf_files()}}, \code{\link[=pf_files_size]{pf_files_size()}}, \code{\link[=pf_plot_history]{pf_plot_history()}}, \code{\link[=pf_diag_convergence]{pf_diag_convergence()}} and \code{\link[=pf_diag_summary]{pf_diag_summary()}}.
}
}
\author{
Edward Lavender
}
