% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_path-internal.R
\name{pf_path_}
\alias{pf_path_}
\alias{.pf_path_pivot_checks}
\alias{.pf_path_join}
\alias{.pf_path_chain}
\title{PF: path reconstruction helpers}
\usage{
.pf_path_pivot_checks(.obs, .cols)

.pf_path_join(.data, .current, .t, .pb = NULL, .pb_step)

.pf_path_chain(.history, .read)
}
\arguments{
\item{.obs, .cols}{Arguments for \code{\link[=.pf_path_pivot_checks]{.pf_path_pivot_checks()}}}

\item{.data, .current., t., pb}{Arguments for \code{\link[=.pf_path_join]{.pf_path_join()}}.}

\item{.history, .read}{Arguments for \code{\link[=.pf_path_chain]{.pf_path_chain()}}.}
}
\description{
These functions facilitate the reconstruction of movement paths.
}
\author{
Edward Lavender
}
\keyword{internal}
