% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_path.R
\name{pf_path_pivot}
\alias{pf_path_pivot}
\title{PF: pivot path matrices}
\usage{
pf_path_pivot(.mat)
}
\arguments{
\item{.mat}{A wide-format \code{\link{data.table}}, from \code{\link[=pf_path]{pf_path()}} with \code{.return = "wide"}, in which:
\itemize{
\item columns (\code{x1}, etc.) represent time steps;
\item rows represent time steps;
}}
}
\value{
The function returns a \code{\link{data.table}} with three columns:
\itemize{
\item \code{id}---An \code{integer} that defines each path (row in \code{.mat});
\item \code{timestep}---An \code{integer} that defines each time step (column in \code{.mat});
\item \code{cell}---An integer that defines each cell (value in \code{.mat});
}
}
\description{
This function converts paths in 'wide format' to paths in 'long format'.
}
\details{
This function uses the fast \code{\link[collapse:pivot]{collapse::pivot()}} function.
}
\examples{
# Define a hypothetical set of paths in wide-format
require(data.table)
mat <- matrix(1:25, ncol = 5)
mat <- as.data.table(mat)
colnames(mat) <- c("x0", "x1", "x2", "x3", "x4")

# Convert to long format
pf_path_pivot(mat)

}
\author{
Edward Lavender
}
