% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_setup-internal.R
\name{pf_setup-internal}
\alias{pf_setup-internal}
\alias{.setup_arrays}
\alias{.setup_acoustics}
\alias{.setup_archival}
\alias{.setup_alignment}
\alias{.setup_series}
\alias{.add_acoustics_arrays}
\alias{.add_acoustics_flags}
\alias{.add_acoustics_obs}
\alias{.add_acoustics_ridnxt}
\alias{.add_acoustics_containers}
\alias{.add_archival}
\title{PF: set up internals}
\usage{
.setup_arrays(.dlist)

.setup_acoustics(.acoustics, .step)

.setup_archival(.archival, .step)

.setup_alignment(.acoustics, .archival)

.setup_series(.acoustics, .archival, .period, .step, .mobility)

.add_acoustics_arrays(.obs, .dlist, .step)

.add_acoustics_flags(.obs, .acoustics)

.add_acoustics_obs(.obs, .dlist, .acoustics)

.add_acoustics_ridnxt(.receiver_id)

.add_acoustics_containers(.obs, .dlist, .mobility)

.add_archival(.obs, .archival)
}
\description{
These are internal functions that support \code{\link[=pf_setup_obs]{pf_setup_obs()}}.
}
\keyword{internal}
