% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-setup.R
\name{pf_setup_files}
\alias{pf_setup_files}
\title{PF set up: list files for PF}
\usage{
pf_setup_files(.root, ...)
}
\arguments{
\item{.root}{A string that defines the directory in which files are located.}

\item{...}{Additional arguments passed to \code{\link[=list.files]{list.files()}}, such as \code{pattern}, excluding \code{full.names}.}
}
\value{
The function returns an ordered \code{list} of file paths.
}
\description{
This function creates an ordered \code{list} of files for \code{\link[=pf_backward_*]{pf_backward_*()}}.
}
\examples{
# List files from pf_forward()
pff_folder <- system.file("extdata", "acpf", "forward", "history",
                          package = "patter", mustWork = TRUE)
files <- pf_setup_files(pff_folder)


# List files from pf_backward_killer()
pfbk_folder <- system.file("extdata", "acpf", "backward", "killer",
                           package = "patter", mustWork = TRUE)
files <- pf_setup_files(pfbk_folder)

}
\author{
Edward Lavender
}
