% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-setup.R
\name{pf_setup_kick}
\alias{pf_setup_kick}
\title{PF set up: a template movement model}
\usage{
pf_setup_kick(.particles, .obs = NULL, .t = NULL, .bathy = NULL, ...)
}
\arguments{
\item{.particles}{A \code{\link{data.table}}, from \code{\link[=pf]{pf()}}, that defines the current particle samples:
\itemize{
\item \code{cell_now} is an integer vector of cell IDs;
\item \code{x_now} is a numerical vector of \code{x} coordinates;
\item \code{y_now} is a numerical vector of \code{y} coordinates;
}}

\item{.obs, .t, .bathy}{(optional) The \code{.obs} \code{\link{data.table}}, an integer that indexes \code{.obs} and the \code{.bathy} \code{\link{SpatRaster}} (see \code{\link[=pf]{pf()}}). These inputs are unused in this template movement model but supported within \code{\link[=pf]{pf()}}.}

\item{...}{Additional arguments passed from \code{\link[=pf]{pf()}} (unused here).}
}
\value{
The function returns a \code{\link{data.table}}, as supplied, with two additional columns:
\itemize{
\item \code{x_next}---the \code{x} coordinate of the next proposal location(s);
\item \code{y_next}---the \code{y} coordinates of the next proposal location(s);
}
}
\description{
This function is an example movement model, of the kind required by \code{.kick} in \code{\link[=pf]{pf()}}.
}
\details{
This template movement model is a biased random walk. Step lengths are simulated from a Gamma distribution via \code{\link[stats:GammaDist]{stats::rgamma()}}. Turning angles are simulated from a wrapped normal distribution via \code{\link[circular:wrappednormal]{circular::rwrappednormal()}}. See the the code for the parameters used.
}
\section{Warning}{
\itemize{
\item This function is used to streamline examples and does not represent a generically suitable detection probability model.
\item The function does not check user inputs.
}
}

\examples{
# Load packages
require(graphics)
require(data.table)
require(dtplyr)
require(dplyr, warn.conflicts = FALSE)

# Set seed
set.seed(1)

# Define hypothetical `.particles` data.table
p <-
  dat_gebco() |>
  terra::spatSample(size = 10L, cells = TRUE, xy = TRUE, as.df = FALSE,
                    na.rm = TRUE) |>
  as.data.table() |>
  select(cell_now = cell, x_now = x, y_now = y) |>
  as.data.table()

# Kick particles into new locations
p <- pf_setup_kick(p)

# Visualise current & next locations
terra::plot(dat_gebco())
arrows(x0 = p$x_now, x1 = p$x_next,
       y0 = p$y_now, y1 = p$y_next,
       length = 0.02)

}
\author{
Edward Lavender
}
