% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-internal.R
\name{pf_utils-internal}
\alias{pf_utils-internal}
\alias{.pf_sink_folder}
\alias{.pf_snapshot}
\alias{.pf_write_particles}
\title{PF: internal utilities}
\usage{
.pf_sink_folder(.sink, .folder)

.pf_snapshot(.dt, .save, .select, .cols)

.pf_write_particles(.particles, .sink, .filename, .write)
}
\description{
These functions are shared internal utilities that support particle filtering routines, especially \code{\link[=pf_forward]{pf_forward()}} and \code{\link[=pf_backward_*]{pf_backward_*()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
