% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-internal.R
\name{pf_utils_internal}
\alias{pf_utils_internal}
\alias{.pf_sink_folder}
\title{PF: internal utilities}
\usage{
.pf_sink_folder(.sink, .folder)
}
\description{
These functions are internal utilities that support particle filtering routines.
}
\author{
Edward Lavender
}
\keyword{internal}
