% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot.ModelMove}
\alias{plot.ModelMove}
\alias{plot.ModelMoveXY}
\alias{plot.ModelMoveXYZ}
\alias{plot.ModelMoveCXY}
\alias{plot.ModelMoveCXYZ}
\title{Movement model plots}
\usage{
\method{plot}{ModelMoveXY}(x, .panel_length = list(), .panel_heading = list(), .par = list(), ...)

\method{plot}{ModelMoveXYZ}(
  x,
  .panel_length = list(),
  .panel_heading = list(),
  .panel_z = list(),
  .par = list(),
  ...
)

\method{plot}{ModelMoveCXY}(
  x,
  .panel_length = list(),
  .panel_heading_delta = list(),
  .par = list(),
  ...
)

\method{plot}{ModelMoveCXYZ}(
  x,
  .panel_length = list(),
  .panel_heading_delta = list(),
  .panel_z_delta = list(),
  .par = list(),
  ...
)
}
\arguments{
\item{x}{A \code{\link{ModelMove}}-class object from a \verb{model_move_*()} function (e.g., \code{\link[=model_move_xy]{model_move_xy()}}.}

\item{.panel_length, .panel_heading, .panel_heading_delta, .panel_z, .panel_z_delta}{Panel properties:
\itemize{
\item \code{NULL} suppresses the panel;
\item \code{list()} uses default graphical arguments;
\item A named \code{list} of arguments, passed to \code{\link[=plot]{plot()}}, customises the panel;
}}

\item{.par}{Graphical parameters:
\itemize{
\item \code{NULL} uses current graphical parameters;
\item \code{list()} uses default graphical parameters;
\item A named \code{list} of arguments, passed to \code{\link[=par]{par()}}, customises parameters;
}}

\item{...}{Additional arguments, passed to \code{\link[=plot]{plot()}}, which affect all panels;}
}
\value{
The functions produce a \code{\link{plot}}. \code{invisible(NULL)} is returned.
}
\description{
\code{\link[=plot]{plot()}} methods for movement models (see \code{\link{ModelMove}}).
}
\details{
For each in-built \code{\link{ModelMove}}-class (e.g., \code{\link{ModelMoveXY}}), a corresponding \code{\link[=plot]{plot()}} method is provided that plots the component probability-density distributions (e.g., the distribution of step lengths and turning angles) for any model instance. Under default options, a multi-panelled plot is produced, with one panel for each model dimension (e.g., the step length and the heading). Use \verb{.panel_*} arguments to customise individual panels, \code{.par} to set graphical parameters and \code{...} to customise all panels. Set \verb{.panel_*} \code{xlim} to specify the x-axis range over which distributions are shown. If unspecified, probability density is plotted along a predefined range or from the 0.0001 to 0.9999 quantiles of the distribution.
}
\examples{
if (patter_run(.julia = TRUE, .geospatial = FALSE)) {

  julia <- julia_connect()
  map <- dat_gebco(.return = "character")
  set_map(map)

  #### Example (1): Plot movement models
  # Plot default models
  plot(model_move_xy())
  plot(model_move_xyz())
  plot(model_move_cxy())
  plot(model_move_cxyz())
  # Plot customised model
  plot(model_move_cxyz(.dbn_length = "truncated(Normal(0.0, 100.0),
                                        lower = 0.0, upper = 1000.0)"))

  #### Example (2): Use `...` to pass graphical arguments to all panels
  plot(model_move_xy(), col = "red")

  #### Example (3): Use `.panel_* = NULL` to suppress selected panels
  plot(model_move_xy(),
       .panel_heading = NULL)

  #### Example (4): Customise individual panels via `.panel_*` lists
  plot(model_move_xy(),
       .panel_length = list(main = "A"),
       .panel_heading = list(main = "B"))

  #### Example (5): Control graphical parameters via the `.par` list
  # Use default options
  plot(model_move_xy(), .par = list())
  # Specify par options
  plot(model_move_xy(), .par = list(oma = c(3, 3, 3, 3)))
  # Set `.par = NULL` to leave `par` unchanged
  pp <- par(mfrow = c(2, 1))
  plot(model_move_xy(), .par = NULL)
  par(pp)

  #### Example (6): Simulate reasliations of the movement model
  # See `?sim_path_walk()`
  # Set `.map` argument to visualise trajectories
  # (On Linux, this is only possible if JULIA_SESSION = "FALSE")
  sim_path_walk(.timeline = seq(as.POSIXct("2016-01-01", tz = "UTC"),
                                as.POSIXct("2016-01-01 12:00:00", tz = "UTC"),
                                by = "2 mins"),
                .state = "StateXY",
                .model_move = model_move_xy())

  JuliaSwitch::julia_stop()

}
}
\seealso{
Use \code{\link[=sim_path_walk]{sim_path_walk()}} to visualise realisations of a movement model (i.e., trajectories).
}
\author{
Edward Lavender
}
