% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers.R
\name{rast_template}
\alias{rast_template}
\title{Spatial helper: create a template \code{\link{SpatRaster}}}
\usage{
rast_template(
  .xmin = 0,
  .xmax = 1000,
  .ymin = 0,
  .ymax = 1000,
  .res = c(10, 10),
  .crs = "+proj=utm +zone=1 +datum=WGS84",
  .value = 0,
  .units = "m",
  ...
)
}
\arguments{
\item{.xmin, .xmax, .ymin, .ymax}{Numbers that define the \code{\link{SpatRaster}} limits.}

\item{.res}{A numeric vector that defines \code{\link{SpatRaster}} resolution.}

\item{.crs}{A \code{character} that defines the Coordinate Reference System.}

\item{.value}{The value of the \code{\link{SpatRaster}} cells.}

\item{.units}{A character that defines the units.}

\item{...}{Additional arguments passed to \code{\link{SpatRaster}}.}
}
\value{
The function returns a \code{\link{SpatRaster}}.
}
\description{
This function creates a template \code{\link{SpatRaster}}.
}
\examples{
r <- rast_template()
terra::plot(r)
r <- rast_template(.value = 1)
terra::plot(r)
}
\author{
Edward Lavender
}
