% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers.R
\name{serial_distance}
\alias{serial_distance}
\title{Calculate serial distances}
\usage{
serial_distance(.xy, .lonlat = FALSE)
}
\arguments{
\item{.xy}{A two-column matrix of coordinates.}

\item{.lonlat}{A \code{logical} variable that defines whether or not the coordinates are in longitude/latitude format.}
}
\description{
This function calculates distances between sequential coordinates in a matrix.
}
\details{
This function is a simple wrapper for \code{\link[terra:distance]{terra::distance()}}.
}
