% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_set-seed.R
\name{set_seed}
\alias{set_seed}
\title{\code{Julia}: set the seed}
\usage{
set_seed(.seed = 123L)
}
\arguments{
\item{.seed}{An \code{integer} that defined the seed (see \code{\link[=set.seed]{set.seed()}}).}
}
\value{
The function returns \code{invisible(NULL)}.
}
\description{
Use \code{\link[=set_seed]{set_seed()}} to set the seed in \code{R} and \code{Julia} at the start of your workflow.
}
\examples{
if (patter_run(.geospatial = FALSE)) {
  julia_connect()
  set_seed(2024L)
  JuliaSwitch::julia_stop()
}

}
\seealso{
\code{\link[=set.seed]{set.seed()}}
}
\author{
Edward Lavender
}
