% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{sf-helper}
\alias{sf-helper}
\alias{st_invert}
\title{Spatial helper: \verb{st_*} functions}
\usage{
st_invert(.x, .bbox = sf::st_bbox(.x))
}
\description{
Internal \code{\link[sf:sf-package]{sf::sf-package}} helper functions.
}
\author{
Edward Lavender
}
\keyword{internal}
