% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-internal.R
\name{sim_detections_internals}
\alias{sim_detections_internals}
\alias{.sim_detections}
\alias{.sim_detections_call}
\title{Simulation: acoustic detections}
\usage{
.sim_detections(
  .path,
  .array,
  .calc_distance = terra::distance,
  .lonlat = FALSE,
  .rdet = rdet,
  ...,
  .return = c("array_id", "path_id", "timestep", "timestamp", "receiver_id", "dist",
    "pr")
)

.sim_detections_call(.path, .array, .args)
}
\description{
Internal functions that support the simulation of acoustic detections at receivers.
}
\details{
\itemize{
\item \code{\link[=.sim_detections]{.sim_detections()}} simulates detections for a single path/array.
\item \code{\link[=.sim_detections_call]{.sim_detections_call()}} is a \code{\link[=do.call]{do.call()}} wrapper for \code{\link[=.sim_detections]{.sim_detections()}}.
}
}
\author{
Edward Lavender
}
\keyword{internal}
