% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_helpers}
\alias{sim_helpers}
\title{Simulation helpers \code{d}, \code{r}, \code{c}}
\description{
These functions support the generation of animal movement paths and observations in \emph{de novo} simulations (\verb{sim_*()} functions) and simulation-based reconstructions of movement paths (\code{\link[=pf_forward_*]{pf_forward_*()}} and \code{\link[=pf_backward_*]{pf_backward_*()}}).
\itemize{
\item \verb{r*()} functions simulate random variates;
\item \verb{c*()} functions calculate outcomes from random-variate inputs;
\item \verb{d*()} functions return densities;
}

\verb{r*()} and \verb{c*()} functions are used in \emph{de novo} simulations (via \verb{sim_*()} functions such as \code{\link[=sim_path_walk]{sim_path_walk()}}) and the forward simulation-based reconstruction of movement paths (via \code{\link[=pf_kick]{pf_kick()}} in \code{\link[=pf_forward_*]{pf_forward_*()}}).

\verb{d*()} functions are used in the simulation-based reconstruction of movement paths as part of the backward sampler via \code{\link[=pf_backward_p]{pf_backward_p()}}.
}
\section{Movement}{
\subsection{Core functions}{
\subsection{Step lengths}{
\itemize{
\item \code{\link[=rtruncgamma]{rtruncgamma()}} simulates step length(s) from a truncated gamma distribution with a \code{.mobility} parameter.
\item \code{\link[=dtruncgamma]{dtruncgamma()}} returns the densit(ies) of specified step length(s).
}
}

\subsection{Turning angles}{
\itemize{
\item \code{\link[=rwn]{rwn()}} simulates turning angle(s) from a wrapped normal distribution.
\item \code{\link[=dwn]{dwn()}} returns the densit(ies) of specified angle(s).
}
}

}

\subsection{Wrappers}{

The following wrapper functions are provided in the form required by front-end functions (e.g., \code{\link[=sim_path_walk]{sim_path_walk()}}, \code{\link[=pf_kick]{pf_kick()}} and \code{\link[=pf_backward_p]{pf_backward_p()}}):
\itemize{
\item \code{\link[=rlen]{rlen()}} and \code{\link[=dlen]{dlen()}} are wrappers for \code{\link[=rtruncgamma]{rtruncgamma()}} and \code{\link[=dtruncgamma]{dtruncgamma()}}.
\item \code{\link[=rangrw]{rangrw()}}, \code{\link[=rangcrw]{rangcrw()}}, \code{\link[=dangrw]{dangrw()}} and \code{\link[=dangcrw]{dangcrw()}} are wrappers for \code{\link[=rwn]{rwn()}} and \code{\link[=dwn]{dwn()}} for random walks and correlated random walks.
}
}

\subsection{Extensions}{
\itemize{
\item \code{\link[=cstep]{cstep()}} steps from previous location(s) into future location(s), given simulated step length(s) and turning angle(s).
\item \code{\link[=dstep]{dstep()}} returns the density of steps between two locations.
}
}
}

\seealso{
\itemize{
\item \verb{sim_*} functions implement \emph{de novo} simulation of movements and observations;
\item \code{\link[=pf_forward_*]{pf_forward_*()}} implements forward simulation-based reconstruction of movement paths;
\item \code{\link[=pf_backward]{pf_backward()}} implements backward simulation-based reconstruction of movement paths;
}
}
\author{
Edward Lavender
}
