% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spats-internal.R
\name{spat}
\alias{spat}
\alias{spatContainsNA}
\alias{spatSampleDT}
\alias{spatIntersect}
\alias{spatIntersect.SpatRaster}
\alias{spatIntersect.SpatVector}
\alias{spatIsEmpty}
\alias{spatNormalise}
\title{Spatial helper: \verb{spat*} functions}
\usage{
spatContainsNA(.x)

spatSampleDT(.x, .spatcell = .x, .size = 1e+06, .method = "random")

spatIntersect(.x, .value = 1, .fun = NULL)

spatIntersect.SpatRaster(.x, .value, .fun)

spatIntersect.SpatVector(.x)

spatIsEmpty(.x)

spatNormalise(x)
}
\description{
Internal \code{terra} helpers for \href{s}{\code{SpatRaster}} and \href{s}{\code{SpatVector}}.
}
\keyword{internal}
