% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers.R
\name{spatIntersect}
\alias{spatIntersect}
\title{Spatial helper: intersect \code{SpatRaster}s in a list}
\usage{
spatIntersect(.x, .value = 1, .fun = NULL)
}
\arguments{
\item{.x}{A list.}

\item{.value}{A number that defines the value.}

\item{.fun}{A function.}
}
\description{
This function identifies the cells on a \code{SpatRaster}s where all \code{SpatRaster} layers in a \code{list} have the same value.
}
\keyword{internal}
