% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spats.R
\name{spatNormalise}
\alias{spatNormalise}
\title{Spatial helper: normalise \code{\link{SpatRaster}}s}
\usage{
spatNormalise(x)
}
\arguments{
\item{x}{A \code{\link{SpatRaster}}.}
}
\description{
This function normalises a \code{\link{SpatRaster}}.
}
\section{Warning}{
\itemize{
\item This function does not implement any internal checks.
\item \code{NA}s are ignored.
}
}

\author{
Edward Lavender
}
