% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spats.R
\name{spatTemplate}
\alias{spatTemplate}
\title{Spatial helper: template \code{\link{SpatRaster}}s}
\usage{
spatTemplate(
  .xmin = 0,
  .xmax = 1000,
  .ymin = 0,
  .ymax = 1000,
  .res = c(10, 10),
  .crs = "+proj=utm +zone=1 +datum=WGS84",
  .value = 0,
  .units = "m",
  ...
)
}
\arguments{
\item{.xmin, .xmax, .ymin, .ymax}{Numbers that define the \code{\link{SpatRaster}}'s extent.}

\item{.res}{A \code{numeric} vector that defines the \code{\link{SpatRaster}}'s resolution.}

\item{.crs}{A \code{character} that defines the Coordinate Reference System.}

\item{.value}{The value of the \code{\link{SpatRaster}} cells.}

\item{.units}{A \code{character} that defines the units.}

\item{...}{Additional arguments passed to \code{\link{SpatRaster}}.}
}
\value{
The function returns a \code{\link{SpatRaster}}.
}
\description{
This function creates a template \code{\link{SpatRaster}}.
}
\examples{
r <- spatTemplate()
terra::plot(r)
r <- spatTemplate(.value = 1)
terra::plot(r)
}
\author{
Edward Lavender
}
