% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats-patter.R
\name{unwrap_elm}
\alias{unwrap_elm}
\title{Unwrap \code{\link{SpatRaster}}s \code{list} elements}
\usage{
unwrap_elm(e)
}
\description{
Unwrap \code{\link{SpatRaster}}s \code{list} elements
}
\keyword{internal}
