% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{utils-lists}
\alias{utils-lists}
\alias{list_compact}
\alias{list_merge}
\title{Utilities: list helpers}
\usage{
list_compact(l)

list_merge(...)
}
\description{
These are internal list helpers in \code{\link{patter}}.
}
\author{
Edward Lavender
}
\keyword{internal}
