% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{utils-logs}
\alias{utils-logs}
\alias{create_log}
\alias{cat_helper}
\title{Utilities: logs}
\usage{
create_log(.file, .verbose)

cat_helper(.verbose, .txt)
}
\description{
Utilities: logs
}
\keyword{internal}
