% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{utils-signal}
\alias{utils-signal}
\alias{msg}
\alias{warn}
\alias{abort}
\title{Utilities: signal messages, warnings or errors}
\usage{
msg(...)

warn(...)

abort(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[glue:glue]{glue::glue()}}.}
}
\value{
Returned values follow parent functions.
}
\description{
These functions are wrappers for \code{\link[=message]{message()}}, \code{\link[=warning]{warning()}} and \code{\link[=stop]{stop()}}.
}
\details{
\itemize{
\item \code{\link[=msg]{msg()}} is a \code{\link[=message]{message()}} wrapper;
\item \code{\link[=warn]{warn()}} is a \code{\link[=warning]{warning()}} wrapper for immediate, clean warnings;
\item \code{\link[=abort]{abort()}} is a \code{\link[=stop]{stop()}} wrapper for clean errors;
}
}
\author{
Edward Lavender
}
\keyword{internal}
