% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{utils-stats}
\alias{utils-stats}
\alias{normalise}
\alias{lognormalise}
\alias{colProds.matrix}
\title{Utilities: statistics helpers}
\usage{
normalise(x)

lognormalise(x)

colProds.matrix(.data, .cols = seq_col(.data))
}
\description{
These are internal helpers for statistical calculations in \code{\link{patter}}.
}
\author{
Edward Lavender
}
\keyword{internal}
