% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-checks.R
\name{acs_check}
\alias{acs_check}
\alias{.acs_check_obs}
\alias{.acs_check_bathy}
\alias{.acs_check_detection_overlaps}
\alias{.acs_check_detection_kernels}
\alias{.acs_check_write_record}
\alias{.acs_check_present}
\title{AC* checks}
\usage{
.acs_check_obs(.obs)

.acs_check_bathy(.bathy)

.acs_check_detection_overlaps(.detection_overlaps)

.acs_check_detection_kernels(.detection_kernels, .bathy)

.acs_check_write_record(.write_record, .element = "filename")

.acs_check_present(.p, .t, .type = c("acs", "dc"))
}
\description{
These are internal check functions.
}
\author{
Edward Lavender
}
\keyword{internal}
