% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internals.R
\name{acs_check}
\alias{acs_check}
\alias{.acs_check_obs}
\alias{.acs_check_bathy}
\alias{.acs_check_detection_overlaps}
\alias{.acs_check_detection_kernels}
\alias{.acs_check_write_record}
\title{AC* helper: internal checks}
\usage{
.acs_check_obs(.o)

.acs_check_bathy(.b)

.acs_check_detection_overlaps(.ov)

.acs_check_detection_kernels(.k, .b)

.acs_check_write_record(.w)
}
\arguments{
\item{.o}{The \code{.obs} \code{\link{data.table}}.}

\item{.b}{The \code{.bathy} \code{\link{SpatRaster}}.}

\item{.ov}{The \code{detection_overlaps} \code{list}.}

\item{.k}{The \code{.detection_kernels} \code{list}.}

\item{.w}{The \code{write_record} \code{list}.}
}
\description{
These are internal check functions.
}
\author{
Edward Lavender
}
\keyword{internal}
