% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internals.R
\name{acs_container}
\alias{acs_container}
\alias{.acs_container}
\alias{.acs_container_1}
\title{AC* helper: define AC* container(s)}
\usage{
.acs_container(
  .moorings,
  .coords = c("receiver_x", "receiver_y"),
  .receivers,
  .bathy,
  .buffer
)

.acs_container_1(.obs, .moorings, .bathy)
}
\arguments{
\item{.moorings}{A \code{\link{data.table}} that defines receiver locations.}

\item{.coords}{A character vector that defines the name of the columns in \code{.moorings} that define receiver coordinates.}

\item{.receivers}{An \code{integer} vector of receivers for which to define containers, used to subset \code{.moorings}.}

\item{.bathy}{The bathymetry \code{\link{SpatRaster}}.}

\item{.buffer}{A \code{numeric} value that defines the container radius.}

\item{.obs}{The \code{.obs} \code{\link{data.table}}.}
}
\value{
The function returns an \code{\link[sf:sf]{sf::sf}} object (the acoustic container).
}
\description{
These functions define AC* containers in \code{\link[=pf_forward_2]{pf_forward_2()}}.
}
\details{
In \code{\link[=acs]{acs()}}, acoustic containers are defined using \code{\link[terra:buffer]{terra::buffer()}}.

These functions are used in \code{\link[=pf_forward_2]{pf_forward_2()}}.
\itemize{
\item \code{\link[=.acs_container]{.acs_container()}} defines the acoustic container around a particular set of receivers.
\item \code{\link[=.acs_container_1]{.acs_container_1()}} defines the acoustic container at the first time step, which is the intersection between the container(s) around the receivers that recorded the first detection (if applicable) and the container(s) around the receivers that recorded the next detection.
}
}
\author{
Edward Lavender
}
\keyword{internal}
