% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internal.R
\name{acs_container}
\alias{acs_container}
\alias{.acs_container_detection.SpatRaster}
\alias{.acs_container_detection.SpatVector}
\alias{.acs_container_detection}
\alias{.acs_container_future.SpatRaster}
\alias{.acs_container_future.SpatVector}
\alias{.acs_container_future}
\alias{.acs_container_1}
\title{AC* helper: define AC* container(s)}
\usage{
.acs_container_detection.SpatRaster(.receivers, .detection_kernels)

.acs_container_detection.SpatVector(.receivers, .moorings)

.acs_container_detection(.receivers, .detection_kernels, .moorings)

.acs_container_future.SpatRaster(.receivers, .detection_kernels, .buffer)

.acs_container_future.SpatVector(.receivers, .moorings, .buffer)

.acs_container_future(.receivers, .detection_kernels, .moorings, .buffer)

.acs_container_1(.obs, .detection_kernels, .moorings)
}
\arguments{
\item{.detection_kernels}{A \code{\link{list}} of detection kernels.}

\item{.moorings}{The \code{.moorings} \code{\link{data.table}}, including \code{receiver_x} and \code{receiver_y} columns.}

\item{.obs}{The \code{.obs} \code{\link{data.table}}.}
}
\value{
The function returns a \code{\link{SpatRaster}} or a \code{\link{SpatVector}}.
}
\description{
\code{\link[=.acs_container_1]{.acs_container_1()}} defines the first AC* container in \code{\link[=pf_forward]{pf_forward()}} (specifically \code{\link[=.pf_rpropose_origin]{.pf_rpropose_origin()}}.
}
\details{
These functions are used in \code{\link[=pf_forward]{pf_forward()}}. \code{\link[=.acs_container_1]{.acs_container_1()}} defines the acoustic container at the first time step, which is the intersection between the container(s) around the receivers that recorded the first detection (if applicable) and the container(s) around the receivers that recorded the next detection.
}
\author{
Edward Lavender
}
\keyword{internal}
