% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_detection_overlaps}
\alias{acs_setup_detection_overlaps}
\title{AC* set up: define detection container overlaps}
\source{
This function supersedes the \href{https://edwardlavender.github.io/flapper/reference/get_detection_containers_overlap.html}{\code{get_detection_containers_overlaps}} function in the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package.
}
\usage{
acs_setup_detection_overlaps(.dlist)
}
\arguments{
\item{.dlist}{A named \code{list} of data and parameters from \code{\link[=pat_setup_data]{pat_setup_data()}}. This function requires:
\itemize{
\item \code{.dlist$data$moorings}, with the following columns: \code{receiver_id}, \code{receiver_x}, \code{receiver_y}, \code{receiver_start}, \code{receiver_end} and \code{receiver_range}.
\item (optional) \code{.dlist$data$services}, with the following columns: \code{receiver_id}, \code{service_start} and \code{service_end}.
}}
}
\value{
The function returns a nested \code{list}, with one element for all integers from \code{1:max(.moorings$receiver_id)}. Any elements that do not correspond to receivers contain a \code{NULL} element. List elements that correspond to receivers contain a \code{NULL} or a \code{list} that defines, for each deployment date with overlapping receiver(s), a vector of overlapping receiver(s).
}
\description{
This function identifies receivers with overlapping detection containers in space and time for the AC* algorithms.
}
\details{
In the AC* algorithms, at the moment of detection, the set of possible locations depends on the receiver(s) at which an individual is, and is not, detected. The outputs of this function are used to restrict the probability calculations to the set of receivers that overlap with the receiver(s) at which an individual is detected for improved efficiency.
}
\examples{
#### Example (1): Basic implementation
dlist <- pat_setup_data(.moorings = dat_moorings,
                       .bathy = dat_gebco(),
                       .lonlat = FALSE)
overlaps <- acs_setup_detection_overlaps(dlist)

}
\author{
Edward Lavender
}
