% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_detection_overlaps}
\alias{acs_setup_detection_overlaps}
\title{AC* set up: detection container overlaps}
\usage{
acs_setup_detection_overlaps(.dlist)
}
\arguments{
\item{.dlist}{A named \code{list} of data and parameters from \code{\link[=pat_setup_data]{pat_setup_data()}}. This function requires:
\itemize{
\item \code{.dlist$data$moorings}, with the following columns: \code{receiver_id}, \code{receiver_x}, \code{receiver_y}, \code{receiver_start}, \code{receiver_end} and \code{receiver_range}.
\item (optional) \code{.dlist$data$services}, with the following columns: \code{receiver_id}, \code{service_start} and \code{service_end}.
}}
}
\value{
The function returns a nested \code{list}, with one element for all integers from \code{1:max(.moorings$receiver_id)}. Any elements that do not correspond to receivers contain a \code{NULL} element. List elements that correspond to receivers contain a \code{NULL} or a \code{list} that defines, for each deployment date with overlapping receiver(s), a vector of overlapping receiver(s).
}
\description{
This function is part of a set of \verb{acs_setup_*()} functions that prepare the layers required to evaluate the likelihood of acoustic observations in an AC*PF algorithm using the inbuilt functions. The function identifies receivers with overlapping detection containers in space and time.
}
\details{
An AC*PF algorithm is a particle filtering algorithm that incorporates acoustic observations to reconstruct the possible movements of an individual (see \code{\link[=pf_forward]{pf_forward()}}). \code{\link[=pf_setup_obs]{pf_setup_obs()}} prepares the timeline of observations for particle filtering. Information from \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}} is used in this routine to assemble the acoustic time series: at each time step with a detection, we define an observation matrix (0, 1) comprising the subset of receiver(s) that recorded detections and the overlapping receiver(s). In \code{\link[=pf_forward]{pf_forward()}}, the likelihood of each acoustic observation is evaluated, given particle samples. Thanks to \code{\link[=acs_setup_detection_containers]{acs_setup_detection_containers()}}, at the moment of detection, likelihood calculations are restricted to the set of receivers that overlap with the receiver(s) at which an individual is detected. This improves efficiency.

In \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}}, receiver deployments that overlap in time (accounting for deployment and servicing dates) are identified by \code{\link[=make_matrix_receivers]{make_matrix_receivers()}}. Spatially overlapping receiver pairs are defined as those for which the Euclidean distance between receiver coordinates is less than the combined detection range. This approach is fast but crude because it ignores the influence of other variables, such as land barriers, on detectability. This means that some 'overlapping receivers' may not in reality overlap. In this situation, downstream calculations may be a little less efficient. However, since overlapping receivers tend to be few in number, the efficiency penalty for this approximation should be negligible. We formerly used detection kernels (see \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}) to identify receiver overlaps, but this is much more expensive in situations with large numbers of receivers and high-resolution grids and this approach is no longer used.

This function replaces \href{https://edwardlavender.github.io/flapper/reference/get_detection_containers_overlap.html}{\code{flapper::get_detection_containers_overlaps()}}.
}
\examples{
# Set up data list(see `?pat_setup_data()`)
dlist <- dat_dlist()

# Prepare detection overlaps
# * Store detection overlaps in `dlist$algorithms$detection_overlaps`
# * It is expected in this slot by `pf_setup_obs()`
dlist$algorithm$detection_overlaps <- acs_setup_detection_overlaps(dlist)

}
\seealso{
To implement an AC*PF algorithm, see:
\enumerate{
\item \code{\link[=pat_setup_data]{pat_setup_data()}} to set up datasets;
\item \verb{acs_setup_*()} functions to prepare layers required for likelihood calculations, i.e.:
\itemize{
\item \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}}, which identifies detection overlaps;
\item \code{\link[=acs_setup_detection_kernel]{acs_setup_detection_kernel()}}, which prepares a detection kernel;
\item \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}, which prepares detection kernels;
}
\item \code{\link[=pf_setup_obs]{pf_setup_obs()}} to define a timeline of observations;
\item \code{\link[=pf_forward]{pf_forward()}} to run the simulation (using \code{\link[=pf_lik_ac]{pf_lik_ac()}} to evaluate the likelihood of acoustic observations);
}
}
\author{
Edward Lavender
}
