% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_detection_pr}
\alias{acs_setup_detection_pr}
\title{AC* set up: calculate detection probability around a receiver}
\usage{
acs_setup_detection_pr(
  .mooring,
  .bathy,
  .calc_detection_pr = calc_detection_pr_logistic,
  ...
)
}
\arguments{
\item{.mooring}{A one-row \code{\link{data.table}} that defines the location of the receiver and associated information used by the model of detection probability.}

\item{.bathy}{A \code{\link{SpatRaster}} that defines the grid over which detection probability is calculated.}

\item{.calc_detection_pr}{A function that calculates detection probability. In this implementation, the function is used to translate a \code{\link{SpatRaster}} of distances (m) (from each grid cell to the receiver in \code{.data}) via \code{\link[terra:app]{terra::app()}}.}

\item{...}{Additional arguments passed to \code{.calc_detection_pr} (\code{\link[=calc_detection_pr_logistic]{calc_detection_pr_logistic()}}  by default.)}
}
\value{
The function returns a \code{\link{SpatRaster}} that defines the probability of detection in each cell, according to a pre-defined model.
}
\description{
This function is an example detection probability function, of the kind required by \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}.
}
\details{
In the AC* algorithms, a model of the detection process informs the set of possible locations for an individual. The information provided by this model is represented in the form of kernels, which are created via \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}. For any one receiver, the form of the kernel depends on the input to \code{.calc_detection_pr}. This function exemplifies one possible input to this argument, which is a model in which detection probability declines logistically with distance from a receiver.
}
\section{Warning}{
\itemize{
\item This function is used to streamline examples and does not represent a generically suitable detection probability model.
\item The function does not check user inputs.
}
}

\examples{
data <- pat_setup_data(.moorings = moorings,
                       .bathy = dat_gebco(),
                       .lonlat = FALSE)
m <- data$data$moorings[1, ]
b <- data$spatial$bathy
p <- acs_setup_detection_pr(m, n)
terra::plot(p)
points(m$receiver_x, m$receiver_y, pch = ".")

}
\seealso{
To implement an AC*PF algorithm, use:
\enumerate{
\item \code{\link[=acs_setup_obs]{acs_setup_obs()}} to set up observations;
\item \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}} to identify receiver overlaps (used in detection probability calculations);
\item \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}} and \code{\link[=acs_setup_detection_pr]{acs_setup_detection_pr()}} to define detection probability kernels;
\item \code{\link[=pf_forward]{pf_forward()}} to run the simulation;
}
}
\author{
Edward Lavender
}
