% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-setup.R
\name{acs_setup_obs}
\alias{acs_setup_obs}
\title{AC* set up: set up movement datasets}
\usage{
acs_setup_obs(
  .acoustics = NULL,
  .archival = NULL,
  .trim = TRUE,
  .step,
  .period = NULL,
  .mobility,
  .detection_range
)
}
\arguments{
\item{.acoustics}{(optional) A \code{\link{data.table}} that defines passive acoustic telemetry detections (see \code{\link{dat_acoustics}} for an example) for a single individual. At a minimum, this must contain two columns:
\itemize{
\item \code{timestamp}---an ordered, \code{POSIXct} vector that defines the times of detections;
\item \code{receiver_id}---an \code{integer} vector that defines the receiver(s) that recorded detections;
}}

\item{.archival}{(optional) A \code{\link{data.table}} that defines depth (m) observations (see \code{\link{dat_archival}} for an example) for the same individual. At a minimum, this must contain two columns:
\itemize{
\item \code{timestamp}---an ordered, \code{POSIXct} vector that defines the times of observations (which may be regular or irregular);
\item \code{depth}---a positive-valued \code{numeric} vector that defines the individual's depth (m) below the surface at each time step;
}}

\item{.trim}{If \code{.archival} is supplied, \code{.trim} is \code{logical} variable that defines whether or not to trim \code{.acoustic} and \code{.archival} time series to the time period for which they overlap.}

\item{.step}{An character, passed to \code{\link[lubridate:period]{lubridate::period()}}, \code{\link[lubridate:round_date]{lubridate::round_date()}} and \code{\link[=seq]{seq()}} that defines the duration between sequential time steps (e.g., \code{"2 mins"}).}

\item{.period}{(optional) A length-two \code{POSIXct} vector that defines the start and end time of the observations. If unprovided, this is defined internally according to \code{.acoustics}, \code{.archival} and \code{.trim}.}

\item{.mobility}{A constant that defines the maximum (Euclidean) distance the individual could move in \code{.step}.}

\item{.detection_range}{A constant that defines the detection range. A constant value across all receivers and time steps is assumed.}
}
\value{
The function returns a \code{\link{data.table}} with the following columns:
\itemize{
\item \code{timestep}---an \code{integer} that defines the time step;
\item \code{timestamp}---a regular sequences of \code{POSIXct} time stamps;
\item \code{date}---a \code{character} that defines the date;
\item \code{mobility}---a \code{double} that defines \code{.mobility};
}

If \code{.acoustics} is provided, the following column(s) are also included:
\itemize{
\item \code{detection_id}---an \code{integer} vector that uniquely defines each detection;
\item \code{detection}---an \code{integer} that distinguishes the time steps at which detections were (1) or were not (0) recorded;
\item \code{receiver_id}---a \code{list} that defines the receiver(s) that recorded detection(s) at each time step;
\item \code{receiver_id_next}---a \code{list} that defines the receiver(s) that recorded the next detection(s);
\item \code{buffer_past}---a \code{double} that controls container growth from the past to the present;
\item \code{buffer_future}---a \code{double} that controls container shrinkage from the future to the present;
\item \code{buffer_future_incl_gamma}---a \code{double} (\code{buffer_future} + \code{.detection_range}) that controls container shrinkage for \code{\link[=pf_forward]{pf_forward()}};
}

If \code{.archival} is provided, the following column(s) are also included:
\itemize{
\item \code{depth}---a number that defines the individual's depth (m) at each time step;
}
}
\description{
This function processes passive acoustic telemetry detections and/or archival time series for \code{\link[=pf_forward]{pf_forward()}}.
}
\details{
This function implements the following routines:
\itemize{
\item Acoustic progressing (if applicable);
\itemize{
\item Acoustic time series are rounded to the nearest \code{.step};
\item All receivers that recorded detection(s) in each time interval are listed;
\item Duplicate detections (at the same receiver in the same time interval) are dropped;
}
\item Archival processing (if applicable);
\itemize{
\item Archival time series are rounded to the nearest \code{.step};
\item Duplicate records (in the same time interval) are not permitted;
}
\item Alignment (if applicable)
\itemize{
\item Acoustic and archival time series are optionally trimmed to the time period for which they overlap;;
}
\item Time series
\itemize{
\item Acoustic and/or archival time series are mapped onto a regular time series;
\item Information required by \code{\link[=pf_forward_2]{pf_forward_2()}} is added
}
}
}
\examples{
#### Define example datasets
acoustics <- dat_acoustics[dat_acoustics$individual_id == 25, ]
archival  <- dat_archival[dat_archival$individual_id == 25, ]

#### Example (1): Implement the function for acoustic time series only
obs <- acs_setup_obs(acoustics,
                     .step = "2 mins",
                     .mobility = 500,
                     .detection_range = 750)
utils::head(obs)

#### Example (2): Use alternative parameters
obs <- acs_setup_obs(acoustics,
                     .step = "4 mins",
                     .mobility = 1000,
                     .detection_range = 500)
utils::head(obs)

#### Example (3): Implement the function for acoustic & archival time series
obs <- acs_setup_obs(acoustics, archival,
                     .step = "2 mins",
                     .mobility = 500,
                     .detection_range = 750)
utils::head(obs)

#### Example (4): Include full acoustic & archival time series
range(acoustics$timestamp)
range(archival$timestamp)
obs <- acs_setup_obs(acoustics, archival, .trim = FALSE,
                     .step = "2 mins",
                     .mobility = 500,
                     .detection_range = 750)
utils::head(obs)
utils::tail(obs)

}
\seealso{
To implement an AC*PF algorithm, use:
\enumerate{
\item \code{\link[=acs_setup_obs]{acs_setup_obs()}} to set up observations;
\item \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}} to identify receiver overlaps (used in detection probability calculations);
\item \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}} and \code{\link[=acs_setup_detection_pr]{acs_setup_detection_pr()}} to define detection probability kernels;
\item \code{\link[=pf_forward]{pf_forward()}} to run the simulation;
}

For \code{\link[=acs_setup_obs]{acs_setup_obs()}} specifically:
\itemize{
\item \href{https://edwardlavender.github.io/flapper/reference/process_receiver_id.html}{\code{process_receiver_ids}} in the \href{https://edwardlavender.github.io/flapper/reference/process_receiver_id.html}{\code{flapper}} package can be used to define receiver deployments using an integer vector.
}
}
\author{
Edward Lavender
}
