% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calls-internal.R
\name{call_}
\alias{call_}
\alias{call_time}
\alias{call_start}
\alias{call_end}
\alias{call_timings}
\title{Utilities: function call properties}
\usage{
call_time(.time, form = "\%Y-\%m-\%d \%H:\%M:\%S")

call_start(.fun = as.character(sys.call(-1L)), .start = Sys.time())

call_end(.fun = as.character(sys.call(-1L)), .start, .end = Sys.time())

call_timings(.start, .end = Sys.time(), ...)
}
\description{
These functions collate information related to function calls.
}
\keyword{internal}
