% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{cat_}
\alias{cat_}
\alias{cat_log_file}
\alias{cat_init}
\title{Utilities: \verb{cat_*()} functions}
\usage{
cat_log_file(.verbose)

cat_init(.verbose)
}
\arguments{
\item{.verbose}{A \code{logical} variable or a \code{string} that defines the path to a text file.
\itemize{
\item \code{.verbose = FALSE} suppresses user outputs;
\item \code{.verbose = TRUE} sends user outputs to the console;
\item \code{.verbose = file.path("path/to/text/file.txt")} sends user outputs to a \code{.txt} file;
}}
}
\description{
\verb{cat_*()} functions control user output in \code{\link{patter}} functions with the \code{.verbose} argument.
}
\details{
These are internal functions.
\itemize{
\item \code{\link[=cat_log_file]{cat_log_file()}} validates \code{.verbose} and, if necessary, creates the \code{.txt} file;
\item \code{\link[=cat_init]{cat_init()}} wraps \code{\link[=cat_log_file]{cat_log_file()}} and defines an appropriate \code{\link[=cat]{cat()}} wrapper for use in \code{\link{patter}} function(s) according to the input to \code{.verbose};
}
}
\examples{
\dontrun{
# Define example function
wrap <- function(.verbose) {
  cat_log <- cat_init(.verbose)
  cat_log("Hello world")
  invisible(NULL)
}

# `.verbose = TRUE` sends user output to the console
wrap(.verbose = TRUE)

# `.verbose = {log}.txt` sends user output to file
log.txt <- tempfile(fileext = ".txt")
wrap(.verbose = log.txt)
readLines(log.txt)
unlink(log.txt)

# `.verbose = TRUE` suppresses user output
wrap(.verbose = FALSE)
}
}
\author{
Edward Lavender
}
\keyword{internal}
