% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_data}
\alias{check_data}
\alias{check_moorings}
\alias{check_acoustics}
\alias{check_archival}
\title{Check input datasets}
\usage{
check_moorings(.moorings)

check_acoustics(.acoustics)

check_archival(.archival)
}
\arguments{
\item{acoustics}{A data.table.}

\item{archival}{A data.table or \code{NULL}.}
}
\description{
These functions check input datasets.
}
\keyword{internal}
