% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_data}
\alias{check_data}
\alias{check_moorings}
\alias{check_services}
\alias{check_acoustics}
\alias{check_archival}
\title{Check input datasets}
\usage{
check_moorings(.moorings)

check_services(.services, .moorings)

check_acoustics(.acoustics)

check_archival(.archival)
}
\arguments{
\item{.moorings, .services, .acoustics, .archival}{User inputs. \code{NULL} inputs are allowed for \code{.services} and \code{.archival}.}
}
\value{
The functions invisibly return the inputted object if all checks are passed, with the following adjustments:
\itemize{
\item \code{.moorings$receiver_id} is silently coerced to an integer if numeric;
}
}
\description{
These functions check input datasets.
}
\author{
Edward Lavender
}
\keyword{internal}
