% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_setup_data-internal.R
\name{check_dlist}
\alias{check_dlist}
\alias{check_map}
\alias{check_detections}
\alias{check_moorings}
\alias{check_services}
\alias{check_archival}
\title{Utilities: check data}
\usage{
check_map(.map)

check_detections(.detections, .moorings = NULL)

check_moorings(.moorings, .detections = NULL, .map)

check_services(.services, .moorings)

check_archival(.archival)
}
\arguments{
\item{.map}{A \code{\link{SpatRaster}} that defines the study area for the simulation (see \code{\link{glossary}}).}

\item{.detections, .moorings, .services, .archival}{Observation \code{\link{data.table}}s.}
}
\value{
Each function returns the input dataset, as described, formatted in line with \code{\link{patter}}'s requirements.
}
\description{
These functions validate user datasets. See Details for \code{\link{patter}} requirements.
}
\details{
These are internal functions that ensure that input dataset(s) meet \code{\link{patter}} requirements. Users with acoustic and/or archival data should implement \code{\link[=pat_setup_data]{pat_setup_data()}} to prepare datasets for \code{\link{patter}} functions. Downstream functions (often) silently assume that input datasets meet all requirements, without subsequent checks. This simplifies internal code and documentation.
\itemize{
\item \code{\link[=check_map]{check_map()}} checks the \code{.map} \code{\link{SpatRaster}}:
\itemize{
\item Class: \code{\link{SpatRaster}};
\item Name(s): \code{"map_value"};
\item Properties:
\itemize{
\item A square grid is recommended;
\item Absolute values (m) are recommended;
\item A planar grid with coordinates in metres is currently required;
}
\item Source:
\itemize{
\item The data source may be in memory or on disk, but the former is faster for pure \code{R} functions;
}
}
\item \code{\link[=check_detections]{check_detections()}} checks detection time series:
\itemize{
\item Class: \code{\link{data.table}};
\item Columns:
\itemize{
\item \code{timestamp}---an ordered, \code{POSIXct} vector of time stamps with a defined \code{tzone} attribute;
\item \code{receiver_id}---see \code{.moorings$receiver_id};
}
\item Properties:
\itemize{
\item Each row defines a detection event;
\item Functions assume accurate clock alignment between receivers;
\item For most functions, a single time series (for one individual) is required;
}
}
\item \code{\link[=check_services]{check_services()}} checks acoustic servicing records:
\itemize{
\item Class: \code{\link{data.table}};
\item Columns:
\itemize{
\item \code{receiver_id}---see \code{.moorings$receiver_id};
\item \code{receiver_start}, \code{receiver_end}---\code{POSIXct} vectors of service start and end dates. Before/after service events, receivers are assumed to have been deployed in the same locations; receiver deployments in different locations before/after servicing should be treated as distinct deployments in \code{.moorings}.
}
\item Properties:
\itemize{
\item Serviced receivers should be found in the \code{.moorings} dataset;
}
}
\item \code{\link[=check_moorings]{check_moorings()}} checks receiver deployment records:
\itemize{
\item Class: \code{\link{data.table}};
\item Columns:
\itemize{
\item \code{receiver_id}---an \code{integer} vector of unique receiver deployments (\verb{1, ..., N});
\item \code{receiver_x} and \verb{receiver_y``---Receiver coordinates on a planar projection, matching }.map`;
}
\item Properties:
\itemize{
\item \code{receiver_start} should precede \code{receiver_end};
\item Receiver deployment periods should (usually) at least partially overlap with the time period of acoustic detections;
\item Receiver coordinates must be valid on \code{.map} (not currently validated);
}
}
\item \code{\link[=check_archival]{check_archival()}} checks archival depth time series data:
\itemize{
\item Class: \code{\link{data.table}};
\item Columns:
\itemize{
\item \code{timestamp}---an ordered, \code{POSIXct} vector of time stamps with a defined \code{tzone} attribute;
\item \code{depth}---a \code{numeric} vector of (absolute) depths (m);
}
\item Properties:
\itemize{
\item Functions assume accurate clock alignment between acoustic and archival tags (if applicable);
\item For most functions, a single time series (for one individual) is required;
}
}
}

For all datasets, \code{NA}s may cause unexpected errors and produce a \code{\link{message}}.

\code{NULL} inputs are permitted in all \verb{check_\{data\}} functions, in which case \code{NULL} is returned.
}
\seealso{
\itemize{
\item See \href{https://edwardlavender.github.io/flapper/reference/process_receiver_id.html}{\code{flapper::process_receiver_ids()}} to define unique receiver IDs;
\item See \code{\link{datasets-mefs}} for example datasets included in \code{\link{patter}};
\item See \code{\link[=pat_setup_data]{pat_setup_data()}} for the front-end function;
}
}
\author{
Edward Lavender
}
\keyword{internal}
