% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_utils}
\alias{check_utils}
\alias{check_dir}
\alias{check_inherits}
\alias{check_names}
\alias{check_named_list}
\alias{check_tz}
\alias{check_dots_allowed}
\alias{check_dots_for_missing_period}
\title{Utility checks}
\usage{
check_dir(input)

check_inherits(input, class)

check_names(
  input,
  req,
  extract_names = names,
  type = all,
  arg = deparse(substitute(input))
)

check_named_list(input, ignore_empty = TRUE, arg = deparse(substitute(input)))

check_tz(input, arg = deparse(substitute(input)))

check_dots_allowed(not_allowed, ...)

check_dots_for_missing_period(args, dots)
}
\description{
These functions check user inputs.
}
\author{
Edward Lavender
}
\keyword{internal}
