% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_utils}
\alias{check_utils}
\alias{check_dir}
\alias{check_dir_is_empty}
\alias{check_contents_ext}
\alias{check_inherits}
\alias{check_names}
\alias{check_named_list}
\alias{check_tz}
\alias{check_dots_allowed}
\alias{check_dots_for_missing_period}
\alias{check_new_colnames}
\alias{check_verbose_and_log}
\title{Utilities: check functions}
\usage{
check_dir(input)

check_dir_is_empty(input)

check_contents_ext(input, ext, ...)

check_inherits(input, class)

check_names(
  input,
  req,
  extract_names = names,
  type = all,
  error = abort,
  arg = deparse(substitute(input))
)

check_named_list(input, ignore_empty = TRUE, arg = deparse(substitute(input)))

check_tz(input, arg = deparse(substitute(input)))

check_dots_allowed(not_allowed, ...)

check_dots_for_missing_period(args, dots)

check_new_colnames(.data, .new)

check_verbose_and_log(.verbose, .txt)
}
\description{
These functions check user inputs.
}
\author{
Edward Lavender
}
\keyword{internal}
