% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{cleanup}
\alias{cleanup}
\title{Utilities: clean up}
\usage{
cleanup(x)
}
\description{
This function deletes temporary files and or directories recursively.
}
\details{
This function is a simple wrapper for \code{\link{unlink}} that is used at the end of examples.
}
\author{
Edward Lavender
}
