% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coa.R
\name{coa}
\alias{coa}
\title{The centres of activity (COA) algorithm}
\usage{
coa(
  .data,
  .delta_t,
  .split = NULL,
  .plot_weights = TRUE,
  ...,
  .one_page = TRUE
)
}
\arguments{
\item{.data}{A named \code{list} of data and parameters from \code{\link[=pat_setup_data]{pat_setup_data()}}. This function requires:
\itemize{
\item \code{.data$data$acoustics}, with the following columns: \code{receiver_id} and \code{timestamp};
\item \code{.data$data$moorings}, with the following columns: \code{receiver_id}, \code{receiver_x} and  \code{receiver_y};
\item \code{.data$pars$lonlat}, which specifies the coordin
}}

\item{.delta_t}{The time interval over which to calculate COAs. This can be specified in any way understood by \code{\link[=cut.POSIXt]{cut.POSIXt()}} (see the \code{breaks} argument).}

\item{.split}{(optional) A \code{character} that defines the name of the grouping factor in \code{.data$data$acoustics} (e.g., \code{individual_id} for \code{\link{dat_acoustics}}).}

\item{.plot_weights}{A \code{logical} variable that defines whether or not to plot the frequency distribution of weights for each \code{.split} value.}

\item{...}{Additional arguments passed to \code{\link[graphics:hist]{graphics::hist()}}.}

\item{.one_page}{A \code{logical} variable that defines whether or not to plot all histograms on one page.}
}
\description{
This function calculates centres of activity (COAs).
}
\details{
COAs are calculated as a weighted mean of the locations of receivers at which individuals are detected over consecutive time intervals, weighted by the frequency of detections at each of those receivers. COAs are calculated via \code{\link[stats:weighted.mean]{stats::weighted.mean()}} (for planar coordinates) or \code{\link[geosphere:geomean]{geosphere::geomean()}} (for longitude/latitude coordinates).
}
\examples{
require(data.table)

#### Example (1): Calculate COAs for an example individual
acc   <- dat_acoustics[individual_id == dat_acoustics$individual_id[1], ]
gebco <- dat_gebco()
data  <- pat_setup_data(.acoustics = acc,
                        .moorings = dat_moorings,
                        .bathy = gebco,
                        .lonlat = FALSE)
coa(data, .delta_t = "2 hours")
coa(data, .delta_t = "4 hours")

#### Example (2): Calculate COAs for multiple individuals via .split
data <- pat_setup_data(.acoustics = dat_acoustics,
                       .moorings = dat_moorings,
                       .bathy = gebco,
                       .lonlat = FALSE)
coa(data, .delta_t = "6 hours", .split = "individual_id")

#### Example (3): Use lon/lat coordinates
gebco_ll <- terra::project(dat_gebco(), "EPSG:4326")
data <- pat_setup_data(.acoustics = dat_acoustics,
                       .moorings = dat_moorings,
                       .bathy = gebco_ll,
                       .lonlat = TRUE)
coa(data, .delta_t = "6 hours", .split = "individual_id")

}
\seealso{
\itemize{
\item For reconstructing movement paths and patterns of space use, see \code{\link[=pf_forward]{pf_forward()}};
\item For mapping utilisation distributions, see \code{\link[=pf_map_dens]{pf_map_dens()}};
}
}
\author{
Edward Lavender
}
