% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coa.R
\name{coa}
\alias{coa}
\title{COA: centres of activity}
\usage{
coa(
  .acoustics,
  .moorings = NULL,
  .delta_t,
  .split = NULL,
  .plot_weights = TRUE,
  ...,
  .one_page = TRUE
)
}
\arguments{
\item{.acoustics, .moorings}{Acoustic detection \href{s}{\code{data.table}}.
\itemize{
\item \code{.acoustics} is a \code{\link{data.table}} of acoustic detections, with the following columns: \code{receiver_id} (or \code{sensor_id}), \code{timestamp} and (optionally) \code{receiver_x} and \code{receiver_y} columns;
\item (optional) \code{.moorings} is a \code{\link{data.table}} of receiver coordinates, which should be provided if unavailable in \code{.acoustics}, with \code{receiver_id}, \code{receiver_x} and \code{receiver_y} columns;
}

Receiver coordinates \strong{must be planar}.}

\item{.delta_t}{The time interval over which to calculate COAs. This can be specified in any way understood by \code{\link[=cut.POSIXt]{cut.POSIXt()}} (see the \code{breaks} argument).}

\item{.split}{(optional) A \code{character} that defines the name of the grouping factor in \code{.acoustics} (e.g., \code{individual_id} for \code{\link{dat_acoustics}}).}

\item{.plot_weights, ..., .one_page}{Plot arguments.
\itemize{
\item \code{.plot_weights} is a \code{logical} variable that defines whether or not to plot the frequency distribution of weights for each \code{.split} value (i.e., the frequency distribution of the number of detections at each receiver in each time interval, excluding time intervals without detections).
\item \code{...} is a placeholder for arguments passed to \code{\link[graphics:hist]{graphics::hist()}}, excluding \code{main}.
\item \code{.one_page} A \code{logical} variable that defines whether or not to plot all histograms on one page.
}}
}
\value{
The function returns a \code{\link{data.table}} with the following columns:
\itemize{
\item \code{.split}---a \code{character} vector that distinguishes groups, if applicable;
\item \code{timestamp}---a \code{POSIXt} vector of time stamps;
\item \code{coa_x}, \code{coa_y}---the coordinates of the COAs;
}

Data are arranged by \code{.split} and \code{timestamp}.
}
\description{
This function calculates centres of activity (COAs) from detections at acoustic receivers.
}
\details{
COAs are calculated as a weighted mean of the locations of receivers at which individuals are detected over consecutive time intervals, weighted by the frequency of detections at each of those receivers. COAs are calculated via \code{\link[stats:weighted.mean]{stats::weighted.mean()}}, which assumes planar coordinates, for consistency with other \code{\link{patter}} routines. To handle longitude/latitude coordinates, use \code{geosphere::geomean()}.

This function replaces \href{https://edwardlavender.github.io/flapper/reference/coa.html}{\code{flapper::coa()}}. See  \href{https://edwardlavender.github.io/flapper/reference/coa_setup_delta_t.html}{\code{flapper::coa_setup_delta_t()}} to evaluate alternative time internals.
}
\examples{
library(data.table)

#### Set up example
acoustics <- dat_acoustics
moorings  <- dat_moorings

#### Example (1): Calculate COAs for an example individual
id     <- dat_acoustics$individual_id[1]
acc    <- dat_acoustics[individual_id == id, ]
coa(acc, moorings, .delta_t = "2 hours")

#### Example (2): Customise the time interval via `.delta_t`
coa(acc, moorings, .delta_t = "4 hours")

#### Example (3): Calculate COAs for multiple individuals via `.split`
# Calculate COAs
coa(acoustics, dat_moorings, .delta_t = "6 hours", .split = "individual_id")
# Use one page for plots via `.one_page = TRUE`
coa(acoustics, dat_moorings, .delta_t = "6 hours", .split = "individual_id",
    .one_page = FALSE)
# Suppress plots via `.plot = FALSE`
coa(acoustics, dat_moorings, .delta_t = "6 hours", .split = "individual_id",
    .one_page = TRUE)

#### Example (4): Specify `.acoustics` only
# `.moorings` is not required if `.acoustics` contains receiver coordinates
acoustics |>
  merge(moorings, by = "receiver_id") |>
  coa(.delta_t = "6 hours", .split = "individual_id")
}
\seealso{
\itemize{
\item To derive location samples from a particle filtering algorithm, see \code{\link[=pf_filter]{pf_filter()}} and associates;
\item For mapping utilisation distributions from coordinates, see \verb{map_*()} functions (i.e., \code{\link[=map_pou]{map_pou()}} and \code{\link[=map_dens]{map_dens()}});
}
}
\author{
Edward Lavender
}
