% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coa.R
\name{coa}
\alias{coa}
\title{COA: centres of activity}
\source{
This function replaces the \href{https://edwardlavender.github.io/flapper/reference/coa.html}{\code{coa()}} function in the \href{https://edwardlavender.github.io/flapper/}{\code{flapper}} package. See the \href{https://edwardlavender.github.io/flapper/reference/coa_setup_delta_t.html}{\code{coa_setup_delta_t()}} function to evaluate alternative time internals.
}
\usage{
coa(
  .dlist,
  .delta_t,
  .split = NULL,
  .plot_weights = TRUE,
  ...,
  .one_page = TRUE
)
}
\arguments{
\item{.dlist}{A named \code{list} of data and parameters from \code{\link[=pat_setup_data]{pat_setup_data()}}. This function requires:
\itemize{
\item \code{.dlist$data$acoustics}, with the following columns: \code{receiver_id} and \code{timestamp};
\item \code{.dlist$data$moorings}, with the following columns: \code{receiver_id}, \code{receiver_x} and  \code{receiver_y};
\item \code{.dlist$pars$lonlat}, which specifies the coordinate format (longitude/latitude or planar);
}}

\item{.delta_t}{The time interval over which to calculate COAs. This can be specified in any way understood by \code{\link[=cut.POSIXt]{cut.POSIXt()}} (see the \code{breaks} argument).}

\item{.split}{(optional) A \code{character} that defines the name of the grouping factor in \code{.dlist$data$acoustics} (e.g., \code{individual_id} for \code{\link{dat_acoustics}}).}

\item{.plot_weights, ..., .one_page}{Plot arguments.
\itemize{
\item \code{.plot_weights} is a \code{logical} variable that defines whether or not to plot the frequency distribution of weights for each \code{.split} value (i.e., the frequency distribution of the number of detections at each receiver in each time interval, excluding time intervals without detections).
\item \code{...} is a placeholder for arguments passed to \code{\link[graphics:hist]{graphics::hist()}}.
\item \code{.one_page} A \code{logical} variable that defines whether or not to plot all histograms on one page.
}}
}
\value{
The function returns a \code{\link{data.table}} with the following columns:
\itemize{
\item \code{.split}---a \code{character} vector that distinguishes groups, if applicable;
\item \code{timestamp}---a \code{POSIXt} vector of time stamps;
\item \code{coa_x},\code{coa_y}---the coordinates of the COAs;
}

Data are arranged by \code{.split} and \code{timestamp}.
}
\description{
This function calculates centres of activity (COAs) from detections at acoustic receivers.
}
\details{
COAs are calculated as a weighted mean of the locations of receivers at which individuals are detected over consecutive time intervals, weighted by the frequency of detections at each of those receivers. COAs are calculated via \code{\link[stats:weighted.mean]{stats::weighted.mean()}} (for planar coordinates) or \code{\link[geosphere:geomean]{geosphere::geomean()}} (for longitude/latitude coordinates).
}
\examples{
require(data.table)

#### Example (1): Calculate COAs for an example individual
acc   <- dat_acoustics[individual_id == dat_acoustics$individual_id[1], ]
gebco <- dat_gebco()
dlist  <- pat_setup_data(.acoustics = acc,
                        .moorings = dat_moorings,
                        .bathy = gebco,
                        .lonlat = FALSE)
coa(dlist, .delta_t = "2 hours")
coa(dlist, .delta_t = "4 hours")

#### Example (2): Calculate COAs for multiple individuals via .split
dlist <- pat_setup_data(.acoustics = dat_acoustics,
                       .moorings = dat_moorings,
                       .bathy = gebco,
                       .lonlat = FALSE)
coa(dlist, .delta_t = "6 hours", .split = "individual_id")

#### Example (3): Use lon/lat coordinates
gebco_ll <- terra::project(dat_gebco(), "EPSG:4326")
dlist <- pat_setup_data(.acoustics = dat_acoustics,
                       .moorings = dat_moorings,
                       .bathy = gebco_ll,
                       .lonlat = TRUE)
coa(dlist, .delta_t = "6 hours", .split = "individual_id")

}
\seealso{
\itemize{
\item To derive location samples from the forward-filter backward-sampler, see \code{\link[=pf_forward]{pf_forward()}}, \code{\link[=pf_backward_sampler]{pf_backward_sampler()}} and \code{\link[=pf_coord]{pf_coord()}};
\item For mapping utilisation distributions from coordinates, see \verb{map_*()} functions (i.e., \code{\link[=map_pou]{map_pou()}} and \code{\link[=map_dens]{map_dens()}};
}
}
\author{
Edward Lavender
}
