% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coa.R
\name{coa_check}
\alias{coa_check}
\alias{.coa_check_acoustics}
\alias{.coa_check_lonlat}
\title{COA internals}
\usage{
.coa_check_acoustics(.o, .split)

.coa_check_lonlat(.acoustics)
}
\description{
COA internals
}
\author{
Edward Lavender
}
\keyword{internal}
