% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{create_log}
\alias{create_log}
\title{Utilities: create a log file}
\usage{
create_log(.file, .verbose)
}
\arguments{
\item{.file}{A character path to a file. \code{NULL} and \code{""} are permitted.}

\item{.verbose}{A logical variable that defines whether or not to act on \code{.file}.}
}
\value{
The function returns \code{invisible(TRUE)}.
}
\description{
This function creates a .txt file (\code{.file}) via \code{\link{file.create}}.
}
\author{
Edward Lavender
}
\keyword{internal}
