% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\name{dat_gebco}
\alias{dat_gebco}
\title{Dataset: A bathymetry dataset}
\source{
Raw data were sourced from GEBCO Compilation Group (2019) GEBCO 2019 Grid (doi:10.5285/836f016a-33be-6ddc-e053-6c86abc0788e)
}
\usage{
dat_gebco()
}
\value{
The function returns a \code{\link{SpatRaster}} with 264, rows, 190 columns and 1 layer, with the following properties:
\itemize{
\item \code{dimensions}---264, 190, 1 (nrow, ncol, nlyr)
\item \code{resolution}---100, 100  (x, y)
\item \code{extent}---695492.1, 714492.1, 6246657, 6273057  (xmin, xmax, ymin, ymax)
\item \verb{coord. ref.}---WGS 84 / UTM zone 29N (EPSG:32629)
}
}
\description{
This function loads a processed bathymetry (m) dataset for the west coast of Scotland.
}
