% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats-patter.R
\name{datasets-algorithms}
\alias{datasets-algorithms}
\alias{dat_path}
\alias{dat_coa}
\alias{dat_pff}
\title{Datasets: example algorithm outputs}
\source{
For full details on the algorithm parameters used to generate these datasets, see \url{https://github.com/edwardlavender/patter/blob/main/data-raw/003-add-data-algorithms.R}.
}
\usage{
dat_path()

dat_coa()

dat_pff()
}
\value{
The functions return a dataset or a \code{character} string that defines the path to a dataset. See the corresponding function documentation for full details.
}
\description{
These functions load example outputs from key \code{\link{patter}} functions. They are included in the package to streamline function examples and tests.
}
\details{
\itemize{
\item \code{\link[=dat_path]{dat_path()}} reads an example output from \code{\link[=sim_path_walk]{sim_path_walk()}};
\item \code{\link[=dat_coa]{dat_coa()}} reads an example output from \code{\link[=coa]{coa()}};
\item \code{\link[=dat_pff]{dat_pff()}} reads an example output from \code{\link[=pf_filter]{pf_filter()}};
}
}
\examples{
# Load inbuilt datasets
dat_path() |> summary()
dat_coa() |> summary()
dat_pff() |> summary()

}
\author{
Edward Lavender
}
