% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats-patter.R
\name{datasets-algorithms}
\alias{datasets-algorithms}
\alias{dat_obs}
\alias{dat_pff}
\alias{dat_pfbk}
\alias{dat_pfp}
\alias{dat_coa}
\title{Datasets: example algorithm outputs}
\source{
For full details on the algorithm parameters used to generate these datasets, see \url{https://github.com/edwardlavender/patter/blob/main/data-raw/003-add-data-algorithms.R}.
}
\usage{
dat_obs()

dat_pff()

dat_pfbk()

dat_pfp()

dat_coa()
}
\description{
These functions load example outputs from key \code{\link{patter}} functions. They are included in the package to streamline function examples and tests.
}
\details{
\itemize{
\item \code{\link[=dat_pff]{dat_pff()}} reads an example output from \code{\link[=pf_setup_obs]{pf_setup_obs()}};
\item \code{\link[=dat_pff]{dat_pff()}} reads an example output from \code{\link[=pf_forward]{pf_forward()}};
\item \code{\link[=dat_pfbk]{dat_pfbk()}} reads an example output from \code{\link[=pf_backward_killer]{pf_backward_killer()}};
\item \code{\link[=dat_pfp]{dat_pfp()}} reads an example output from \code{\link[=pf_path]{pf_path()}};
\item \code{\link[=dat_coa]{dat_coa()}} reads an example output from \code{\link[=coa]{coa()}};
}
}
\examples{
dat_obs() |> dplyr::glimpse()
dat_pff() |> summary()
dat_pfbk() |> summary()
dat_pfp() |> summary()
dat_coa() |> summary()

}
\author{
Edward Lavender
}
