% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats-patter.R
\name{datasets-algorithms}
\alias{datasets-algorithms}
\alias{dat_obs}
\alias{dat_overlaps}
\alias{dat_kernels}
\alias{dat_pff}
\alias{dat_pfb}
\alias{dat_pfp}
\title{Datasets: example algorithm outputs}
\source{
For full details on the algorithm parameters used to generate these datasets, see \url{https://github.com/edwardlavender/patter/blob/main/data-raw/002-add-data-algorithms.R}.
}
\usage{
dat_obs()

dat_overlaps()

dat_kernels()

dat_pff()

dat_pfb()

dat_pfp()
}
\description{
These functions load example outputs from key \code{\link{patter}} functions. They are included in the package to streamline function examples and tests.
}
\details{
\itemize{
\item \code{\link[=dat_obs]{dat_obs()}} reads an example input dataset from \code{\link[=acs_setup_obs]{acs_setup_obs()}};
\item \code{\link[=dat_overlaps]{dat_overlaps()}} reads an example output from \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}};
\item \code{\link[=dat_kernels]{dat_kernels()}} reads an example output from \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}};
\item \code{\link[=dat_pff]{dat_pff()}} reads an example output from \code{\link[=pf_forward]{pf_forward()}};
\item \code{\link[=dat_pfb]{dat_pfb()}} reads an example output from \code{\link[=pf_backward_killer]{pf_backward_killer()}};
\item \code{\link[=dat_pfp]{dat_pfp()}} reads an example output from \code{\link[=pf_path]{pf_path()}};
}
}
\examples{
dat_obs() |> summary()
dat_overlaps() |> summary()
dat_kernels() |> summary()
dat_pff() |> summary()
dat_pfb() |> summary()
dat_pfp() |> summary()

}
\author{
Edward Lavender
}
