% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc_setup_model}
\alias{dc_setup_model}
\title{DC: a template DC model}
\usage{
dc_setup_model(
  .obs,
  .t,
  .bathy,
  .unwrap = ifelse(inherits(.bathy, "SpatRaster"), FALSE, TRUE),
  ...
)
}
\arguments{
\item{.obs}{The \code{.obs} \code{\link{data.table}} from \code{\link[=dc]{dc()}}.}

\item{.t}{An \code{integer} that defines the time step (used to index \code{.obs}).}

\item{.bathy}{The \code{.bathy} \code{\link{SpatRaster}}, posibly as a \code{PackedSpatRaster}, from \code{\link[=dc]{dc()}}.}

\item{.unwrap}{A logical variable that defines whether or not to unwrap \code{.bathy} (via \code{\link[terra:wrap]{terra::unwrap()}}). This is required if \code{.bathy} is passed to this function as a \code{PackedSpatRaster}, which is the case for implementations of \code{\link[=dc]{dc()}} using socket clusters.}

\item{...}{Additional arguments passed from \code{\link[=dc]{dc()}} (unused here).}
}
\value{
The function returns a normalised \code{\link{SpatRaster}}.
}
\description{
This function is an example DC model, of the kind required by \code{.model} in \code{\link[=dc]{dc()}}.
}
\details{
This function assigns equal probability to all locations between two depth limits (as defined by \code{depth_shallow} and \code{depth_deep} columns in \code{.obs}).
}
\section{Warning}{
\itemize{
\item This function is used to streamline examples and does not represent a generically suitable model.
\item The function does not check user inputs.
}
}

\examples{
# For examples, see `?dc()`

}
\author{
Edward lavender
}
