% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dist_along_path}
\alias{dist_along_path}
\title{Distances: along a path}
\usage{
dist_along_path(.xy, .lonlat = FALSE)
}
\arguments{
\item{.xy}{A two-column matrix of coordinates.}

\item{.lonlat}{A \code{logical} variable that defines whether or not the coordinates are in longitude/latitude format.}
}
\description{
This function calculates distances between sequential coordinates in a matrix.
}
\details{
This function is a simple wrapper for \code{\link[terra:distance]{terra::distance()}}.
}
