% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-internal.R
\name{.acs_absences}
\alias{.acs_absences}
\title{AC* helper: define active, overlapping receivers with absences}
\usage{
.acs_absences(.date, .detections, .overlaps)
}
\arguments{
\item{.date}{A \code{character} that defines the date.}

\item{.detections}{An \code{integer} vector that defines the receiver(s) that recorded detection(s) at the current time step.}

\item{.overlaps}{A named \code{list} from \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}} that defines receiver overlaps. \code{NULL} is permitted.}
}
\value{
The function returns an \code{integer} vector that defines the set of receivers that overlap with \code{.detections} but did not record detections. \code{NULL} indicates no overlapping receivers.
}
\description{
Given a detection at one or more receivers on a given date, this function defines the set of remaining, active, overlapping receivers that did not record detections.
}
\details{
In the AC* algorithms, at the moment of detection, likelihood of the acoustic data given a particle sample depends on depends on both the receivers that record detections and those that did not (eqn S5 in Lavender et al., 2023). This function is used to restrict the set of receivers to which eqn S5 needs to be applied.
}
\section{Warning}{
For speed, this function performs no internal checks.
}

\seealso{
This function defines the \code{absences} argument for \code{\link[=.acs_given_detection_particles]{.acs_given_detection_particles()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
