% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-generic-internals.R
\name{.pf_history_dt}
\alias{.pf_history_dt}
\title{PF: Compile particle histories into a \code{\link{data.table}}}
\usage{
.pf_history_dt(.history, .collect = TRUE)
}
\arguments{
\item{.history}{The particle samples, supplied as:
\itemize{
\item A \code{\linkS4class{pf}} object;
\item The \code{history} element of a \code{\linkS4class{pf}} object;
\item A string that defines the directory containing parquet files;
\item A list of strings that defines parquet files;
}}
}
\description{
This function compiles particle histories into a \code{\link{data.table}}.
}
\author{
Edward Lavender
}
\keyword{internal}
