% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers-internal.R
\name{geomean}
\alias{geomean}
\title{Spatial helper: calculate the centre of mass of weighted coordinates}
\usage{
geomean(xy, w = NULL)
}
\arguments{
\item{xy, w}{Arguments passed to \code{geosphere::geomean()}.}
}
\description{
This is a wrapper for \code{geosphere::geomean()} that handles one-row matrices.
}
\details{
This function uses the internal code of \code{geosphere::geomean()} without the checks of the \code{.pointsToMatrix()} function.
}
\author{
Edward Lavender
}
\keyword{internal}
