% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{glossary}
\alias{glossary}
\title{Glossary}
\description{
This is a glossary of key arguments in \code{\link{patter}}.
}
\section{\code{.map}}{
\code{.map} is \code{\link{SpatRaster}} that defines the area of interest. \code{NAs} define inhospitable regions (such as land).
}

\section{\code{.model_obs}}{
\code{.model_obs} is a \code{character} that defines a \code{\link{ModelObs}} sub-type.
}

\section{\code{.model_move}}{
\code{.model_move} is a \code{character} string that defines a \code{\link{ModelMove}} instance.
}

\section{\code{.state}}{
\code{.state} is a \code{character} that defines a \code{\link{State}} sub-type.
}

\section{\code{.yobs}}{
\code{.yobs} is a \code{list} of observational datasets, one for each data type. Each element must be a \code{\link{data.table}} with the following columns:
\itemize{
\item \code{timestamp}---a \code{POSIXct} vector of time stamps;
\item \code{sensor_id}---an \code{integer} vector of sensor IDs, such as receivers;
\item \code{obs}---a vector of observations;
\item Additional columns that define the parameters of the observation model, as defined by a \code{\link{ModelObs}} structure;
}

No other columns should be included.

For real-world analyses, see \code{\link{assemble}}\verb{_*()} functions to format datasets as required. Other types of datasets require manual preparation.
}

\seealso{
\code{\link{State}}, \code{\link{ModelMove}}, \code{\link{ModelObs}}, \code{\link[=sim_path_walk]{sim_path_walk()}}, \code{\link[=sim_observations]{sim_observations()}}, \code{\link[=pf_filter]{pf_filter()}}
}
\author{
Edward Lavender
}
