% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-helpers.R
\name{julia_helper}
\alias{julia_helper}
\alias{julia_session}
\alias{julia_works}
\alias{julia_option}
\alias{julia_proj_path}
\alias{julia_proj_generate}
\alias{julia_proj_activate}
\alias{julia_proj_temp}
\alias{julia_pkg_list_full}
\alias{julia_pkg_list_req}
\alias{julia_pkg_list_installed}
\alias{julia_pkg_list_update}
\alias{julia_pkg_list_load}
\alias{julia_pkg_patter_source}
\alias{julia_pkg_install_Patter}
\alias{julia_pkg_install_deps}
\alias{julia_pkg_library}
\alias{julia_pkg_setup}
\alias{julia_threads}
\alias{julia_glimpse}
\alias{julia_print}
\alias{julia_summary}
\alias{julia_save}
\alias{julia_load}
\alias{julia_timeline}
\alias{julia_check_exists}
\alias{julia_code}
\title{Julia: helpers}
\usage{
julia_session()

julia_works(.action = abort)

julia_option(VALUE)

julia_proj_path(JULIA_PROJ)

julia_proj_generate(JULIA_PROJ)

julia_proj_activate(JULIA_PROJ)

julia_proj_temp()

julia_pkg_list_full(.pkg_install)

julia_pkg_list_req()

julia_pkg_list_installed()

julia_pkg_list_update(.pkg_update)

julia_pkg_list_load(.pkg_load)

julia_pkg_patter_source(JULIA_PATTER_SOURCE)

julia_pkg_install_Patter(JULIA_PATTER_SOURCE, .pkg_update)

julia_pkg_install_deps(.pkg_install, .pkg_update)

julia_pkg_library(.pkg_load)

julia_pkg_setup(JULIA_PATTER_SOURCE, .pkg_install, .pkg_update, .pkg_load)

julia_threads(JULIA_NUM_THREADS)

julia_glimpse(.x)

julia_print(.x)

julia_summary(.x)

julia_save(.x, .file = .x)

julia_load(.file, .x = basename(tools::file_path_sans_ext(.file)))

julia_timeline(.x)

julia_check_exists(...)

julia_code(.x)
}
\description{
A set of \code{Julia} helper functions.
}
\author{
Edward Lavender
}
\keyword{internal}
