% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-helpers.R
\name{julia_helper}
\alias{julia_helper}
\alias{julia_works}
\alias{julia_proj_generate}
\alias{julia_proj_activate}
\alias{julia_packages_install}
\alias{julia_packages_library}
\alias{julia_packages}
\alias{julia_threads}
\alias{julia_glimpse}
\alias{julia_print}
\alias{julia_summary}
\alias{julia_save}
\alias{julia_timeline}
\alias{julia_check_exists}
\alias{julia_code}
\title{Julia: helpers}
\usage{
julia_works(.action = abort)

julia_proj_generate(JULIA_PROJ)

julia_proj_activate(JULIA_PROJ)

julia_packages_install(.packages, .update)

julia_packages_library(.packages)

julia_packages(.packages, .update)

julia_threads(.threads)

julia_glimpse(x)

julia_print(x)

julia_summary(x)

julia_save(x, file = x)

julia_timeline(x)

julia_check_exists(...)

julia_code(x)
}
\description{
A set of \code{Julia} helper functions.
}
\author{
Edward Lavender
}
\keyword{internal}
