% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-set.R
\name{julia_set}
\alias{julia_set}
\alias{set_JULIA_NUM_THREADS}
\alias{set_timeline}
\alias{set_state_type}
\alias{set_xinit}
\alias{set_model_move}
\alias{set_path}
\alias{set_model_obs_types}
\alias{set_model_obs}
\alias{set_yobs_vect}
\alias{set_yobs_dict}
\alias{set_yobs_dict_via_sim}
\alias{set_n_particle}
\alias{set_direction}
\alias{set_states_init}
\alias{set_t_resample}
\alias{set_batch}
\alias{set_progress}
\alias{set_pf_filter}
\alias{set_cache}
\alias{set_smoother_two_filter}
\title{\code{Julia}: set objects in \code{Julia}}
\usage{
set_JULIA_NUM_THREADS(JULIA_NUM_THREADS)

set_timeline(.timeline)

set_state_type(.state)

set_xinit(.xinit)

set_model_move(.model_move)

set_path()

set_model_obs_types(.datasets)

set_model_obs(.model_obs)

set_yobs_vect(.timeline, .yobs)

set_yobs_dict(.yobs)

set_yobs_dict_via_sim()

set_n_particle(.n_particle)

set_direction(.direction = c("forward", "backward"))

set_states_init(
  .timeline,
  .state,
  .xinit,
  .model_move,
  .yobs,
  .n_particle,
  .direction
)

set_t_resample(.t_resample)

set_batch(.batch, .type = c("fwd", "bwd", "smo"))

set_progress(.progress)

set_pf_filter(
  .n_move,
  .n_resample,
  .t_resample,
  .n_record,
  .n_iter,
  .direction,
  .batch,
  .progress
)

set_cache(.cache)

set_smoother_two_filter(.n_particle, .n_sim, .cache, .batch, .progress)
}
\description{
These are internal functions that assign objects to variable names in \code{Julia}.
}
\author{
Edward Lavender
}
\keyword{internal}
