% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-set.R
\name{julia_set}
\alias{julia_set}
\alias{set_env}
\alias{set_initial_states}
\alias{set_move}
\alias{set_path}
\title{Julia: set objects in Julia}
\usage{
set_env(x)

set_initial_states(.state = "StateXY", .xinit = NULL, .bathy = NULL, .n = 100L)

set_move(cmd)

set_path(.n_step)
}
\description{
These are internal functions that assign objects to variable names in Julia.
}
\author{
Edward Lavender
}
\keyword{internal}
