% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-set.R
\name{julia_set}
\alias{julia_set}
\alias{set_map}
\alias{set_states_init}
\alias{set_move}
\alias{set_timeline}
\alias{set_path}
\title{Julia: set objects in Julia}
\usage{
set_map(x)

set_states_init(.state = "StateXY", .map = NULL, .xinit = NULL, .n = 100L)

set_move(cmd)

set_timeline(.timeline)

set_path()
}
\description{
These are internal functions that assign objects to variable names in Julia.
}
\author{
Edward Lavender
}
\keyword{internal}
