% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-set.R
\name{julia_set}
\alias{julia_set}
\alias{set_map}
\alias{set_states_init}
\alias{set_move}
\alias{set_timeline}
\alias{set_path}
\alias{set_model_types}
\alias{set_models}
\alias{set_parameters}
\alias{set_yobs_via_sim}
\alias{set_datasets}
\title{Julia: set objects in Julia}
\usage{
set_map(x)

set_states_init(.map, .timeline, .datasets, .models, .pars, .state, .xinit, .n)

set_move(cmd)

set_timeline(.timeline)

set_path()

set_model_types(.models)

set_models(.models)

set_parameters(.parameters)

set_yobs_via_sim()

set_datasets(.datasets)
}
\description{
These are internal functions that assign objects to variable names in Julia.
}
\author{
Edward Lavender
}
\keyword{internal}
