% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-metrics.R
\name{make_matrix_receivers}
\alias{make_matrix_receivers}
\title{Study metrics: receiver deployments}
\source{
This function is based on the \href{https://edwardlavender.github.io/flapper/reference/make_matrix_receivers.html}{\code{make_matrix_receivers}} function in the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package. Other \verb{flapper::make_matrix_*()} routines are not currently available in \code{\link{patter}}.
}
\usage{
make_matrix_receivers(
  .dlist,
  .start = NULL,
  .end = NULL,
  .delta_t = "days",
  .as_POSIXct = NULL,
  .set_names = TRUE
)
}
\arguments{
\item{.dlist}{A named \code{list} of data and parameters from \code{\link[=pat_setup_data]{pat_setup_data()}}. This function requires:
\itemize{
\item \code{.dlist$data$moorings}, with the following columns: \code{receiver_id}, \code{receiver_start}, and \code{receiver_end}.
\item (optional) \code{.dlist$data$services}, with the following columns: \code{receiver_id}, \code{service_start} and \code{service_end}.
}}

\item{.start, .end}{\code{Date} or \code{POSIXct} objects that define the start and end time. If unspecified, these are taken from the range of deployment times in \code{.moorings}.}

\item{.delta_t}{A \code{numeric} or \code{character} input that defines the time interval between successive time steps. This is passed to the \code{by} argument of \code{\link[base:seq.POSIXt]{base::seq.POSIXt()}} or \code{\link[base:seq.Date]{base::seq.Date()}} (depending on \code{.as_POSIXct}, below).}

\item{.as_POSIXct}{(optional) A \code{function} that coerces supplied any supplied times (\code{.dlist$data$moorings$receiver_start}, \verb{.dlist$data$$moorings$receiver_end}, \code{.dlist$data$services$service_start}, \code{.dlist$data$services$service_end}, \code{start} and \code{end}) that are not \code{POSIXct} objects to \code{POSIXct} objects. This can be suppressed via \code{.as_POSIXct = NULL} if supplied times are \code{Date} objects and \code{.delta_t} >= one day.}

\item{.set_names}{A \code{logical} variable that defines whether or not to set the row and column names of the matrix to the time steps and the receiver IDs respectively.}
}
\value{
The function returns a \code{\link{matrix}} with one row for each time step and one column for each receiver. Each cell defines whether (1) or not (0) each receiver was at active during that time step. A \code{bins} attribute is included, which defines the time steps as a \code{Date} or \code{POSIXct} vector.
}
\description{
This function creates a matrix that, for each time step (matrix row) in a sequence of user-defined times, defines whether or not each receiver (matrix column) was active, optionally accounting for servicing dates. The timeline is defined from a user-provided \code{start} and \code{end} date (or the range of deployment times in the input data) and an interval (\code{.delta_t}) parameter that defines the duration between successive time steps.
}
\seealso{
This function is used internally in \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
