% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_hr}
\alias{map_hr}
\alias{map_hr_prop}
\alias{map_hr_core}
\alias{map_hr_home}
\alias{map_hr_full}
\title{Map: animal home ranges}
\usage{
map_hr_prop(.map, .prop = 0.5, .add = FALSE, ...)

map_hr_core(.map, .add = FALSE, ...)

map_hr_home(.map, .add = FALSE, ...)

map_hr_full(.map, .add = FALSE, ...)
}
\arguments{
\item{.map}{A \code{\link{SpatRaster}} (utilisation distribution).}

\item{.prop}{For \code{\link[=map_hr_prop]{map_hr_prop()}}, \code{.prop} is a number that defines the range proportion.}

\item{.add}{A logical variable that defines whether or not to add a polygon of the range to an existing map.}

\item{...}{If \code{.add = TRUE}, \code{...} is a place holder for additional arguments passed to \code{\link[terra:plot]{terra::plot()}}.}
}
\value{
The functions return a \code{\link{SpatRaster}}. Cells with a value of one are inside the specified range boundaries; cells with a value of zero are beyond range boundaries. If \code{.add} is \code{TRUE}, the boundaries are added to an existing plot.
}
\description{
These functions extract 'home range' estimates from a \code{\link{SpatRaster}} that describes the intensity of movements within an area.
}
\details{
These functions are modelled on \href{https://edwardlavender.github.io/flapper/reference/get_hr.html}{\verb{flapper::map_hr_*()}} functions, where full details are provided.
}
\examples{
library(terra)

#### Set up example
# Define hypothetical input SpatRaster
r <- rast()
n <- ncell(r)
i <- 2e4
r[i] <- 1
r <- distance(r)
r <- r / global(r, "sum")[1, 1]
plot(r)

#### Examples
map <- map_hr_full(r, .add = TRUE, lwd = 5)
map <- map_hr_home(r, .add = TRUE, border = "blue")
map <- map_hr_core(r, .add = TRUE, border = "orange")
map <- map_hr_prop(r, .prop = 0.2, .add = TRUE, border = "red")
}
\seealso{
\verb{map_*()} functions build maps of space use:
\itemize{
\item \code{\link[=map_pou]{map_pou()}} maps probability-of-use;
\item \code{\link[=map_dens]{map_dens()}} maps point density;
\item \code{\link{map_hr}}\verb{_*()} functions map home ranges;
}

All maps are represented as \code{\link{SpatRaster}}s.

To derive coordinates for mapping patterns of space use for tagged animals, see:
\itemize{
\item \code{\link[=coa]{coa()}} to calculate centre-of-activity;
\item \code{\link[=pf_filter]{pf_filter()}} and associates to sample locations using particle filtering;
}
}
\author{
Edward Lavender
}
