% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_hr}
\alias{map_hr}
\alias{map_hr_prop}
\alias{map_hr_core}
\alias{map_hr_home}
\alias{map_hr_full}
\title{Get animal 'home ranges'}
\usage{
map_hr_prop(.map, .prop = 0.5, .add = FALSE, ...)

map_hr_core(.map, .add = FALSE, ...)

map_hr_home(.map, .add = FALSE, ...)

map_hr_full(.map, .add = FALSE, ...)
}
\arguments{
\item{.map}{A \code{\link{SpatRaster}} (utilisation distribution).}

\item{.prop}{For \code{\link[=map_hr_prop]{map_hr_prop()}}, \code{.prop} is a number that defines the range proportion.}

\item{.add}{A logical variable that defines whether or not to add a polygon of the range to an existing map.}

\item{...}{If \code{.add = TRUE}, \code{...} is a place holder for additional arguments passed to \code{\link[terra:plot]{terra::plot()}}.}
}
\value{
The functions return a \code{\link{SpatRaster}}. Cells with a value of one are inside the specified range boundaries; cells with a value of zero are beyond range boundaries. If \code{.add} is \code{TRUE}, the boundaries are added to an existing plot.
}
\description{
These functions extract 'home range' estimates from a \code{\link{SpatRaster}} that describes the intensity of movements within an area.
}
\details{
These functions are modelled on the \href{https://edwardlavender.github.io/flapper/reference/map_hr.html}{\verb{map_hr_*()}} functions in the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package, where full details are provided. The \code{spatialEco} package is required.
}
\examples{
#### Set up example
# Define hypothetical input SpatRaster
require(terra)
r <- rast()
n <- ncell(r)
i <- 2e4
r[i] <- 1
r <- distance(r)
r <- r / global(r, "sum")[1, 1]
plot(r)

#### Examples
map <- map_hr_full(r, .add = TRUE, lwd = 5)
map <- map_hr_home(r, .add = TRUE, border = "blue")
map <- map_hr_core(r, .add = TRUE, border = "orange")
map <- map_hr_prop(r, .prop = 0.2, .add = TRUE, border = "red")

}
\seealso{
\itemize{
\item For reconstructing movement paths and patterns of space use, see \code{\link[=pf_forward]{pf_forward()}};
\item For mapping utilisation distributions, see \code{\link[=pf_map_pou]{pf_map_pou()}} and \code{\link[=pf_map_dens]{pf_map_dens()}};
}
}
\author{
Edward Lavender
}
