% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_pou}
\alias{map_pou}
\title{Map: map probability-of-use}
\usage{
map_pou(.map, .coord, .plot = TRUE, ..., .verbose = TRUE)
}
\arguments{
\item{.map}{A \code{\link{SpatRaster}} that defines the grid for probability-of-use estimation. \code{NAs} on \code{.map} are used as a mask.}

\item{.coord}{Coordinates, provided in any format accepted by \code{\link[=.map_marks]{.map_marks()}}}

\item{.plot}{A logical input that defines whether or not to plot the \code{\link{SpatRaster}}.}

\item{...}{If \code{.plot = TRUE}, \code{...} is a place holder for additional arguments passed to \code{\link[terra:plot]{terra::plot()}}.}
}
\value{
The function returns a \code{\link{SpatRaster}} (utilisation distribution) in which cell values define probability-of-use.
}
\description{
This function builds a 'probability-of-use' utilisation distribution.
}
\details{
Probability-of-use is the proportion of samples of each unique cell (out of the total number of samples across all time steps).
}
\seealso{
\itemize{
\item The PF (forward simulation) is implemented by \code{\link[=pf_forward]{pf_forward()}};
\item PF is supported by:
\itemize{
\item Setup helpers, namely \code{\link[=pf_files]{pf_files()}};
}
\item The backward pass is implemented by \code{\link[=pf_backward_*]{pf_backward_*()}};
\item Movement paths are built from PF outputs via \code{pf_path()} functions:
\itemize{
\item \code{\link[=pf_path]{pf_path()}} reconstructs paths;
\item \code{\link[=pf_path_pivot]{pf_path_pivot()}} supports path reconstruction;
}
\item To reconstruct maps of space use, see:
\itemize{
\item \code{\link[=pf_coords]{pf_coords()}} to extract particle coordinates;
\item \code{\link[=map_pou]{map_pou()}} for probability-of-use maps;
\item \code{\link[=map_dens]{map_dens()}} for smooth utilisation distributions;
\item \code{\link[=get_hr]{get_hr()}} for home range estimates;
}
}
}
\author{
Edward Lavender
}
