% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat.R
\name{pat_setup_data}
\alias{pat_setup_data}
\title{Set up data for \code{\link{patter}}}
\usage{
pat_setup_data(
  .acoustics = NULL,
  .moorings = NULL,
  .services = NULL,
  .archival = NULL,
  .bathy = NULL,
  .lonlat = NULL
)
}
\arguments{
\item{.acoustics, .services, .archival, .moorings}{(optional) Movement datasets.
\itemize{
\item \code{.acoustics} contains the detection time series (see \code{\link{dat_acoustics}} for an example);
\item \code{.moorings} contains the receiver deployments (see \code{\link{dat_moorings}}) for an example);
\item \code{.services} contains receiver servicing information (times during the deployment period of a receiver when it was not active due to servicing);
\item \code{.archival} contains archival (depth) time series (see \code{\link{dat_archival}} for an example);
}}

\item{.bathy}{(optional) A \code{\link{SpatRaster}} that defines a (bathymetry) grid over the study area.}

\item{.lonlat}{(optional) A \code{logical} variable that defines whether spatial data is in longitude/latitude format.}
}
\value{
The function returns a named \code{list} with three elements:
\itemize{
\item \code{data}---a named list, with one element for each movement dataset;
\item \code{spatial}---a named list, with one element for each spatial dataset;
\item \code{pars}---a named list, with required parameters;
}
}
\description{
This function is used to set up movement datasets, spatial datasets and associated parameters for all \code{\link{patter}} functions. In general, this function should be called before other \code{\link{patter}} functions.
}
\details{
The preparation of datasets for \code{\link{patter}} is a one-off inconvenience. See the \code{\link{check_data}} function documentation for full details as to the required properties of each input dataset. All requirements are kept to a minimum and straightforward to address. To minimise inconvenience, all inputs are optional in \code{\link[=pat_setup_data]{pat_setup_data()}}. If you are only interested in  specific routine(s), check the associated documentation to see which inputs are essential for those routine(s). The output of \code{\link[=pat_setup_data]{pat_setup_data()}} is widely passed to downstream \code{\link{patter}} functions, which streamlines the API, documentation and internal code.
}
\examples{
data <- pat_setup_data(.acoustics = dat_acoustics,
                      .services = NULL,
                      .archival = dat_archival,
                      .moorings = dat_moorings,
                      .bathy = dat_gebco(),
                      .lonlat = FALSE)
summary(data)

}
\author{
Edward Lavender
}
