% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter-package.R
\docType{package}
\name{patter-package}
\alias{patter}
\alias{patter-package}
\title{\code{\link{patter}} for passive acoustic telemetry}
\description{
\code{\link{patter}} is a re-implementation of the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} family of algorithms for passive acoustic telemetry that is simpler, faster and better tested than its \href{https://github.com/edwardlavender/flapper}{predecessor}. The aim of the package is to facilitate the reconstruction of movement paths and patterns of space use in passive acoustic telemetry systems.
}
\section{Vignettes}{
For an introduction to \code{\link{patter}}, use:
\itemize{
\item \code{vignette("workflow", package = "patter")} for an overview of the analytical workflow;
\item \code{vignette("faqs", package = "patter")} for FAQs;
}

For a full list of all functions, see \code{help(package = "patter")}.
}

\section{Datasets}{
For example datasets from the Movement Ecology of Flapper Skate project (\code{\link{datasets-mefs}}), see:
\itemize{
\item \code{\link{dat_moorings}} for receiver locations and associated information;
\item \code{\link{dat_acoustics}} for acoustic time series;
\item \code{\link{dat_archival}} for archival (depth) time series;
\item \code{\link[=dat_gebco]{dat_gebco()}} for a bathymetry grid;
}

For example algorithm outputs (\code{\link{datasets-algorithms}}), see:
\itemize{
\item \code{\link[=dat_obs]{dat_obs()}} for an example output dataset from \code{\link[=acs_setup_obs]{acs_setup_obs()}};
\item \code{\link[=dat_overlaps]{dat_overlaps()}} for an example output from \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}};
\item \code{\link[=dat_kernels]{dat_kernels()}} for an example output from \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}};
\item \code{\link[=dat_pff]{dat_pff()}} for an example output from \code{\link[=pf_forward]{pf_forward()}};
\item \code{\link[=dat_pfb]{dat_pfb()}} for an example output from \code{\link[=pf_backward_killer]{pf_backward_killer()}};
\item \code{\link[=dat_pfp]{dat_pfp()}} for an example output from \code{\link[=pf_path]{pf_path()}};
}
}

\section{Simulation}{
To simulate passive acoustic telemetry data, see:
\itemize{
\item \code{\link[=sim_array]{sim_array()}} to simulate an acoustic array;
\item \code{\link[=sim_path_walk]{sim_path_walk()}} to simulate a movement path from a walk model, with the help of:
\itemize{
\item \code{\link[=rtruncgamma]{rtruncgamma()}} and \code{\link[=rlen]{rlen()}} to simulate step lengths;
\item \code{\link[=rwn]{rwn()}}, \code{\link[=rangrw]{rangrw()}} and \code{\link[=rangcrw]{rangcrw()}} to simulate turning angles;
}
\item \code{\link[=sim_detections]{sim_detections()}} to simulate detections at receivers, with the help of:
\itemize{
\item \code{\link[=calc_detection_pr_logistic]{calc_detection_pr_logistic()}} and \code{\link[=calc_detection_pr]{calc_detection_pr()}}, which represent example detection probability models;
}
}

To evaluate model skill in reconstructing simulated patterns, see \code{skill_()} functions, specifically:
\itemize{
\item \code{\link[=skill_mb]{skill_mb()}} to calculate mean bias;
\item \code{\link[=skill_me]{skill_me()}} to calculate mean error;
\item \code{\link[=skill_rmse]{skill_rmse()}} to calculate root mean squared error;
\item \code{\link[=skill_R]{skill_R()}} to calculate Spearman's rank correlation coefficient;
\item \code{\link[=skill_d]{skill_d()}} to calculate the index of agreement;
}
}

\section{Data preparation}{
For help with data acquisition, processing, checking and preliminary analyses, see the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package. This facilitates:
\itemize{
\item Simulation;
\item Data preparation;
\item Spatial operations;
\item Distance calculations;
\item Movement analyses;
}

Please submit a \href{https://github.com/edwardlavender/patter/issues}{feature request} if you would like functions from \href{https://github.com/edwardlavender/flapper}{\code{flapper}} in \href{https://github.com/edwardlavender/patter}{\code{patter}}.
}

\section{Modelling workflow}{
The main thrust of \href{https://github.com/edwardlavender/patter}{\code{patter}} is the provision of an integrated modelling workflow for reconstructing animal movement paths and emergent patterns of space use in passive acoustic telemetry systems.

To implement the centre-of-activity algorithm use:
\itemize{
\item \code{\link[=coa]{coa()}} to calculate centres of activity;
}

To implement the acoustic-centroid* (AC*) algorithm(s) (e.g., AC and ACDC), use:
\itemize{
\item \code{\link[=acs_setup_obs]{acs_setup_obs()}} to set up observations;
\item \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}} to set up detection overlaps;
\item \code{\link[=acs_setup_detection_pr]{acs_setup_detection_pr()}} to define a detection probability model;
\item \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}} to set up detection kernels;
}

To implement particle filtering (PF), use:
\itemize{
\item \code{\link[=pf_setup_files]{pf_setup_files()}} to set up files;
\item \code{\link[=pf_forward]{pf_forward()}} to implement the forward simulation;
\item \code{\link[=pf_backward_*]{pf_backward_*()}} functions implement the backward pass;
\item \code{\link[=pf_coords]{pf_coords()}} to collate particle coordinates;
\item \code{\link[=pf_path]{pf_path()}} (and \code{\link[=pf_path_pivot]{pf_path_pivot()}}) to reconstruct movement paths;
\item \code{\link[=pf_map_pou]{pf_map_pou()}} to map probability-of-use;
\item \code{\link[=pf_map_dens]{pf_map_dens()}} to create smooth maps;
}

For home ranges, use:
\itemize{
\item \code{\link[=get_hr_prop]{get_hr_prop()}} for a custom range;
\item \code{\link[=get_hr_core]{get_hr_core()}} for the 'core' range;
\item \code{\link[=get_hr_home]{get_hr_home()}} for the 'home' range;
\item \code{\link[=get_hr_full]{get_hr_full()}} for the full range;
}
}

\section{Miscellaneous helpers}{
The following convenience functions are also made available to users of \href{https://github.com/edwardlavender/patter}{\code{patter}}:
\itemize{
\item Operational statistics. See:
\itemize{
\item \code{\link[=make_matrix_receivers]{make_matrix_receivers()}} to matricise receiver deployment time series;
}
\item Distance calculations. See:
\itemize{
\item \code{\link[=dist_along_path]{dist_along_path()}} to calculate distances along a movement path;
\item \code{\link[=dist_btw_cells]{dist_btw_cells()}} to calculate distances between cells on a \code{\link{SpatRaster}};
\item \code{\link[=degrees]{degrees()}} to create circular angles;
}
\item \code{terra} helpers. See:
\itemize{
\item \code{\link[=spatTemplate]{spatTemplate()}} to create a template \code{\link{SpatRaster}};
\item \code{\link[=spatNormalise]{spatNormalise()}} to normalise a \code{\link{SpatRaster}};
\item \code{\link{as.im.SpatRaster}} and \code{\link{as.owin.SpatRaster}} for interfaces to \code{spatstat} function(s);
}
\item Parallelisation. See:
\itemize{
\item \code{\link[=cl_lapply]{cl_lapply()}} and associated \verb{cl_*() functions for parallelisation routines};
}
}
}

\references{
Lavender, E. et al. (2023). An integrative modelling framework for passive acoustic telemetry. Methods in Ecology and Evolution. \url{https://doi.org/10.1111/2041-210X.14193}.
}
\seealso{
\itemize{
\item For information on the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} algorithms, see Lavender et al. (\href{https://doi.org/10.1111/2041-210X.14193}{2023}).
\item For information on \code{\link{patter}}'s predecessor, see \url{https://github.com/edwardlavender/flapper}.
\item For further information on  \code{\link{patter}}, see \url{https://github.com/edwardlavender/patter}.
\item For feature requests and bug reports, see \url{https://github.com/edwardlavender/patter/issues}.
\item For support, contact \href{mailto:edward.lavender@eawag.ch}{edward.lavender@eawag.ch}.
}
}
\author{
Edward Lavender (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})
}
\keyword{internal}
