% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter-progress.R
\name{patter-progress}
\alias{patter-progress}
\title{\code{\link{patter}} options: progress}
\description{
\code{\link{patter}} functions enable progress monitoring via function arguments and global options. At the time of writing, there are three main tools to monitor and enhance function progress:
\itemize{
\item User output messages (via the \code{patter.verbose} option and the \code{.verbose} argument);
\item Progress bars (via \code{\link[pbapply:pboptions]{pbapply::pboptions()}});
\item Parallelisation (via \code{JULIA_NUM_THREADS} and \code{.cl_} arguments);
}

Only selected \code{\link{patter}} functions support these options but this may expand in the future, depending on user feedback.
}
\section{User outputs}{
User output messages are controlled via the \code{.verbose} argument. There is a global option \code{patter.verbose} that can be set to suppress user output messages. See the internal \code{\link{cat_}} documentation for permitted inputs to \code{.verbose}.
}

\section{Progress bars}{
For \code{R} functions with a \code{Julia} backend, progress bars are implemented via \code{Julia} and cannot be modified by the user.

For pure \code{R} functions, progress bars are implemented via the \code{pbapply} package and controlled globally via \code{\link[pbapply:pboptions]{pbapply::pboptions()}}. See the internal \code{\link{pb_}} function documentation for examples.
}

\section{Parallelisation}{
For \code{R} functions with a \code{Julia} backend, parallelisation is implemented via the \code{JULIA_NUM_THREADS} argument to \code{\link[=julia_connect]{julia_connect()}}. This can only be set once per \code{R} session.

For pure \code{R} functions, parallelisation is implemented via \code{.cl_} arguments passed to \code{\link[=cl_lapply]{cl_lapply()}}, which wraps \code{\link[pbapply:pbapply]{pbapply::pblapply()}}. See the \code{\link[=cl_lapply]{cl_lapply()}} function documentation for full details.
}

\author{
Edward Lavender
}
