% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter-package.R
\docType{package}
\name{patter}
\alias{patter-package}
\alias{patter}
\title{\code{\link{patter}} for passive acoustic telemetry}
\description{
\code{\link{patter}} is a \code{R} implementation of a forward filtering--backward sampling algorithm for passive acoustic telemetry. This methodology enables the reconstruction of movement paths and patterns of space use in passive acoustic telemetry systems. \code{\link{patter}} unifies a suite of methods formerly known as the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} algorithms (Lavender et al., 2023) and supersedes the experimental \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package.
}
\section{Vignettes}{
For an introduction to \code{\link{patter}}, use:
\itemize{
\item \code{vignette("a-methodology", package = "patter")} for a conceptual introduction to the methodology;
\item \code{vignette("b-workflow-outline", package = "patter")} for an overview of the workflow;
\item \code{vignette("c-workflow-example", package = "patter")} for an example workflow;
\item \code{vignette("d-demos", package = "patter")} for more involved demonstrations;
\item \code{vignette("e-faqs", package = "patter")} for FAQs;
}

For a full list of all functions, see \code{help(package = 'patter')}.
}

\section{Datasets}{
For example datasets from the Movement Ecology of Flapper Skate project (\code{\link{datasets-mefs}}), which inspired \code{\link{patter}}, see:
\itemize{
\item \code{\link{dat_moorings}} for receiver locations and associated information;
\item \code{\link{dat_acoustics}} for acoustic time series;
\item \code{\link{dat_archival}} for archival (depth) time series;
\item \code{\link[=dat_gebco]{dat_gebco()}} for a bathymetry grid;
}

For example algorithm outputs (\code{\link{datasets-algorithms}}), see:
\itemize{
\item \code{\link[=dat_obs]{dat_obs()}} for an example output dataset from \code{\link[=pf_setup_obs]{pf_setup_obs()}};
\item \code{\link[=dat_pff]{dat_pff()}} and \code{\link[=dat_pff_src]{dat_pff_src()}} for example outputs from \code{\link[=pf_forward]{pf_forward()}};
\item \code{\link[=dat_pfbk]{dat_pfbk()}} and \code{\link[=dat_pfbk_src]{dat_pfbk_src()}} for example outputs from \code{\link[=pf_backward_killer]{pf_backward_killer()}};
\item \code{\link[=dat_pfp]{dat_pfp()}} for an example output from \code{\link[=pf_path]{pf_path()}};
\item \code{\link[=dat_coa]{dat_coa()}} for an example output from \code{\link[=coa]{coa()}};
}
}

\section{Simulation}{
To simulate passive acoustic telemetry data, see:
\itemize{
\item \code{\link[=sim_array]{sim_array()}} to simulate an acoustic array;
\item \code{\link[=sim_path_walk]{sim_path_walk()}} to simulate a movement path from a walk model;
\item \code{\link[=sim_detections]{sim_detections()}} to simulate detections at receivers;
}

These functions are supported by a set of simulation helpers, including:
\itemize{
\item \code{\link[=rbern]{rbern()}}, \code{\link[=rdet]{rdet()}}, \code{\link[=dbern]{dbern()}}, \code{\link[=ddetlogistic]{ddetlogistic()}} and \code{\link[=ddet]{ddet()}} for the simulation of detections;
\item \code{\link[=dtruncgamma]{dtruncgamma()}}, \code{\link[=rtruncgamma]{rtruncgamma()}}, \code{\link[=rlen]{rlen()}} and \code{\link[=clen]{clen()}} for the simulation of step lengths;
\item \code{\link[=rwn]{rwn()}}, \code{\link[=rangrw]{rangrw()}}, \code{\link[=rangcrw]{rangcrw()}} and \code{\link[=cang]{cang()}} for the simulation of turning angles;
\item \code{\link[=rstep]{rstep()}}, \code{\link[=dstep]{dstep()}} and \code{\link[=cstep]{cstep()}} for the simulation of steps into new locations;
\item \code{\link[=dkick]{dkick()}} and \code{\link[=rkick]{rkick()}} for \verb{*step()} wrappers;
}

To evaluate model skill in reconstructing simulated patterns, see \code{skill_()} functions, specifically:
\itemize{
\item \code{\link[=skill_mb]{skill_mb()}} to calculate mean bias;
\item \code{\link[=skill_me]{skill_me()}} to calculate mean error;
\item \code{\link[=skill_rmse]{skill_rmse()}} to calculate root mean squared error;
\item \code{\link[=skill_R]{skill_R()}} to calculate Spearman's rank correlation coefficient;
\item \code{\link[=skill_d]{skill_d()}} to calculate the index of agreement;
}
}

\section{Data exploration}{
For help with data acquisition, processing, checking and preliminary analyses, see the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package. This facilitates:
\itemize{
\item Data preparation;
\item Spatial operations;
\item Distance calculations;
\item Movement analyses;
}

Please submit a \href{https://github.com/edwardlavender/patter/issues}{feature request} if you would like functions from \href{https://github.com/edwardlavender/flapper}{\code{flapper}} in \code{\link{patter}}.
}

\section{Modelling workflow}{
The main thrust of \code{\link{patter}} is the provision of an integrated modelling workflow for reconstructing animal movement paths and emergent patterns of space use in passive acoustic telemetry systems.

\strong{To set up data for \code{\link{patter}}}, use
\itemize{
\item \code{\link[=pat_setup_data]{pat_setup_data()}} to set up data;
}

\strong{To implement the centre-of-activity algorithm}, use:
\itemize{
\item \code{\link[=coa]{coa()}} to calculate centres of activity;
}

\strong{To implement the particle filter (PF)}, use:
\itemize{
\item \code{\link[=pf_setup_obs]{pf_setup_obs()}} to set up a timeline of observations;
\item \code{\link[=pf_forward]{pf_forward()}} to implement the forward simulation;
}

\strong{PF is supported by}:
\itemize{
\item Proposal functions (see \code{\link{pf_propose}}) for the generation of new (candidate) locations, including:
\itemize{
\item \code{\link[=pf_rpropose_kick]{pf_rpropose_kick()}}, which uses stochastic kicks;
\item \code{\link[=pf_rpropose_reachable]{pf_rpropose_reachable()}}, which supports directed sampling;
}
\item Likelihood functions (see \code{\link{pf_lik}}) for evaluating the likelihood of the data, given proposal locations, including:
\itemize{
\item \code{\link{acs_filter_land}}, which filters proposals on land;
\item \code{\link{acs_filter_container}}, which filters proposals incompatible with acoustic container dynamics;
\item \code{\link{pf_lik_ac}}, which calculates the likelihood of acoustic data;
\item \code{\link{pf_lik_dc}}, which calculates the likelihood of depth observations;
}
\item Likelihood helpers, including:
\itemize{
\item \code{\link[=acs_setup_detection_overlaps]{acs_setup_detection_overlaps()}}, which pre-calculates detection overlaps;
\item \code{\link[=acs_setup_detection_kernel]{acs_setup_detection_kernel()}}, which prepares a detection kernel;
\item \code{\link[=acs_setup_detection_kernels]{acs_setup_detection_kernels()}}, which pre-calculates detection kernels;
}
\item (Re)sampling functions (see \code{\link{pf_sample}}) for the (re)sampling of valid proposal locations, including:
\itemize{
\item \code{\link[=pf_sample_multinomial]{pf_sample_multinomial()}}, which implements multinomial resampling;
\item \code{\link[=pf_sample_systematic]{pf_sample_systematic()}}, which implements systematic resampling;
}
\item Option functions (\code{\link{pf_opt}}) for tuning the forward simulation, including:
\itemize{
\item \code{\link[=pf_opt_trial]{pf_opt_trial()}}, which sets convergence parameters;
\item \code{\link[=pf_opt_rerun_from]{pf_opt_rerun_from()}}, which sets re-run parameters;
\item \code{\link[=pf_opt_control]{pf_opt_control()}}, which sets control parameters;
\item \code{\link[=pf_opt_record]{pf_opt_record()}}, which sets output properties;
}
}

\strong{To implement the backward pass (\code{\link[=pf_backward_*]{pf_backward_*()}})}, use:
\itemize{
\item \code{\link[=pf_backward_killer]{pf_backward_killer()}} to prune dead-ends;
\item \code{\link[=pf_backward_sampler]{pf_backward_sampler()}} to run the backward sampler;
}

\strong{For particle diagnostics}, see:
\itemize{
\item \code{\link[=pf_diag_convergence]{pf_diag_convergence()}} to collate convergence diagnostics;
\item \code{\link[=pf_diag_summary]{pf_diag_summary()}} to collate summary diagnostics;
}

\strong{For convenience plotting functions}, see:
\itemize{
\item \code{\link[=pf_plot_history]{pf_plot_history()}} to plot particle histories;
}

\strong{For common utility functions}, see:
\itemize{
\item \code{\link[=pf_files]{pf_files()}} to list particle-sample files;
\item \code{\link[=pf_files_size]{pf_files_size()}} to measure file size;
}

\strong{For movement-path reconstruction}, use:
\itemize{
\item \code{\link[=pf_path]{pf_path()}} (and \code{\link[=pf_path_pivot]{pf_path_pivot()}}) to reconstruct movement paths;
}

\strong{For mapping utilisation distributions}, use:
\itemize{
\item \code{\link[=pf_coord]{pf_coord()}} to collate particle coordinates for mapping;
\item \code{\link[=map_pou]{map_pou()}} to map probability-of-use;
\item \code{\link[=map_dens]{map_dens()}} to create smooth maps using \code{spatstat}, plus the supporting functions:
\itemize{
\item \code{\link[=as.im.SpatRaster]{as.im.SpatRaster()}}, to convert \code{\link{SpatRaster}}s to pixel images;
\item \code{\link[=as.owin.SpatRaster]{as.owin.SpatRaster()}}, to convert  \code{\link{SpatRaster}}s to observation windows;
\item \code{\link[=as.owin.sf]{as.owin.sf()}}, to convert \code{sf} objects to observation windows;
}
\item \verb{map_hr_*()} to map home ranges, specifically:
\itemize{
\item \code{\link[=map_hr_prop]{map_hr_prop()}} for a custom range;
\item \code{\link[=map_hr_core]{map_hr_core()}} for the 'core' range;
\item \code{\link[=map_hr_home]{map_hr_home()}} for the 'home' range;
\item \code{\link[=map_hr_full]{map_hr_full()}} for the full range;
}
}
}

\section{Miscellaneous helpers}{
The following convenience functions are also exported. Use:
\itemize{
\item \code{\link[=ssf]{ssf()}} and \code{\link[=ssv]{ssv()}} to \code{\link[=set.seed]{set.seed()}};
\item \code{\link[=dist_along_path]{dist_along_path()}} to calculate distances along a movement path;
\item \code{\link[=degrees]{degrees()}} to create circular angles;
\item \code{\link[=spatTemplate]{spatTemplate()}} to create a template \code{\link{SpatRaster}};
}
}

\section{Options}{
For additional options in \code{\link{patter}}, see:
\itemize{
\item \code{\link{patter-progress}} to monitor function progress;
}
}

\references{
Lavender, E. et al. (2023). An integrative modelling framework for passive acoustic telemetry. Methods in Ecology and Evolution. \url{https://doi.org/10.1111/2041-210X.14193}.
}
\seealso{
\itemize{
\item For information on the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} algorithms, see Lavender et al. (\href{https://doi.org/10.1111/2041-210X.14193}{2023}).
\item For information on \code{\link{patter}}'s predecessor, see \url{https://github.com/edwardlavender/flapper}.
\item For further information on  \code{\link{patter}}, including useful resources, see \url{https://github.com/edwardlavender/patter}.
\item For feature requests and bug reports, see \url{https://github.com/edwardlavender/patter/issues}.
\item For support, contact \href{mailto:edward.lavender@eawag.ch}{edward.lavender@eawag.ch}.
}
}
\author{
Edward Lavender (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})
}
