% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patter-package.R
\docType{package}
\name{patter}
\alias{patter-package}
\alias{patter}
\title{\code{\link{patter}}: particle algorithms for animal movement}
\description{
\code{\link{patter}} provides particle filtering, smoothing and sampling algorithms for animal movement modelling, with a focus on passive acoustic telemetry systems. This wraps and enhances a fast \href{https://julialang.org}{\code{Julia}} backend (\href{https://edwardlavender.github.io/Patter.jl}{\code{Patter.jl}}). The methodology enables the reconstruction of movement paths and patterns of space use. \code{\link{patter}} unifies a suite of methods formerly known as the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} algorithms  and supersedes the experimental \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package (Lavender et al., \href{https://doi.org/10.1111/2041-210X.14193}{2023}).
}
\section{Vignettes}{
For an introduction to \code{\link{patter}}, use:
\itemize{
\item \code{vignette("a-methodology", package = "patter")} for a conceptual introduction to the methodology;
\item \code{vignette("b-workflow-outline", package = "patter")} for an overview of the workflow;
}

For a full list of all functions, see \code{help(package = 'patter')}.

For a glossary of key arguments, see \code{\link{glossary}}.
}

\section{Datasets}{
For example datasets from the Movement Ecology of Flapper Skate project (\code{\link{datasets-mefs}}), which inspired \code{\link{patter}}, see:
\itemize{
\item \code{\link{dat_moorings}} for acoustic receiver deployments;
\item \code{\link{dat_detections}} for acoustic detection time series;
\item \code{\link{dat_archival}} for archival (depth) time series;
\item \code{\link[=dat_gebco]{dat_gebco()}} for a bathymetry grid;
}

To validate new datasets for use with \code{\link{patter}}, see \code{\link[=pat_setup_data]{pat_setup_data()}} and/or the \code{\link{assemble}}\verb{_*()} function documentation.

For example algorithm outputs (\code{\link{datasets-algorithms}}), see:
\itemize{
\item \code{\link[=dat_path]{dat_path()}} for an example output from \code{\link[=sim_path_walk]{sim_path_walk()}};
\item \code{\link[=dat_coa]{dat_coa()}} for an example output from \code{\link[=coa]{coa()}};
\item \code{\link[=dat_pff]{dat_pff()}} and \code{\link[=dat_pfb]{dat_pfb()}} for an example output from \code{\link[=pf_filter]{pf_filter()}};
\item \code{\link[=dat_tff]{dat_tff()}} for an example output from \code{\link[=pf_smoother_two_filter]{pf_smoother_two_filter()}};
}
}

\section{Set up \code{Julia}}{
To link \code{\link{patter}} and the \href{https://edwardlavender.github.io/Patter.jl}{\code{Patter.jl}} \code{Julia} backend, use:
\itemize{
\item \code{\link[=julia_connect]{julia_connect()}} to connect to \code{R} to \code{Julia};
\item \code{\link[=julia_validate]{julia_validate()}} to validate the \code{R}---\code{Julia} connection;
\item \code{\link[=set_seed]{set_seed()}} to set the seed in \code{R} and \code{Julia};
\item \code{\link[=set_map]{set_map()}} to make a \code{\link{SpatRaster}} of the study area available in \code{Julia};
}

These functions should be run at the start of every \code{R} session.
}

\section{Abstract Types}{
\code{\link{patter}} is based on three Abstract Types, defined in \code{Julia}:
\itemize{
\item \code{\link{State}} structures hold the state (location) of an animal at a given time step;
\item \code{\link{ModelMove}} structures hold movement model, used to simulate new states;
\item \code{\link{ModelObs}} structures hold observation model parameters, used to evaluate the correspondence between simulated states and observations;
}
}

\section{Simulation}{
To simulate animal movement time series, see:
\itemize{
\item \code{\link[=sim_path_walk]{sim_path_walk()}} to simulate a movement path from a walk model (via \code{\link{ModelMove}});
\item \code{\link[=sim_array]{sim_array()}} to simulate an acoustic array;
\item \code{\link[=sim_observations]{sim_observations()}} to simulate observational time series (via \code{\link{ModelObs}});
}

To evaluate model skill in reconstructing simulated patterns, see \verb{skill_*()} functions:
\itemize{
\item \code{\link[=skill_mb]{skill_mb()}} to calculate mean bias;
\item \code{\link[=skill_me]{skill_me()}} to calculate mean error;
\item \code{\link[=skill_rmse]{skill_rmse()}} to calculate root mean squared error;
\item \code{\link[=skill_R]{skill_R()}} to calculate Spearman's rank correlation coefficient;
\item \code{\link[=skill_d]{skill_d()}} to calculate the index of agreement;
}
}

\section{Data exploration}{
For help with data acquisition, processing, checking and preliminary analyses, see the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package. This facilitates:
\itemize{
\item Data preparation;
\item Spatial operations;
\item Distance calculations;
\item Movement analyses;
}

Please submit a \href{https://github.com/edwardlavender/patter/issues}{feature request} if you would like functions from \href{https://github.com/edwardlavender/flapper}{\code{flapper}} in \code{\link{patter}}.
}

\section{Algorithms}{
The main thrust of \code{\link{patter}} is the provision of fast, integrated modelling workflow based on particle filtering for reconstructing animal movement paths and emergent patterns of space use from observational time series (with a focus on passive acoustic telemetry systems).

\strong{To assemble datasets for particle filtering}, use \code{\link{assemble}}\verb{_*()} functions:
\itemize{
\item \code{\link[=assemble_timeline]{assemble_timeline()}} assembles a timeline;
\item \code{\link[=assemble_acoustics]{assemble_acoustics()}} assembles an acoustic time series;
\item \code{\link[=assemble_acoustics_containers]{assemble_acoustics_containers()}} assembles a corresponding time series of acoustic containers;
\item \code{\link[=assemble_archival]{assemble_archival()}} assembles an archival time series;
\item \code{\link[=assemble_custom]{assemble_custom()}} assembles custom time series;
}

Ancillary time series should be structured in the same way for inclusion in the particle filter.

\strong{To implement particle filtering (PF) routines}, use:
\itemize{
\item \code{\link[=pf_filter]{pf_filter()}} to implement the particle filter;
\item \code{\link[=pf_smoother_two_filter]{pf_smoother_two_filter()}} to implement the two-filter smoother;
}

These functions return \code{\linkS4class{pf_particles}} objects.

\strong{For convenience plotting functions}, see:
\itemize{
\item \code{\link[=pf_plot_xy]{pf_plot_xy()}} to plot particle locations;
}

\strong{For mapping utilisation distributions}, use:
\itemize{
\item \code{\link[=map_pou]{map_pou()}} to map probability-of-use;
\item \code{\link[=map_dens]{map_dens()}} to create smooth maps using \code{spatstat}, plus the supporting functions:
\itemize{
\item \code{\link[=as.im.SpatRaster]{as.im.SpatRaster()}}, to convert \code{SpatRaster}s to pixel images;
\item \code{\link[=as.owin.SpatRaster]{as.owin.SpatRaster()}}, to convert  \code{SpatRaster}s to observation windows;
\item \code{\link[=as.owin.sf]{as.owin.sf()}}, to convert \code{sf} objects to observation windows;
}
\item \verb{map_hr_*()} to map home ranges, specifically:
\itemize{
\item \code{\link[=map_hr_prop]{map_hr_prop()}} for a custom range;
\item \code{\link[=map_hr_core]{map_hr_core()}} for the 'core' range;
\item \code{\link[=map_hr_home]{map_hr_home()}} for the 'home' range;
\item \code{\link[=map_hr_full]{map_hr_full()}} for the full range;
}
}
}

\section{Options}{
For additional options in \code{\link{patter}}, see:
\itemize{
\item \code{\link{patter-progress}} to monitor function progress;
}
}

\section{Miscellaneous}{
\itemize{
\item See \code{\link[=cl_lapply]{cl_lapply()}} for \code{R} parallelisation;
\item See \code{\link[=example_setup]{example_setup()}} for functions used to streamline examples;
\item See \verb{file_*()} functions (e.g., \code{\link[=file_list]{file_list()}}) for system file helpers;
}
}

\references{
Lavender, E. et al. (2023). An integrative modelling framework for passive acoustic telemetry. Methods in Ecology and Evolution. \url{https://doi.org/10.1111/2041-210X.14193}.
}
\seealso{
\itemize{
\item For information on the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} algorithms, see Lavender et al. (\href{https://doi.org/10.1111/2041-210X.14193}{2023}).
\item For information on \code{\link{patter}}'s predecessor, see \url{https://github.com/edwardlavender/flapper}.
\item For further information on  \code{\link{patter}}, including useful resources, see \url{https://github.com/edwardlavender/patter}.
\item For feature requests and bug reports, see \url{https://github.com/edwardlavender/patter/issues}.
\item For the \code{Julia} backend, see \url{https://edwardlavender.github.io/Patter.jl}.
\item For support, contact \href{mailto:edward.lavender@eawag.ch}{edward.lavender@eawag.ch}.
}
}
\author{
Edward Lavender (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})
}
