% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_setup.R
\name{patter_run}
\alias{patter_run}
\alias{patter_run_expensive}
\title{Examples: run code}
\usage{
patter_run(.julia = TRUE, .geospatial = TRUE)

patter_run_expensive()
}
\arguments{
\item{.julia}{A \code{logical} variable that defines whether or not the code requires connection to a \code{Julia} session.}

\item{.geospatial}{A \code{logical} variable that defines whether or not \code{R}'s geospatial packages (e.g., \code{terra}) are required.}
}
\value{
The functions return a \code{logical} value.
}
\description{
These functions switch code blocks on/off.
}
\details{
\code{\link[=patter_run]{patter_run()}} is the main switch:
\itemize{
\item If \code{.julia = TRUE}, the 'switch' is off if:
\itemize{
\item \code{AUTO_JULIA_INSTALL != "true"};
\item You are on a Linux platform and geospatial packages are currently in use;
\item You are on a Linux platform and \code{.geospatial = TRUE};
}
\item Otherwise, the switch if off if:
\itemize{
\item You are on a Linux platform, \code{JULIA_SESSION = "TRUE"} and \code{.geospatial = TRUE};
}
}

(On Linux, geospatial libraries cannot be used simultaneously in \code{R} and \code{Julia}.)

\code{\link[=patter_run_expensive]{patter_run_expensive()}} is a subsidiary switch for particularly expensive routines:
\itemize{
\item If \code{PATTER_RUN_EXPENSIVE = "TRUE"}, this returns \code{TRUE};
\item Otherwise, \code{FALSE} is returned;
}
}
\examples{
if (patter_run(.julia = TRUE, .geospatial = TRUE)) {
  # Connect to `Julia` session
  julia_connect()
  # Use a geospatial library in `R`
  map <- dat_gebco()
  terra::plot(map)
  # Set the map in `Julia`
  set_map(map)
  # Stop `Julia`
  JuliaSwitch::julia_stop()
}
}
\author{
Edward Lavender
}
