% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_setup.R
\name{patter_run}
\alias{patter_run}
\title{Examples: run code}
\usage{
patter_run(.julia = TRUE, .geospatial = TRUE)
}
\arguments{
\item{.julia}{A \code{logical} variable that defines whether or not the code requires connection to a \code{Julia} session.}

\item{.geospatial}{A \code{logical} variable that defines whether or not \code{R}'s geospatial packages (e.g., \code{terra}) are required.}
}
\value{
The function returns a \code{logical} value.
}
\description{
This function switches code blocks on/off.
}
\details{
\itemize{
\item If \code{.julia = TRUE}, the 'switch' is off if:
\itemize{
\item AUTO_JULIA_INSTALL != "true";
\item You are on a Linux platform and \code{.geospatial = TRUE};
}
\item Otherwise, the switch if off if:
\itemize{
\item You are on a Linux platform, \code{JULIA_SESSION = "TRUE"} and \code{.geospatial = TRUE};
}
}

(On Linux, geospatial libraries cannot be used simultaneously in \code{R} and \code{Julia}.)
}
\examples{
if (patter_run(.julia = TRUE, .geospatial = TRUE)) {
  # Connect to `Julia` session
  julia_connect()
  # Use a geospatial library in `R`
  map <- dat_gebco()
  terra::plot(map)
  # Set the map in `Julia`
  set_map(map)
}
}
\author{
Edward Lavender
}
