% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward.R
\name{pf_backward_*}
\alias{pf_backward_*}
\title{PF: run the backward pass}
\description{
\code{\link{patter}} provides two options for the backward simulation:
\itemize{
\item \code{\link[=pf_backward_killer]{pf_backward_killer()}}
\item \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}
}
}
\seealso{
\itemize{
\item The PF (forward simulation) is implemented by \code{\link[=pf_forward]{pf_forward()}}.
\item PF is supported by:
\itemize{
\item Setup helpers, namely \code{\link[=pf_setup_files]{pf_setup_files()}};
}
\item The backward pass is implemented by \code{\link[=pf_backward_*]{pf_backward_*()}};
\item Movement paths are built from PF outputs via \code{pf_path()} functions:
\itemize{
\item \code{\link[=pf_path]{pf_path()}} reconstructs paths;
\item \code{\link[=pf_path_pivot]{pf_path_pivot()}} supports path reconstruction;
}
\item To reconstruct maps of space use, see:
\itemize{
\item \code{\link[=pf_coords]{pf_coords()}} to extract particle coordinates;
\item \code{\link[=pf_map_pou]{pf_map_pou()}} for probability-of-use maps;
\item \code{\link[=pf_map_dens]{pf_map_dens()}} for smooth utilisation distributions;
\item \code{\link[=get_hr]{get_hr()}} for home range estimates;
}
}
}
