% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward_sampler-internal.R
\name{pf_backward_internals}
\alias{pf_backward_internals}
\alias{.pf_backward_dens_mem}
\alias{.pf_backward_dens_calc}
\alias{.pf_backward_dens_nest}
\alias{.pf_backward_dens_spark}
\title{PF: Internal functions for backwards pass}
\usage{
.pf_backward_dens_mem(.history, .dens_step, ..., .verbose, .txt)

.pf_backward_dens_calc(.pairs, .dens_step, ..., .slice_now = FALSE)

.pf_backward_dens_nest(.pairs)

.pf_backward_dens_spark(
  .history,
  .dens_step,
  ...,
  .collect = 1e+09,
  .store = NULL,
  .cl,
  .varlist,
  .verbose,
  .txt
)
}
\description{
These functions support the backwards pass.
}
\author{
Edward Lavender
}
\keyword{internal}
