% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward_killer.R
\name{pf_backward_killer_diagnostics}
\alias{pf_backward_killer_diagnostics}
\title{PF: backward killer diagnostics}
\usage{
pf_backward_killer_diagnostics(.history, ...)
}
\arguments{
\item{.history}{Particle samples, provided in any format accepted by \code{\link[=.pf_history_dt]{.pf_history_dt()}}.}
}
\value{
The function returns a \code{\link{data.table}} with the following columns:
\itemize{
\item \code{timestep}---an \code{integer} that defines the time step;
\item \code{n}---an \code{integer} that defines the number of particles;
\item \code{n_u}---an \code{integer} that defines the number of unique location samples (see \code{\link[=.pf_diag_unique]{.pf_diag_unique()}});
\item \code{ess}---a \code{double} that defines the effective sample size (see \code{\link[=.pf_diag_ess]{.pf_diag_ess()}});
}
}
\description{
This function calculates diagnostics from particle samples from \code{\link[=pf_backward_killer]{pf_backward_killer()}}.
}
\details{
Particle diagnostics are fully described in \code{\link{pf_diag}}.
}
\seealso{
\itemize{
\item \code{\link[=pf_forward]{pf_forward()}}, \code{\link[=pf_backward_killer]{pf_backward_killer()}} and \code{\link[=pf_backward_sampler]{pf_backward_sampler()}} implement the forward simulation and the backward pass;
\item \code{\link[=pf_forward_diagnostics]{pf_forward_diagnostics()}} and \code{\link[=pf_backward_killer_diagnostics]{pf_backward_killer_diagnostics()}} collect diagnostics;
\item \code{\link{.pf_diag}}\verb{_()} functions are internal routines that calculate diagnostic statistics;
}
}
\author{
Edward Lavender
}
