% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-internals.R
\name{pf_check}
\alias{pf_check}
\alias{.pf_check_obs}
\alias{.pf_check_write_history}
\title{PF helper: internal checks}
\usage{
.pf_check_obs(.o)

.pf_check_write_history(.w)
}
\description{
These are internal check functions.
}
\author{
Edward Lavender
}
\keyword{internal}
