% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-internal-utils.R
\name{pf_check}
\alias{pf_check}
\alias{.pf_check_obs}
\title{PF helper: internal check functions}
\usage{
.pf_check_obs(.obs)
}
\description{
These are internal check functions for \code{\link[=pf_forward]{pf_forward()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
