% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-generic-internals.R
\name{pf_check}
\alias{pf_check}
\alias{.pf_check_obs}
\alias{.pf_check_write_history}
\alias{.pf_check_rows}
\alias{.pf_path_pivot_checks}
\title{PF helper: internal checks}
\usage{
.pf_check_obs(.obs)

.pf_check_write_history(.write_history, .element = "sink")

.pf_check_rows(.data, .filter, .t)

.pf_path_pivot_checks(.obs, .cols)
}
\description{
These are internal check functions.
}
\author{
Edward Lavender
}
\keyword{internal}
