% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map.R
\name{pf_coord}
\alias{pf_coord}
\title{PF: particle coordinates}
\usage{
pf_coord(.history, .bathy, .obs = NULL, .cols = NULL)
}
\arguments{
\item{.history}{Particle samples, provided in any format accepted by \code{\link[=.pf_history_dt]{.pf_history_dt()}}. Particle samples may be sourced from:
\itemize{
\item \code{\link[=pf_forward]{pf_forward()}} (a marginal distribution);
\item \code{\link[=pf_backward_killer]{pf_backward_killer()}} (a 'partial' joint distribution);
\item \code{\link[=pf_backward_sampler]{pf_backward_sampler()}} (the full joint distribution);
}

Particle samples must contain the \code{timestep} and \code{cell_now} columns.}

\item{.bathy}{The bathymetry \code{\link{SpatRaster}}.}

\item{.obs, .cols}{(optional) A \code{\link{data.table}} and a character vector of column names in \code{.obs} to match onto the output. \code{.obs} must contain a \code{timestep} column for matching.}
}
\value{
The function returns a \code{\link{data.table}} that defines time steps, sampled locations and, optionally, information from \code{.obs}, with the following columns:
\itemize{
\item \code{timestep}---an \code{integer} vector that defines the time step;
\item \code{cell_id}---an \code{integer} vector that defines the cell ID on \code{.bathy};
\item \code{cell_x}, \code{cell_y}, \code{cell_z}---\code{double}s that define cell coordinates;
}

If \code{.obs} is supplied, the output also contains any columns specified in \code{.cols}.
}
\description{
This function collects particle samples and extracts coordinates.
}
\details{
The \code{cell_now} column is used to extract coordinates on \code{.bathy}.

This function is not memory safe, since the entire time series of coordinates is collected in memory.
}
\examples{
#### Example (1): Use particle samples in memory or on file
# Particles can be provided in any format accepted by `?.pf_history_dt()`
# Here, we use precomputed samples:
gebco <- dat_gebco()
pf_coord(dat_pff(), .bathy = gebco)
pf_coord(dat_pff()$history, .bathy = gebco)
pf_coord(dat_pff_src(), .bathy = gebco)
pf_coord(pf_files(dat_pff_src()), .bathy = gebco)

#### Example (2): Use particle samples from the forward run or the backward pass
pf_coord(dat_pff(), .bathy = gebco)
pf_coord(dat_pfbk(), .bathy = gebco)
# pf_coord(dat_pfbs(), .bathy = gebco)

#### Example (3): Include additional columns in the output
obs <- dat_obs()
pxy <- pf_coord(dat_pff(), .bathy = gebco,
                .obs = obs, .cols = c("detection", "depth"))

#### Example (4): Use particle coordinates for mapping
# See `map_*()` functions e.g. `?map_dens()`
}
\seealso{
The forward filtering--backward sampling algorithm samples locations (particles) that represent the possible locations of an individual through time, accounting for all data and the individual's movement.
\itemize{
\item To set up data, use \code{\link[=pat_setup_data]{pat_setup_data()}}.
\item \code{\link[=pf_forward]{pf_forward()}} implements the forward filter:
\itemize{
\item To set up an observations timeline, use \code{\link[=pf_setup_obs]{pf_setup_obs()}}.
\item For proposal (movement) models, see \code{\link{pf_propose}}.
\item For likelihood functions, to evaluate the likelihood of the data at proposal locations, see \code{\link{pf_lik}}.
\item For sampling functions, to (re)sample plausible proposal locations, see \code{\link{pf_sample}}.
\item For tuning parameters, see \code{\link{pf_opt}}.
}
\item \code{\link[=pf_backward_*]{pf_backward_*()}} refines outputs from the forward filter:
\itemize{
\item \code{\link[=pf_backward_killer]{pf_backward_killer()}} removes dead-ends;
\item \code{\link[=pf_backward_sampler]{pf_backward_sampler()}} implements the backward sampler;
}
\item To reconstruct movement paths from particle samples, use \code{\link[=pf_path]{pf_path()}}.
\item To map emergent patterns of space use, use \code{\link[=pf_coord]{pf_coord()}} plus a \verb{map_*()} function, such as \code{\link[=map_pou]{map_pou()}}, \code{\link[=map_dens]{map_dens()}} and/or \code{\link{map_hr}}\verb{_()}.
\item For additional utilities, see supporting \verb{pf_*()} functions, such as \code{\link[=pf_plot_history]{pf_plot_history()}}, \code{\link[=pf_files]{pf_files()}} and \code{\link[=pf_files_size]{pf_files_size()}}.
}
}
\author{
Edward Lavender
}
