% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map.R
\name{pf_coord}
\alias{pf_coord}
\title{PF: particle coordinates}
\usage{
pf_coord(.history, .bathy, .obs = NULL, .cols = NULL)
}
\arguments{
\item{.history}{Particle samples, provided in any format accepted by \code{\link[=.pf_history_dt]{.pf_history_dt()}}.}

\item{.bathy}{The bathymetry \code{\link{SpatRaster}}.}

\item{.obs, .cols}{(optional) A \code{\link{data.table}} and a character vector of column names in \code{.obs} to match onto the output. \code{.obs} must contain a \code{timestep} column for matching.}
}
\value{
The function returns a \code{\link{data.table}} that defines time steps, sampled locations and, optionally, information from \code{.obs}, with the following columns:
\itemize{
\item \code{timestep}---an \code{integer} vector that defines the time step;
\item \code{cell_id}---an \code{integer} vector that defines the cell ID on \code{.bathy};
\item \code{cell_x}, \code{cell_y}, \code{cell_z}---\code{double}s that define cell coordinates;
}

If \code{.obs} is supplied, the output also contains any columns specified in \code{.cols}.
}
\description{
This function collects particle samples and extracts coordinates.
}
\details{
The \code{timestep} and \code{cell_now} columns in \code{\link[=pf_forward]{pf_forward()}} outputs are required to extract coordinates. \code{cell_now} is used to extract coordinates on \code{.bathy}.

This function is not memory safe, since the entire time series of coordinates is collected in memory.
}
\examples{
p <- dat_pfbk()
pxy_1 <- pf_coord(p$history, dat_gebco())

pfbk_folder <- system.file("extdata", "acpf", "backward", "killer",
                           package = "patter", mustWork = TRUE)
pxy_2 <- pf_coord(pfbk_folder, dat_gebco())

stopifnot(all.equal(pxy_1, pxy_2))

}
\seealso{
The forward filtering--backward sampling algorithm samples locations (particles) that represent the possible locations of an individual through time, accounting for all data and the individual's movement.
\itemize{
\item To set up data, use \code{\link[=pat_setup_data]{pat_setup_data()}}.
\item \code{\link[=pf_forward]{pf_forward()}} implements the forward filter:
\itemize{
\item To set up an observations timeline, use \code{\link[=pf_setup_obs]{pf_setup_obs()}}.
\item For proposal (movement) models, see \code{\link{pf_propose}}.
\item For likelihood functions, to evaluate the likelihood of the data at proposal locations, see \code{\link{pf_lik}}.
\item For sampling functions, to (re)sample plausible proposal locations, see \code{\link{pf_sample}}.
\item For tuning parameters, see \code{\link{pf_opt}}.
}
\item \code{\link[=pf_backward_*]{pf_backward_*()}} refines outputs from the forward filter:
\itemize{
\item \code{\link[=pf_backward_killer]{pf_backward_killer()}} removes dead-ends;
\item \code{\link[=pf_backward_sampler]{pf_backward_sampler()}} implements the backward sampler;
}
\item To reconstruct movement paths from particle samples, use \code{\link[=pf_path]{pf_path()}}.
\item To map emergent patterns of space use, use \code{\link[=pf_coord]{pf_coord()}} plus a \verb{map_*()} function, such as \code{\link[=map_pou]{map_pou()}}, \code{\link[=map_dens]{map_dens()}} and/or \code{\link{map_hr}}\verb{_()}.
\item For additional utilities, see supporting \verb{pf_*()} functions, such as \code{\link[=pf_files]{pf_files()}} and \code{\link[=pf_files_size]{pf_files_size()}}.
}
}
\author{
Edward Lavender
}
