% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map.R
\name{pf_coords}
\alias{pf_coords}
\title{PF: particle coordinates}
\usage{
pf_coords(.history, .bathy, .obs = NULL, .cols = NULL)
}
\arguments{
\item{.history}{Particle samples from the particle filter, provided either as:
\itemize{
\item A \code{list} of \code{\link{data.table}}s that define cell samples; i.e., the \code{history} element of a \code{\linkS4class{pf}} object. This must contain a column that defines accepted cell samples at each time step (\code{cell_now}).
\item An ordered list of file paths (from \code{\link[=pf_setup_files]{pf_setup_files()}}) that define the directories in which particle samples were written from the forward simulation (as parquet files).
}}

\item{.bathy}{The bathymetry \code{\link{SpatRaster}}.}

\item{.obs, .cols}{(optional) A \code{\link{data.table}} and a character vector of column names in \code{.obs} to match onto the output. \code{.obs} must contain a \code{timestep} column for matching.}
}
\value{
The function returns a \code{\link{data.table}} that defines timesteps, sampled locations and, optionally, information from \code{.obs}, with the following columns:
\itemize{
\item \code{timestep}---an \code{integer} vector that defines the time step;
\item \code{cell_id}---an \code{integer} vector that defines the cell ID on \code{.bathy};
\item \code{cell_x}, \code{cell_y}, \code{cell_z}---\code{double}s that define cell coordinates;
}

If \code{.obs} is supplied, the output also contains any columns specified in \code{.cols}.
}
\description{
This function collects particle samples and extracts coordinates.
}
\details{
This function is not memory safe.
}
\examples{
p <- dat_pfb()
pf_coords(p$history, dat_gebco())

}
\author{
Edward Lavender
}
