% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_backward_sampler.R
\name{pf_count}
\alias{pf_count}
\title{PF: count particles}
\usage{
pf_count(
  .history,
  .cl = NULL,
  .cl_varlist = NULL,
  .cl_chunks = TRUE,
  .record = pf_opt_record(),
  .verbose = getOption("patter.verbose")
)
}
\arguments{
\item{.history}{Particle samples, provided in any format accepted by \code{\link[=.pf_history_list]{.pf_history_list()}}. Particle samples must contain the \code{cell_now} column.}

\item{.cl, .cl_varlist, .cl_chunks}{Parallelisation options, passed to \code{\link[=cl_lapply]{cl_lapply()}}.}

\item{.record}{A named \code{list} of output options, from \code{\link[=pf_opt_record]{pf_opt_record()}}.}

\item{.verbose}{User output control (see \code{\link{patter-progress}} for supported options).}
}
\value{
The function returns a \code{\linkS4class{pf_particles}} object.
}
\description{
This function counts particle (cell) samples. Its primary use is to process outputs from forward filter for \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}.
}
\details{
\code{\link[=pf_count]{pf_count()}} iterates over time steps, counts the number of copies of each cell and saves a \code{\link{data.table}} of counts in memory and/or writes them to file for \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}. This is designed to improve speed in \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}.
}
\author{
Edward Lavender
}
