% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-internal-diagnostics.R
\name{pf_diag}
\alias{pf_diag}
\alias{.pf_diag_any}
\alias{.pf_diag_ess}
\alias{.pf_diag_unique}
\alias{.pf_diag}
\alias{.pf_diag_bind}
\alias{.pf_diag_collect}
\title{PF: particle diagnostic internals}
\usage{
.pf_diag_any(.particles)

.pf_diag_ess(.likelihood)

.pf_diag_unique(.cells)

.pf_diag(.particles, .t, .trial = NA_integer_, .label)

.pf_diag_bind(.diag)

.pf_diag_collect(.diagnostics, .iter_m, .iter_i)
}
\description{
These are internal functions that calculate diagnostic statistics from selected particle samples.
}
\details{
\itemize{
\item \code{\link[=.pf_diag_any]{.pf_diag_any()}} identifies whether or not any particle samples remain;
\item \code{\link[=.pf_diag_ess]{.pf_diag_ess()}} calculates effective sample size;
\item \code{\link[=.pf_diag_unique]{.pf_diag_unique()}} counts the number of unique particle samples;
\item \code{\link[=.pf_diag]{.pf_diag()}} is a wrapper function that collates diagnostics;
}
}
\author{
Edward Lavender
}
\keyword{internal}
