% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward_2-diagnostics.R
\name{pf_diag}
\alias{pf_diag}
\alias{.pf_diag_any}
\alias{.pf_diag_ess}
\alias{.pf_diag_unique}
\alias{.pf_diag}
\title{PF: calculate particle diagnostics}
\usage{
.pf_diag_any(.particles)

.pf_diag_ess(.likelihood)

.pf_diag_unique(.cells)

.pf_diag(.particles, .t, .label)
}
\description{
PF: calculate particle diagnostics
}
\keyword{internal}
