% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-helpers.R
\name{pf_files}
\alias{pf_files}
\title{PF: list files}
\usage{
pf_files(.sink, ...)
}
\arguments{
\item{.sink}{A \code{character} string that defines the directory in which files are located.}

\item{...}{Additional arguments passed to \code{\link[=list.files]{list.files()}}, such as \code{pattern}, excluding \code{full.names}.}
}
\value{
The function returns an ordered \code{list} of file paths.
}
\description{
This function creates an ordered \code{list} of the parquet files contains particle samples (e.g., from \code{\link[=pf_forward]{pf_forward()}} or \code{\link[=pf_backward_killer]{pf_backward_killer()}}).
}
\examples{
# List files from pf_forward()
pff_folder <- system.file("extdata", "acpf", "forward", "history",
                          package = "patter", mustWork = TRUE)
files <- pf_files(pff_folder)


# List files from pf_backward_killer()
pfbk_folder <- system.file("extdata", "acpf", "backward", "killer",
                           package = "patter", mustWork = TRUE)
files <- pf_files(pfbk_folder)

}
\author{
Edward Lavender
}
