% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward.R
\name{pf_forward}
\alias{pf_forward}
\title{PF: forward simulation}
\usage{
pf_forward(
  .obs,
  .bathy,
  .origin = .bathy,
  .lonlat = FALSE,
  .moorings = NULL,
  .detection_overlaps = NULL,
  .detection_kernels = NULL,
  .update_ac = NULL,
  .rpropose = pf_rpropose_kick,
  .dpropose = pf_dpropose,
  .n = 100L,
  .sample_origin = pf_sample_multinomial,
  .sample = .sample_origin,
  .trial_origin_crit = 1L,
  .trial_origin = 1L,
  .trial_kick_crit = 1L,
  .trial_kick = 2L,
  .trial_sampler_crit = 2L,
  .trial_sampler = 2L,
  .trial_revert_crit = 2L,
  .trial_revert_steps = 10L,
  .trial_revert = 2L,
  .rerun = list(),
  .rerun_from = pf_setup_rerun(.rerun),
  .record_opts = list(save = FALSE, cols = NULL, sink = NULL),
  .verbose = TRUE
)
}
\description{
PF: forward simulation
}
