% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward.R
\name{pf_forward_*}
\alias{pf_forward_*}
\title{PF: run the forward simulation}
\description{
\code{\link{patter}} provides two options for the forward simulation of possible locations.
\itemize{
\item \strong{Two-step implementation.} In the two-step implementation, an AC-branch algorithm (e.g., \code{\link[=acs]{acs()}}, \code{\link[=dc]{dc()}}) is used to reconstruct the possible locations of an individual over a grid. The resultant surfaces are passed to \code{\link[=pf_forward_1]{pf_forward_1()}}, which uses a particle filtering algorithm to refine the set of possible locations for the individual through time. This is the original implementation in both \href{https://github.com/edwardlavender/flapper}{\code{flapper}} and \code{\link{patter}}. A benefit of this approach is that the deterministic and stochastic parts of the algorithm are separated, which facilitates diagnosing the causes of any convergence failures (e.g., \code{\link[=acs]{acs()}} failures suggest overly restrictive detection probability or mobility parameters whereas \code{\link[=pf_forward_1]{pf_forward_1()}} failures suggest incorrect mobility parameters or insufficient sampling). \code{\link[=pf_forward_1]{pf_forward_1()}} is also fast because pre-calculated weights are used in calculations. However, with high-resolution and/or large grids, this approach may become prohibitively expensive because \code{\link[=acs]{acs()}} and \code{\link[=dc]{dc()}} are implemented over the entire grid, whereas in actuality we only require AC-branch weights at sampled (particle) locations.
\item \strong{Integrated implementation.} In the integrated implementation, AC-branch and PF-branch algorithms are implemented via \code{\link[=pf_forward_2]{pf_forward_2()}}. In this approach, AC-branch calculations are only implemented at particle locations, which is much more efficient. \strong{This is generally the recommended approach.}
}
}
\seealso{
\itemize{
\item \code{\link[=acs]{acs()}} and \code{\link[=dc]{dc()}} implement AC-branch algorithms;
\item \code{\link[=pf_forward_1]{pf_forward_1()}} and \code{\link[=pf_forward_2]{pf_forward_2()}} implement the forward simulation;
\item \code{\link[=pf_backward]{pf_backward()}} implements the backward pass;
\item \code{\link[=pf_path]{pf_path()}} reconstructs movement paths;
\item \code{\link[=pf_pou]{pf_pou()}} and \code{\link[=pf_dens]{pf_dens()}} generate maps of space use;
\item \code{\link[=pf_setup_files]{pf_setup_files()}}, \code{\link[=pf_kick]{pf_kick()}} and \code{\link[=pf_coords]{pf_coords()}} are helper functions;
}
}
