% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward_2-utils-internal.R
\name{pf_forward_2_utils}
\alias{pf_forward_2_utils}
\alias{.pf_utils_dirs}
\alias{.pf_util_startup}
\alias{.pf_util_write}
\alias{pf_util_next}
\title{PF helper: internal utilities}
\usage{
.pf_utils_dirs(.write_opts)

.pf_util_startup(
  .rerun,
  .obs,
  .lonlat,
  .bathy,
  .moorings,
  .detection_overlaps,
  .detection_kernels,
  .update_ac,
  .write_opts
)

.pf_util_write(.particles, .diagnostics, .psink, .dsink, .write)

pf_util_next(.particles)
}
\description{
PF helper: internal utilities
}
\keyword{internal}
