% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward.R
\name{pf_forward_diagnostics}
\alias{pf_forward_diagnostics}
\title{PF: forward run diagnostics}
\usage{
pf_forward_diagnostics(.sink)
}
\arguments{
\item{.sink}{A character string that defines the directory in which \code{\link[=pf_forward]{pf_forward()}} outputs are saved. \code{.sink} should contain the \verb{diagnostics/} directory that is created by \code{\link[=pf_forward]{pf_forward()}}.}
}
\value{
The function returns the diagnostics \code{\link{data.table}}.
}
\description{
This function collates particle diagnostics from \code{\link[=pf_forward]{pf_forward()}}.
}
\details{
Particle diagnostics are fully described in \code{\link{pf_diag}}.
}
\seealso{
\itemize{
\item \code{\link[=pf_forward]{pf_forward()}}, \code{\link[=pf_backward_killer]{pf_backward_killer()}} and \code{\link[=pf_backward_sampler]{pf_backward_sampler()}} implement the forward simulation and the backward pass;
\item \code{\link[=pf_forward_diagnostics]{pf_forward_diagnostics()}} and \code{\link[=pf_backward_killer_diagnostics]{pf_backward_killer_diagnostics()}} collect diagnostics;
\item \code{\link{.pf_diag}}\verb{_()} functions are internal routines that calculate diagnostic statistics;
}
}
\author{
Edward Lavender
}
