% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_kick.R
\name{pf_kick}
\alias{pf_kick}
\title{PF: template movement models}
\usage{
pf_kick(
  .particles,
  .obs = NULL,
  .t = NULL,
  .bathy = NULL,
  .sim_length = sim_length,
  .sim_angle = sim_angle_rw,
  .lonlat = FALSE,
  ...
)
}
\arguments{
\item{.particles}{A \code{\link{data.table}}, from \code{\link[=pf_forward_*]{pf_forward_*()}}, that defines the current particle samples:
\itemize{
\item \code{cell_now} is an integer vector of cell IDs;
\item \code{x_now} is a numerical vector of \code{x} coordinates;
\item \code{y_now} is a numerical vector of \code{y} coordinates;
}}

\item{.obs, .t, .bathy}{(optional) The \code{.obs} \code{\link{data.table}}, an integer that indexes \code{.obs} and the \code{.bathy} \code{\link{SpatRaster}} (see \code{\link[=pf_forward_1]{pf_forward_1()}}). These inputs are unused in this template movement model but supported within \code{\link[=pf_forward_*]{pf_forward_*()}}.}

\item{.sim_length, .sim_angle, ...}{Functions and additional arguments that simulate step lengths and turning angles. The first argument of each function should be the number of step lengths/turning angles to simulate (defined internally).}

\item{.lonlat}{A logical variable that defines whether or not coordinates are longitudes/latitudes.}
}
\value{
The function returns a \code{\link{data.table}}, as supplied, with two additional columns:
\itemize{
\item \code{x_next}---the \code{x} coordinate of the next proposal location(s);
\item \code{y_next}---the \code{y} coordinates of the next proposal location(s);
}
}
\description{
These functions are example movement models, of the kind required by \code{.kick} in \code{\link[=pf_forward_*]{pf_forward_*()}}.
}
\details{
This template movement model is a biased random walk. Step lengths are simulated from a truncated Gamma distribution via \code{\link[=sim_length]{sim_length()}}. Turning angles are simulated from a wrapped normal distribution via \code{\link[=sim_angle_rw]{sim_angle_rw()}}.
}
\section{Warning}{
\itemize{
\item These function are used to streamline examples and do not represent a generically suitable model.
\item The functions do not check user inputs.
}
}

\examples{
# Load packages
require(graphics)
require(data.table)
require(dtplyr)
require(dplyr, warn.conflicts = FALSE)

# Set seed
set.seed(1)

# Define hypothetical `.particles` data.table
p <-
  dat_gebco() |>
  terra::spatSample(size = 10L, cells = TRUE, xy = TRUE, as.df = FALSE,
                    na.rm = TRUE) |>
  as.data.table() |>
  select(cell_now = cell, x_now = x, y_now = y) |>
  as.data.table()

# Kick particles into new locations
p <- pf_kick(p)

# Visualise current & next locations
terra::plot(dat_gebco())
arrows(x0 = p$x_now, x1 = p$x_next,
       y0 = p$y_now, y1 = p$y_next,
       length = 0.02)

}
\author{
Edward Lavender
}
