% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward.R
\name{pf_opt}
\alias{pf_opt}
\alias{pf_opt_trial}
\alias{pf_opt_record}
\alias{pf_opt_control}
\alias{pf_opt_rerun_from}
\title{PF: simulation options}
\usage{
pf_opt_trial(
  .trial_origin_crit = 1L,
  .trial_origin = 1L,
  .trial_kick_crit = 1L,
  .trial_kick = 1L,
  .trial_sampler_crit = 10L,
  .trial_sampler = 1L,
  .trial_revert_crit = 5L,
  .trial_revert_steps = 10L,
  .trial_revert = 10L
)

pf_opt_record(.save = FALSE, .cols = NULL, .sink = NULL)

pf_opt_control(.sampler_batch_size = 2L)

pf_opt_rerun_from(.rerun, .revert = 25L)
}
\arguments{
\item{.trial_origin_crit, .trial_origin, .trial_kick_crit, .trial_kick, .trial_sampler_crit, .trial_sampler, .trial_revert_crit, .trial_revert_steps, .trial_revert}{Trial arguments used to tune convergence properties. All arguments expect \code{integer}s.
\itemize{
\item \verb{.trial_\{step\}} arguments define the number of times to trial a stochastic process at each time step (before giving up).
\item \verb{.trial_\{step\}_crit} arguments define the number of valid, unique proposal locations (grid cells) required to trigger a repeated trial.
}

For example, if after kicking previous particles into proposal locations and evaluating the likelihood of each location only 10 unique particles remain, the stochastic process is repeated up to \code{.trial_kick} times or until the number of valid proposals exceeds \code{.trial_kick_crit} (see Details). In full:
\itemize{
\item \code{.trial_origin_crit} is the critical threshold for the starting samples. If the number of unique, valid starting locations is <= \code{.trial_origin_crit}, sampling is repeated up to \code{.trial_origin} times.
\item \code{trial_kick_crit} is the critical threshold for stochastic kicks. If the number of unique, valid proposals is <= \code{trial_kick_crit}, the process of kicking and sampling particles (via \code{.rpropose} and \code{.sample}) is repeated up to \code{.trial_kick} times. Use \code{.trial_kick = 0L} to suppress stochastic kicks.
\item \code{.trial_sampler_crit} is the critical threshold for directed sampling. Following stochastic kicks, if the number of unique, valid proposals remains <= \code{.trial_sampler_crit}, directed sampling is implemented. Samples are redrawn up to \code{.trial_sampler} times. Use \code{.trial_sampler_crit = 0L} to suppress directed sampling.
\item \code{.trial_revert_crit} is the critical threshold for a reversion. If the number of unique, valid proposal locations is <= \code{.trial_revert_crit}, the algorithm reverts by \code{.trial_revert_steps} time steps to an earlier time step (time step two or greater). \code{.trial_revert} is the total number of reversions permitted. This is reset on algorithm reruns (see \code{.rerun}).
}}

\item{.save, .sink, .cols}{\code{\link[=pf_opt_record]{pf_opt_record()}} arguments, passed to \code{.record} in \code{\link[=pf_forward]{pf_forward()}}.
\itemize{
\item \code{.save}---a \code{logical} variable that defines whether or not to save particle samples and diagnostics in memory. Use with caution.
\item \code{.sink}---a \code{character} string that defines a (usually) empty directory in which to write particle samples and diagnostics (see Value).
\item \code{.cols}---a \code{character} vector that defines the names of the columns in particle-sample \code{\link{data.table}}s to save and/or write to file at each time step. This reduces the space occupied by outputs. At a minimum, you should retain \code{timestep}, \code{cell_now}, \code{x_now} and \code{y_now} for the backward sampler (see \code{\link[=pf_backward_sampler]{pf_backward_sampler()}}). \code{NULL} retains all columns.
}

At least one of \code{.save} and \code{.sink} must be provided.}

\item{.sampler_batch_size}{\code{\link[=pf_opt_control]{pf_opt_control()}} arguments, passed to \code{.control} in \code{\link[=pf_forward]{pf_forward()}}.
\itemize{
\item \code{.sampler_batch_size}---an \code{integer} that controls the batch size (number of particles processed simultaneously) in directed sampling. Increase the batch size to improve speed; decrease the batch size to avoid memory constraints. The appropriate batch size depends on grid resolution and memory availability.
}}

\item{.rerun, .revert}{\code{\link[=pf_opt_rerun_from]{pf_opt_rerun_from()}} arguments.
\itemize{
\item \code{.rerun} is a named \code{list} of algorithm outputs from a previous run.
\item \code{.revert.revert} is an \code{integer} that defines the number of steps to revert.
}}
}
\value{
\itemize{
\item \code{\link[=pf_opt_trial]{pf_opt_trial()}} returns a named \code{list};
\item \code{\link[=pf_opt_record]{pf_opt_record()}} returns a named \code{list};
\item \code{\link[=pf_opt_control]{pf_opt_control()}} returns a named \code{list};
\item \code{\link[=pf_opt_rerun_from]{pf_opt_rerun_from()}} returns an \code{integer};
}
}
\description{
These functions define selected function arguments for \code{\link[=pf_forward]{pf_forward()}}.
}
\details{
These functions are defined separately for convenience of documentation. Note that they do not define global options and must be passed to \code{\link[=pf_forward]{pf_forward()}} arguments.
}
\examples{
pf_opt_trial()
pf_opt_trial(.trial_kick_crit = 10L, .trial_kick = 2L)
pf_opt_record(.save = TRUE)
pf_opt_control()
pf_opt_rerun_from(dat_pff(), .revert = 10L)

}
\author{
Edward lavender
}
