% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-particles.R
\name{pf_particle}
\alias{pf_particle}
\alias{.pf_sample_origin}
\alias{.pf_particles_origin}
\alias{.pf_particles_kick}
\alias{.pf_particles_sampler}
\title{PF: particle helpers}
\usage{
.pf_sample_origin(.particles, .sample, .n, .trial_crit, .trial_count)

.pf_particles_origin(
  .obs,
  .origin,
  .grid = FALSE,
  .detection_kernels,
  .moorings,
  .pf_lik,
  .sample,
  .n,
  .trial_crit,
  .trial_count
)

.pf_particles_kick(
  .particles,
  .rpropose,
  .obs,
  .t,
  .bathy,
  .pf_lik,
  .sample,
  .n,
  .trial_crit,
  .trial_count
)

.pf_particles_sampler(
  .particles,
  .obs,
  .t,
  .bathy,
  .lonlat,
  .pf_lik,
  .dpropose,
  .sample,
  .n,
  .trial_crit,
  .trial_count
)
}
\description{
PF: particle helpers
}
\details{
Hierarchy
pf_propose_
pf_sample_
pf_particles
}
\keyword{internal}
