% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-internal-particles.R
\name{pf_particle}
\alias{pf_particle}
\alias{.pf_rpropose_origin}
\alias{.pf_sample_origin}
\alias{.pf_particles_origin}
\alias{.pf_lik}
\alias{.acs_filter_container}
\alias{.pf_lik_dc}
\alias{.pf_particles_kick}
\alias{.pf_particles_sampler}
\title{PF: particle internals}
\usage{
.pf_rpropose_origin(.particles = NULL, .obs, .t = 1L, .dlist, .rargs = NULL)

.pf_sample_origin(.particles, .n, .sample, .trial_crit)

.pf_particles_origin(
  .particles = NULL,
  .obs,
  .t = 1L,
  .dlist,
  .rpropose = NULL,
  .dpropose = NULL,
  .rargs = NULL,
  .dargs = NULL,
  .likelihood,
  .sample,
  .n,
  .trial_crit,
  .control
)

.pf_lik(.particles, .obs, .t, .dlist, .stack, .control)

.acs_filter_container(.dist, .buffer)

.pf_lik_dc(.x, .a, .b)

.pf_particles_kick(
  .particles,
  .obs,
  .t,
  .dlist,
  .rpropose,
  .rargs = list(),
  .likelihood,
  .control,
  .trial
)

.pf_particles_sampler(
  .particles,
  .obs,
  .t,
  .dlist,
  .dpropose,
  .dargs = list(),
  .likelihood,
  .control
)
}
\description{
These internal functions facilitate proposals, likelihood calculations and sampling in \code{\link[=pf_forward]{pf_forward()}}.
}
\details{
Function hierarchy is as follows:
\itemize{
\item \verb{pf_rpropose_*()} functions generate proposals;
\item \verb{pf_lik_*()} functions calculate likelihoods;
\item \verb{pf_sample_*()} functions sample proposals;
\item \verb{pf_particle_*()} functions combine proposals, likelihood calculations and, if applicable, sampling;
}

Proposal and likelihood functions must accept the following core arguments:
\itemize{
\item \code{.particles}
\item \code{.obs}
\item \code{.t}
\item \code{.dlist}
\item \code{.drop} (except for \code{rpropose} functions);
}

Proposal functions must also accept:
\itemize{
\item \code{.rargs}
\item \code{.dargs}
}

\code{\link[=.pf_lik]{.pf_lik()}} is a wrapper for specified likelihood functions (see \code{\link{pf_lik}}). The function returns a \code{\link{data.table}} that defines proposal locations, log-likelihoods and log-weights. This also requires:
\itemize{
\item \code{.stack}---a named \code{list} of likelihood functions (see \code{\link{pf_lik}}).
\item \code{.control}---a named \code{list} of control options, from \code{\link[=pf_opt_control]{pf_opt_control()}}.
}

Most likelihood functions are directly exported but some internals routines are also documented here for convenience. See also \verb{acs_*()} functions.

Sampling functions must accept:
\itemize{
\item \code{.particles}
\item \code{.n}
}

Wrapper functions must accept:
\itemize{
\item Proposal function arguments (\code{.particles}, \code{.obs}, \code{.t}, \code{.dlist});
\item Proposal functions (\code{.rpropose},\code{.dpropose}) and argument lists (\code{.rargs}, \code{.dargs});
\item Additional likelihood arguments (\code{.likelihood});
\item Sampling arguments (\code{.sample}, \code{.n});
\item Iteration arguments, if applicable (\code{.trial_crit});
\item Control arguments (\code{.control});
}

Internal functions may support additional arguments as required.

At the first time step:
\itemize{
\item \code{\link[=.pf_rpropose_origin]{.pf_rpropose_origin()}} 'proposes' starting locations (quadrature points);
\item \code{\link[=.pf_sample_origin]{.pf_sample_origin()}} samples starting locations;
\item \code{\link[=.pf_lik]{.pf_lik()}} calculates likelihoods;
\item \code{\link[=.pf_particles_origin]{.pf_particles_origin()}} integrates \code{\link[=.pf_rpropose_origin]{.pf_rpropose_origin()}}, \code{\link[=.pf_lik]{.pf_lik()}} and \code{\link[=.pf_particles_origin]{.pf_particles_origin()}};
}

At subsequent time steps:
\itemize{
\item The exported functions \code{\link[=pf_rpropose_kick]{pf_rpropose_kick()}} and \code{\link{pf_rpropose_reachable}} propose locations;
\item \code{\link[=.pf_lik]{.pf_lik()}} calculates likelihoods;
\item \code{\link[=.pf_particles_kick]{.pf_particles_kick()}} and \code{\link[=.pf_particles_sampler]{.pf_particles_sampler()}} are the integration functions;
}

\verb{pf_particles_*()} functions are similar internally, but currently separated for convenience.
}
\author{
Edward Lavender
}
\keyword{internal}
