% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_path.R
\name{pf_path}
\alias{pf_path}
\title{PF: path reconstruction}
\usage{
pf_path(
  .history,
  .bathy = NULL,
  .obs = NULL,
  .cols = NULL,
  .verbose = TRUE,
  .return = c("long", "wide")
)
}
\arguments{
\item{.history}{Particle samples from the particle filter, provided either as:
\itemize{
\item A \code{list} of \code{\link{data.table}}s that define cell samples; i.e., the \code{history} element of a \code{\linkS4class{pf}} object. This must contain columns that define cell samples at each time step (\code{cell_now}) alongside previous samples (\code{cell_past}).
\item An ordered list of file paths (from \code{\link[=pf_setup_files]{pf_setup_files()}}) that define the directories in which particle samples were written from the forward simulation (as parquet files).
}}

\item{.bathy}{(optional) If \code{.return = "long"}, a bathymetry \code{\link{SpatRaster}} can be supplied to define cell coordinates (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).}

\item{.obs, .cols}{(optional) If \code{.return = "long"}, \code{.obs} and \code{.cols} are a \code{\link{data.table}} and a \code{character} vector of column names in \code{.obs} to match onto the output (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).}

\item{.verbose}{Arguments to monitor function progress (see \code{\link[=pf_forward]{pf_forward()}}).}

\item{.return}{A \code{character} that defines the return format:
\itemize{
\item \code{long} specifies a long-format \code{\link{data.table}} that defines path IDs, time steps and associated locations (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).
\item \code{wide} specifies a wide-format \code{\link{data.table}}, with:
\itemize{
\item one row for each path;
\item one column for each time step (named \code{x1}, \code{x2}, etc.);
}
}}
}
\value{
The function returns a long- or wide-format \code{\link{data.table}} (see \code{.return}).
}
\description{
This function implements the path-reconstruction algorithm.
}
\details{
The path reconstruction algorithm 'chains' sequential particle samples into movement paths. This function evolved from \href{https://edwardlavender.github.io/flapper/reference/pf_simplify.html}{\code{pf_simplify()}} in the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package. This implementation uses the fast \code{\link[collapse:join]{collapse::join()}} function.
}
\seealso{
TO DO
}
\author{
Edward Lavender
}
