% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_path.R
\name{pf_path}
\alias{pf_path}
\title{PF: path reconstruction}
\usage{
pf_path(
  .history,
  .bathy = NULL,
  .obs = NULL,
  .cols = NULL,
  .return = c("long", "wide"),
  .verbose = getOption("patter.verbose")
)
}
\arguments{
\item{.history}{Particle samples from the particle filter, provided either as:
\itemize{
\item A \code{list} of \code{\link{data.table}}s that define particle samples; i.e., the \code{history} element of a \code{\linkS4class{pf_particles}} object. This must contain columns that define sampled cells at each time step (\code{cell_now}) alongside previous samples (\code{cell_past}).
\item An ordered \code{list} of file paths (from \code{\link[=pf_files]{pf_files()}}) that define the directories in which particle samples are located (as parquet files).
}}

\item{.bathy}{(optional) If \code{.return = "long"}, a bathymetry \code{\link{SpatRaster}} can be supplied to define cell coordinates (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).}

\item{.obs, .cols}{(optional) If \code{.return = "long"}, \code{.obs} and \code{.cols} are a \code{\link{data.table}} and a \code{character} vector of column names in \code{.obs} to match onto the output (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).}

\item{.return}{A \code{character} that defines the return format:
\itemize{
\item \code{long} specifies a long-format \code{\link{data.table}} that defines path IDs, time steps and associated locations (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).
\item \code{wide} specifies a wide-format \code{\link{data.table}}, with:
\itemize{
\item one row for each path;
\item one column for each time step (named \code{x1}, \code{x2}, etc.);
}
}}

\item{.verbose}{User output control (see \code{\link{patter-progress}} for supported options)}
}
\value{
The function returns a long- or wide-format \code{\link{data.table}} (see \code{.return}).
}
\description{
This function implements the path-reconstruction algorithm.
}
\details{
The path reconstruction algorithm 'chains' sequential particle samples into movement paths. This function evolved from \href{https://edwardlavender.github.io/flapper/reference/pf_simplify.html}{\code{pf_simplify()}} in the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package. This implementation uses the fast \code{\link[collapse:join]{collapse::join()}} function.
}
\examples{
#### Set up examples using pre-defined datasets
gebco       <- dat_gebco()
obs         <- dat_obs()
out_pfbk    <- dat_pfbk()
pfbk_folder <- system.file("extdata", "acpf", "backward", "killer",
                           package = "patter", mustWork = TRUE)

#### Example (1): Implement pf_path() from `pf` object
p1 <- pf_path(out_pfbk$history)

#### Example (2): Implement pf_path from parquet files
p2 <- pf_path(pf_files(pfbk_folder))
stopifnot(all.equal(p1, p2))

#### Example (3): Include cell coordinates/obs columns
p3 <- pf_path(out_pfbk$history, gebco, obs, .cols = "depth")
head(p3)

#### Example (4): Control messages
# Suppress messages
p4 <- pf_path(out_pfbk$history, .verbose = FALSE)
# Write messages to log
log.txt <- tempfile(fileext = ".txt")
p4 <- pf_path(out_pfbk$history, .verbose = log.txt)
readLines(log.txt)

#### Example (5): Examine outputs
# Load packages
require(data.table)
require(dtplyr)
require(dplyr, warn.conflicts = FALSE)
# Compute (Euclidean) distances between sequential samples
gebco <- dat_gebco()
p5 <-
  p3 |>
  group_by(path_id) |>
  mutate(
    dist = terra::distance(cbind(cell_x, cell_y),
                           lonlat = FALSE, sequential = TRUE))
max(p5$dist, na.rm = TRUE)
# Visualise example path
terra::plot(gebco)
path_1 <- p5[p5$path_id == 1, ]
s <- seq_len(nrow(path_1))
graphics::arrows(x0 = path_1$cell_x[s], x1 = path_1$cell_x[s + 1],
                 y0 = path_1$cell_y[s], y1 = path_1$cell_y[s + 1],
                 length = 0.02)

#### Example (6): Change output format
p6 <- pf_path(out_pfbk$history, .return = "wide")
str(p6)
}
\seealso{
TO DO
}
\author{
Edward Lavender
}
