% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_path.R
\name{pf_path}
\alias{pf_path}
\title{PF: path reconstruction}
\usage{
pf_path(
  .history,
  .bathy = NULL,
  .obs = NULL,
  .cols = NULL,
  .verbose = TRUE,
  .txt = "",
  .return = c("long", "wide")
)
}
\arguments{
\item{.history}{Particle samples from the particle filter, provided either as:
\itemize{
\item A \code{list} of \code{\link{data.table}}s that define cell samples; i.e., the \code{history} element of a \code{\linkS4class{pf}} object. This must contain columns that define cell samples at each time step (\code{cell_now}) alongside previous samples (\code{cell_past}).
\item An ordered list of file paths (from \code{\link[=pf_setup_files]{pf_setup_files()}}) that define the directories in which particle samples were written from the forward simulation (as parquet files).
}}

\item{.bathy}{(optional) If \code{.return = "long"}, a bathymetry \code{\link{SpatRaster}} can be supplied to define cell coordinates (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).}

\item{.obs, .cols}{(optional) If \code{.return = "long"}, \code{.obs} and \code{.cols} are a \code{\link{data.table}} and a \code{character} vector of column names in \code{.obs} to match onto the output (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).}

\item{.verbose, .txt}{Arguments to monitor function progress (see \code{\link[=pf_forward_1]{pf_forward_1()}}).}

\item{.return}{A \code{character} that defines the return format:
\itemize{
\item \code{long} specifies a long-format \code{\link{data.table}} that defines path IDs, time steps and associated locations (see \code{\link[=pf_path_pivot]{pf_path_pivot()}}).
\item \code{wide} specifies a wide-format \code{\link{data.table}}, with:
\itemize{
\item one row for each path;
\item one column for each time step (named \code{x1}, \code{x2}, etc.);
}
}}
}
\value{
The function returns a long- or wide-format \code{\link{data.table}} (see \code{.return}).
}
\description{
This function implements the path-reconstruction algorithm.
}
\details{
The path reconstruction algorithm 'chains' sequential particle samples into movement paths. This function evolved from \href{https://edwardlavender.github.io/flapper/reference/pf_simplify.html}{\code{pf_simplify()}} in the \href{https://github.com/edwardlavender/flapper}{\code{flapper}} package. This implementation uses the fast \code{\link[collapse:join]{collapse::join()}} function.
}
\examples{
#### Workflow
# A. Define input datasets (see ?`acs_setup_obs()`)
# B. Implement forward simulation (see ?`acs()`, ?`dc()` and ?`pf_forward_*()`)
# C. Implement backwards pass (see ?`pf_backward()`)
# D. Implement `pf_path()`, `pf_map_pou()` etc.

#### Set up examples using pre-defined datasets
# Define required datasets (e.g., observations, bathymetry grid)
obs        <- dat_obs()
gebco      <- dat_gebco()
# Collect pre-defined particle samples from forward simulation
dat_pff    <- dat_pff()
# Implement backwards pass
con        <- tempdir()
pfb_folder <- file.path(con, "patter", "pf", "backward")
dir.create(pfb_folder, recursive = TRUE)
out_pfb <- pf_backward(dat_pff$history,
                       .save_history = TRUE,
                       .write_history = list(sink = pfb_folder))

#### Example (1): Implement pf_path() from `pf` object
p1 <- pf_path(out_pfb$history)

#### Example (2): Implement pf_path from parquet files
p2 <- pf_path(pf_setup_files(pfb_folder))
stopifnot(all.equal(p1, p2))

#### Example (3): Include cell coordinates/obs columns
p3 <- pf_path(out_pfb$history, gebco, obs, .cols = "depth")
head(p3)

#### Example (4): Control messages
# Suppress messages
p4 <- pf_path(out_pfb$history, .verbose = FALSE)
# Write messages to log
log.txt <- tempfile(fileext = ".txt")
p4 <- pf_path(out_pfb$history, .txt = log.txt)
readLines(log.txt)

#### Example (5): Examine outputs
# Load packages
require(data.table)
require(dtplyr)
require(dplyr, warn.conflicts = FALSE)
# Compute (Euclidean) distances between sequential samples
gebco <- dat_gebco()
p5 <-
  p3 |>
  group_by(path_id) |>
  mutate(
    dist = terra::distance(cbind(cell_x, cell_y),
                           lonlat = FALSE, sequential = TRUE))
max(p5$dist, na.rm = TRUE)
# Visualise example path
terra::plot(gebco)
path_1 <- p5[p5$path_id == 1, ]
s <- seq_len(nrow(path_1))
graphics::arrows(x0 = path_1$cell_x[s], x1 = path_1$cell_x[s + 1],
                 y0 = path_1$cell_y[s], y1 = path_1$cell_y[s + 1],
                 length = 0.02)

#### Example (6): Change output format
p6 <- pf_path(out_pfb$history, .return = "wide")
str(p6)
}
\seealso{
TO DO
}
\author{
Edward Lavender
}
