% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-propose.R
\name{pf_propose}
\alias{pf_propose}
\alias{pf_rpropose_kick}
\alias{pf_rpropose_reachable}
\alias{pf_dpropose}
\title{PF: proposal functions}
\usage{
pf_rpropose_kick(
  .particles,
  .obs,
  .t,
  .dlist,
  .sim_length = rlen,
  .sim_angle = rangrw,
  ...
)

pf_rpropose_reachable(.particles, .obs, .t, .dlist)

pf_dpropose(.particles, .obs, .t, .dlist, ...)
}
\description{
TO DO
}
