% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-propose.R
\name{pf_propose}
\alias{pf_propose}
\alias{pf_rpropose_origin}
\alias{pf_rpropose_kick}
\alias{pf_rpropose_reachable}
\alias{pf_dpropose}
\title{PF: proposal functions}
\usage{
pf_rpropose_origin(.obs, .origin, .grid = FALSE, .detection_kernels, .moorings)

pf_rpropose_kick(
  .particles,
  .obs = NULL,
  .t = NULL,
  .bathy = NULL,
  .sim_length = rlen,
  .sim_angle = rangrw,
  .lonlat = FALSE,
  ...
)

pf_rpropose_reachable(.particles, .obs, .t, .bathy, ...)

pf_dpropose(.particles, .lonlat)
}
\description{
TO DO
}
