% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-sample.R
\name{pf_sample}
\alias{pf_sample}
\alias{pf_sample_multinomial}
\alias{pf_sample_systematic}
\title{PF: sampling functions}
\usage{
pf_sample_multinomial(.particles, .n)

pf_sample_systematic(.particles, .n)
}
\arguments{
\item{.particles}{A \code{\link{data.table}} that defines particle samples. This includes a \code{weight} column that should be used for (re)sampling.}

\item{.n}{An \code{integer} that defines the number of particle samples at each time step.}
}
\value{
The functions return a \code{\link{data.table}}, as inputted, but containing \code{.n} rows, one for each sampled particle.
}
\description{
These are particle sampling functions for \code{\link[=pf_forward]{pf_forward()}}.
}
\details{
\code{\link[=pf_forward]{pf_forward()}} expects a sampling function in the \code{.sample} argument. You can write your own function or use an exported one:
\itemize{
\item \code{\link{pf_sample_multinomial}} implements multinomial (re)sampling.
\item \code{\link[=pf_sample_systematic]{pf_sample_systematic()}} implements systematic (re)sampling.
}

Sampling functions must accept \code{.particles} and \code{.n} arguments and return a \code{\link{data.table}} of \code{.n} particle samples.

Systematic particle sampling (\code{\link[=pf_sample_systematic]{pf_sample_systematic()}}) is generally recommended.
}
\examples{
require(data.table)
p <- data.table(cell_now = c(1, 2, 3), weight = c(0.5, 0.25, 0.25))
pf_sample_multinomial(p, .n = 10L)
pf_sample_systematic(p, .n = 10L)

}
\author{
Edward Lavender
}
