% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-sample.R
\name{pf_sample}
\alias{pf_sample}
\alias{pf_sample_multinomial}
\alias{pf_sample_systematic}
\title{PF: sampling functions}
\usage{
pf_sample_multinomial(.particles, .n)

pf_sample_systematic(.particles, .n)
}
\arguments{
\item{.particles}{A \code{\link{data.table}} that defines particle samples. This includes a \code{weight} column that should be used for (re)sampling.}

\item{.n}{An \code{integer} that defines the number of particle samples at each time step.}
}
\value{
The functions return a \code{\link{data.table}}, as inputted, but containing \code{.n} rows, one for each sampled particle.
}
\description{
These are particle sampling functions for \code{\link[=pf_forward]{pf_forward()}}.
}
\details{
\code{\link[=pf_forward]{pf_forward()}} expects a sampling function in the \code{.sample} argument. You can write your own function or use an exported one:
\itemize{
\item \code{\link{pf_sample_multinomial}} implements multinomial (re)sampling.
\item \code{\link[=pf_sample_systematic]{pf_sample_systematic()}} implements systematic (re)sampling.
}

Sampling functions must accept \code{.particles} and \code{.n} arguments and return a \code{\link{data.table}} of \code{.n} particle samples.

Systematic particle sampling (\code{\link[=pf_sample_systematic]{pf_sample_systematic()}}) is generally recommended.
}
\examples{
require(data.table)
p <- data.table(cell_now = c(1, 2, 3), weight = c(0.5, 0.25, 0.25))
pf_sample_multinomial(p, .n = 10L)
pf_sample_systematic(p, .n = 10L)

}
\seealso{
The forward filtering--backward sampling algorithm samples locations (particles) that represent the possible locations of an individual through time, accounting for all data and the individual's movement.
\itemize{
\item To set up data, use \code{\link[=pat_setup_data]{pat_setup_data()}}.
\item \code{\link[=pf_forward]{pf_forward()}} implements the forward filter:
\itemize{
\item To set up an observations timeline, use \code{\link[=pf_setup_obs]{pf_setup_obs()}}.
\item For proposal (movement) models, see \code{\link{pf_propose}}.
\item For likelihood functions, to evaluate the likelihood of the data at proposal locations, see \code{\link{pf_lik}}.
\item For sampling functions, to (re)sample plausible proposal locations, see \code{\link{pf_sample}}.
\item For tuning parameters, see \code{\link{pf_opt}}.
}
\item \code{\link[=pf_backward_*]{pf_backward_*()}} refines outputs from the forward filter:
\itemize{
\item \code{\link[=pf_backward_killer]{pf_backward_killer()}} removes dead-ends;
\item \code{\link[=pf_backward_sampler]{pf_backward_sampler()}} implements the backward sampler;
}
\item To reconstruct movement paths from particle samples, use \code{\link[=pf_path]{pf_path()}}.
\item To map emergent patterns of space use, use \code{\link[=pf_coord]{pf_coord()}} plus a \verb{map_*()} function, such as \code{\link[=map_pou]{map_pou()}}, \code{\link[=map_dens]{map_dens()}} and/or \code{\link{map_hr}}\verb{_()}.
\item For additional utilities, see supporting \verb{pf_*()} functions, such as \code{\link[=pf_files]{pf_files()}}, \code{\link[=pf_files_size]{pf_files_size()}}, \code{\link[=pf_plot_history]{pf_plot_history()}}, \code{\link[=pf_diag_convergence]{pf_diag_convergence()}} and \code{\link[=pf_diag_summary]{pf_diag_summary()}}.
}
}
\author{
Edward Lavender
}
