% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_forward-sample.R
\name{pf_sample}
\alias{pf_sample}
\alias{pf_sample_multinomial}
\alias{pf_sample_systematic}
\title{PF: sampling functions}
\usage{
pf_sample_multinomial(.particles, .n)

pf_sample_systematic(.particles, .n)
}
\arguments{
\item{.particles}{A \code{\link{data.table}} that defines particle samples. This includes a \code{logwt} column that should be used for (re)sampling. Weights should be normalised.}

\item{.n}{An \code{integer} that defines the number of particle samples at each time step.}
}
\value{
The functions return a \code{\link{data.table}}, as inputted, but containing \code{.n} rows, one for each sampled particle.
}
\description{
These are particle sampling functions for \code{\link[=pf_forward]{pf_forward()}}.
}
\details{
\code{\link[=pf_forward]{pf_forward()}} expects a sampling function in the \code{.sample} argument. You can write your own function or use an exported one:
\itemize{
\item \code{\link{pf_sample_multinomial}} implements multinomial (re)sampling.
\item \code{\link[=pf_sample_systematic]{pf_sample_systematic()}} implements systematic (re)sampling.
}

Sampling functions must accept \code{.particles} and \code{.n} arguments and return a \code{\link{data.table}} of \code{.n} particle samples. The \code{logwt} column must be reset after sampling.

Systematic particle sampling (\code{\link[=pf_sample_systematic]{pf_sample_systematic()}}) is generally recommended.
}
\examples{
require(data.table)
p <- data.table(cell_now = c(1, 2, 3), logwt = log(c(0.5, 0.25, 0.25)))
pf_sample_multinomial(p, .n = 10L)
pf_sample_systematic(p, .n = 10L)

}
\seealso{
The particle filter samples locations (particles) that represent the possible locations of an individual through time, accounting for the data and the individual's movement.
\itemize{
\item To set up data, use \code{\link[=pat_setup_data]{pat_setup_data()}}.
\item To implement the particle filter, use \code{\link[=pf_forward]{pf_forward()}}:
\itemize{
\item To set up an observations timeline, use \code{\link[=pf_setup_obs]{pf_setup_obs()}};
\item For proposal (movement) models, see \code{\link{pf_propose}};
\item For likelihood functions, to evaluate the likelihood of the data at proposal locations, see \code{\link{pf_lik}};
\item For sampling functions, to (re)sample plausible proposal locations, see \code{\link{pf_sample}};
\item For tuning parameters, see \code{\link{pf_opt}};
}
\item To map patterns of space use, use:
\itemize{
\item The backward smoother (TO DO);
\item \code{\link[=pf_coord]{pf_coord()}} to extract coordinates;
\item A \verb{map_*()} function, such as \code{\link[=map_pou]{map_pou()}}, \code{\link[=map_dens]{map_dens()}} and/or \code{\link{map_hr}}\verb{_()};
}
\item To reconstruct movement paths, use the backward sampler (TO DO).
\item For additional utilities, see supporting \verb{pf_*()} functions, such as \code{\link[=pf_files]{pf_files()}}, \code{\link[=pf_files_size]{pf_files_size()}}, \code{\link[=pf_plot_history]{pf_plot_history()}} and \code{\link[=pf_diag_summary]{pf_diag_summary()}}.
}
}
\author{
Edward Lavender
}
