% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-setup.R
\name{pf_setup_files}
\alias{pf_setup_files}
\title{PF set up: list files for PF}
\usage{
pf_setup_files(.root, ...)
}
\arguments{
\item{.root}{A string that defines the directory in which files are located.}

\item{...}{Additional arguments passed to \code{\link[=list.files]{list.files()}}, such as \code{pattern}, excluding \code{full.names}.}
}
\value{
The function returns an ordered \code{list} of file paths.
}
\description{
This function creates an ordered \code{list} of files for \code{\link[=pf_backward_*]{pf_backward_*()}}.
}
\examples{
# Quick implementation of pf_forward()
pff_folder <- file.path(con, "patter", "pf", "forward")
dir.create(pff_folder, recursive = TRUE)
out_pff <- pf_forward(.obs = dat_obs(),
                      .bathy = dat_gebco(),
                      .moorings = dat_moorings, .detection_overlaps = dat_overlaps(),
                      .detection_kernels = dat_kernels(),
                      .write_opts = list(sink = pff_folder))

# List files for pf_backward_*()
files <- pf_setup_files(file.path(pff_folder, "history"))

# Quick implementation of pf_backward_killer()
pfb_folder <- file.path(con, "patter", "pf", "backward")
dir.create(pfb_folder, recursive = TRUE)
out_pfb <- pf_backward_killer(files,
                              .write_history = list(sink = pfb_folder))

# List files from pf_backward_*()
pf_setup_files(pfb_folder)

# Clean up
unlink(file.path(con, "patter"), recursive = TRUE)

}
\seealso{
TO DO
}
\author{
Edward Lavender
}
