% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle-smoother.R
\name{pf_smoother_two_filter}
\alias{pf_smoother_two_filter}
\title{Two-filter smoother}
\usage{
pf_smoother_two_filter(
  .map = NULL,
  .mobility = .map,
  .n_particle = NULL,
  .n_sim = 100L,
  .verbose = getOption("patter.verbose")
)
}
\arguments{
\item{.map, .mobility}{(optional) 'Mobility box' arguments.}

\item{.verbose}{.verbose User output control (see \code{\link{patter-progress}} for supported options).}

\item{.nMC}{An \code{integer} that defines the number of Monte Carlo simulations.}
}
\description{
This function implements the two-filter smoother.
}
\details{
This function wraps \href{https://edwardlavender.github.io/Patter.jl/}{\code{Patter.two_filter_smoother()}}.
}
\seealso{
Particle filters and smoothers sample states (particles) that represent the possible locations of an individual through time, accounting for all data and the individual's movement.
\itemize{
\item To simulate artificial datasets, see \verb{sim_*()} functions (especially \code{\link[=sim_path_walk]{sim_path_walk()}}, \code{\link[=sim_array]{sim_array()}} and \code{\link[=sim_observations]{sim_observations()}}).
\item To assemble real-world datasets for the filter, see \code{\link{assemble}}\verb{_*()} functions.
\item \code{\link[=pf_filter]{pf_filter()}} runs the filter:
\itemize{
\item For state types, see \code{\link{State}};
\item For observation models, see \code{\link{ModelObs}};
\item For movement models, see \code{\link{ModelMove}};
}
\item To run particle smoothing, use \code{\link[=pf_smoother_two_filter]{pf_smoother_two_filter()}}.
\item To map emergent patterns of space use, use a \verb{map_*()} function (such as \code{\link[=map_pou]{map_pou()}}, \code{\link[=map_dens]{map_dens()}} and \code{\link[=map_hr]{map_hr()}}).
}
}
\author{
Edward Lavender
}
