% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle-smoother.R
\name{pf_smoother_two_filter}
\alias{pf_smoother_two_filter}
\title{Two-filter smoother}
\usage{
pf_smoother_two_filter(
  .map = NULL,
  .mobility = .map,
  .nMC = 100L,
  .verbose = getOption("patter.verbose")
)
}
\arguments{
\item{.map, .mobility}{(optional) 'Mobility box' arguments.}

\item{.nMC}{An \code{integer} that defines the number of Monte Carlo simulations.}

\item{.verbose}{.verbose User output control (see \code{\link{patter-progress}} for supported options).}
}
\description{
This function implements the two-filter smoother.
}
\details{
This function wraps \href{https://edwardlavender.github.io/Patter.jl/}{\code{Patter.two_filter_smoother()}}.
}
\author{
Edward Lavender
}
