% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{prettyGraphics}
\alias{prettyGraphics}
\alias{one_page}
\alias{par_mf}
\alias{add_sp_path}
\title{\code{prettyGraphics} helpers in \code{\link{patter}}}
\usage{
one_page(.one_page, .n)

par_mf(.n)

add_sp_path(x, y = NULL, ...)
}
\description{
\code{prettyGraphics} helpers in \code{\link{patter}}
}
\examples{
\dontrun{

#### Example (1): one_page() implementation
f <- function() {
  pp <- one_page(TRUE, 2)
  on.exit(graphics::par(pp), add = TRUE)
  hist(1:10, main = "1")
  hist(1:10, main = "2")
  }
f()
plot(1)
}
}
\keyword{internal}
