% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-internal.R
\name{raster.vol}
\alias{raster.vol}
\title{Spatial helper: raster volume}
\source{
This function is copied, almost exactly, from the \href{https://github.com/jeffreyevans/spatialEco}{\code{spatialEco}} package (v.2.0-2). It is defined separately in \code{\link{patter}} to reduce dependencies.
}
\usage{
raster.vol(
  x,
  p = 0.75,
  sample = FALSE,
  spct = 0.05,
  type = c("random", "regular")
)
}
\description{
The \code{spatialEco::raster.vol()} function.
}
\keyword{internal}
