% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_array}
\alias{sim_array}
\title{Simulate an acoustic array}
\usage{
sim_array(
  .bathy = rast_template(),
  .lonlat = FALSE,
  .arrangement = "random",
  .n_receiver = 10L,
  ...,
  .receiver_start = NULL,
  .receiver_end = NULL,
  .receiver_range = NULL,
  .n_array = 1L,
  .plot = TRUE,
  .one_page = FALSE
)
}
\arguments{
\item{.bathy}{A \code{\link{SpatRaster}} that defines the region of interest.}

\item{.lonlat}{A \code{logical} variable that defines whether or not coordinates on \code{.bathy} are longitudes/latitudes. This input controls the output name of the coordinates column (see Value).}

\item{.arrangement}{A character that defines the receiver arrangement (passed to the \code{method} argument of \code{\link[terra:sample]{terra::spatSample()}}).}

\item{.n_receiver}{An \code{integer} that defines the number of receivers (passed to the \code{size} argument of \code{\link[terra:sample]{terra::spatSample()}}).}

\item{...}{Additional arguments passed to \code{\link[terra:sample]{terra::spatSample()}}.}

\item{.receiver_start, .receiver_end, .receiver_range}{(optional) Additional columns to include in the output:
\itemize{
\item \code{.receiver_start} and \code{.receiver_end} are \code{Date} or \code{POSIXct} inputs that specify the deployment time period;
\item \code{.receiver_range} is a \code{numeric} input that defines the detection range;
}

Single inputs are expected to these arguments, which are constant across all receivers.}

\item{.n_array}{An \code{integer} that defines the number of array designs to simulate.}

\item{.plot}{A \code{logical} variable that defines whether or not to plot simulated arrays.}

\item{.one_page}{If \code{.plot = TRUE}, \code{.one_page} is a \code{logical} variable that defines whether or not to produce plots on a single page.}
}
\value{
The function returns a \code{data.table} with the following columns:
\itemize{
\item \code{array_id}---an \code{integer} vector of array IDs,
\item \code{receiver_id}---an \code{integer} vector of receiver IDs;
\item \code{receiver_easting} (if \code{.lonlat = FALSE}) or \code{receiver_lon} (if \code{.lonlat = TRUE})---a \code{numeric} vector that defines receiver x coordinates;
\item \code{receiver_northing} (if \code{.lonlat = FALSE}) or \code{receiver_lat} (if \code{.lonlat = TRUE})---a \code{numeric} vector that defines receiver y coordinates;
\item \code{receiver_start}---receiver start dates (if defined);
\item \code{receiver_end}---receiver end dates (if defined);
\item \code{receiver_range}---receiver detection ranges (if defined);
}
}
\description{
This function simulates acoustic receivers on a grid.
}
\details{
Receiver locations are simulated using \code{\link[terra:sample]{terra::spatSample()}}.
}
\examples{
#### Example (1): The default implementation
# The function returns a data.table
a <- sim_array()
a

#### Example (2): Customise receiver placement/number
a <- sim_array(.arrangement = "regular", .n_receiver = 100)

#### Example (3): Add additional columns for downstream functions
a <- sim_array(.receiver_start = as.Date("2013-01-01"),
               .receiver_end = as.Date("2014-01-01"),
               .receiver_range = 750)

#### Example (4): Control the plot(s)
sim_array(.plot = FALSE)
sim_array(.n_array = 5L, .plot = TRUE, .one_page = TRUE)
sim_array(.n_array = 5L, .plot = TRUE, .one_page = FALSE)

}
\seealso{
\itemize{
\item \code{\link[=sim_array]{sim_array()}} simulates acoustic array(s);
\item \code{\link[=sim_path_walk]{sim_path_walk()}} simulates movement path(s) via a walk model;
\item [\code{sim_detections()} simulates detection(s) at receivers;
\item \code{\link{skill}} functions compared simulated and reconstructed patterns to evaluate model skill;
}
}
\author{
Edward Lavender
}
