% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_array}
\alias{sim_array}
\title{Simulation: acoustic arrays}
\usage{
sim_array(
  .bathy = spatTemplate(),
  .lonlat = FALSE,
  .arrangement = "random",
  .n_receiver = 10L,
  ...,
  .receiver_start = NULL,
  .receiver_end = NULL,
  .receiver_range = NULL,
  .n_array = 1L,
  .plot = TRUE,
  .one_page = FALSE
)
}
\arguments{
\item{.bathy}{A \code{\link{SpatRaster}} that defines the region of interest. Receivers are not simulated in \code{NA} regions.}

\item{.lonlat}{A \code{logical} variable that defines whether or not coordinates on \code{.bathy} are in longitudes/latitude or planar format. This input controls the output name of the coordinate columns (see Value).}

\item{.arrangement, .n_receiver, ...}{Arguments passed to \code{\link[terra:sample]{terra::spatSample()}}.
\itemize{
\item \code{.arrangement} is a \code{character} that defines the receiver arrangement (passed to the \code{method} argument).
\item \code{.n_receiver} is an \code{integer} that defines the number of receivers to simulate (passed to the \code{size} argument).
\item \code{...} ... Additional arguments, passed to \code{\link[terra:sample]{terra::spatSample()}}, excluding: \code{x}, \code{size}, \code{method}, \code{replace}, \code{na.rm}, \code{xy}, \code{cells} and \code{values}.
}}

\item{.receiver_start, .receiver_end, .receiver_range}{(optional) Additional columns to include in the output:
\itemize{
\item \code{.receiver_start} and \code{.receiver_end} are \code{Date} or \code{POSIXct} inputs that specify the deployment time period;
\item \code{.receiver_range} is a \code{numeric} input that defines the detection range;
}

Single inputs are expected to these arguments, which are constant across all receivers.}

\item{.n_array}{An \code{integer} that defines the number of array designs to simulate with the aforementioned parameters.}

\item{.plot}{A \code{logical} variable that defines whether or not to plot simulated arrays.}

\item{.one_page}{If \code{.plot = TRUE}, \code{.one_page} is a \code{logical} variable that defines whether or not to produce plots on a single page.}
}
\value{
The function returns a \code{data.table} with the following columns:
\itemize{
\item \code{array_id}---an \code{integer} vector of array IDs,
\item \code{receiver_id}---an \code{integer} vector of receiver IDs;
\item \code{receiver_easting} (if \code{.lonlat = FALSE}) or \code{receiver_lon} (if \code{.lonlat = TRUE})---a \code{numeric} vector that defines receiver x coordinates;
\item \code{receiver_northing} (if \code{.lonlat = FALSE}) or \code{receiver_lat} (if \code{.lonlat = TRUE})---a \code{numeric} vector that defines receiver y coordinates;
\item \code{receiver_start}---receiver start dates (if defined);
\item \code{receiver_end}---receiver end dates (if defined);
\item \code{receiver_range}---receiver detection ranges (if defined);
}
}
\description{
This function simulates acoustic arrays (i.e., networks of acoustic receiver(s)) on a grid.
}
\details{
Receiver locations are simulated using \code{\link[terra:sample]{terra::spatSample()}}.

This function replaces \href{https://edwardlavender.github.io/flapper/reference/sim_array.html}{\code{flapper::sim_array()}}.
}
\examples{
#### Example (1): The default implementation
# The function returns a data.table
a <- sim_array()
a

#### Example (2): Customise receiver placement/number
a <- sim_array(.arrangement = "regular", .n_receiver = 100)

#### Example (3): Add additional columns for downstream functions
a <- sim_array(.receiver_start = as.Date("2013-01-01"),
               .receiver_end = as.Date("2014-01-01"),
               .receiver_range = 750)

#### Example (4): Control the plot(s)
sim_array(.plot = FALSE)
sim_array(.n_array = 5L, .plot = TRUE, .one_page = TRUE)
sim_array(.n_array = 5L, .plot = TRUE, .one_page = FALSE)

}
\seealso{
\itemize{
\item \verb{sim_*} functions implement \emph{de novo} simulation of movements and observations:
\itemize{
\item \code{\link{sim_helpers}} are convenience functions for simulations;
\item \code{\link[=sim_array]{sim_array()}} simulates acoustic array(s);
\item \code{\link[=sim_path_walk]{sim_path_walk()}} simulates movement path(s) via a walk model;
\item \code{\link[=sim_detections]{sim_detections()}} simulates detection(s) at receivers;
}
\item \code{\link[=pf_forward]{pf_forward()}} implements forward simulation-based reconstruction of movement paths;
\item \code{\link[=pf_backward_*]{pf_backward_*()}} implements backward simulation-based reconstruction of movement paths;
\item \code{\link{skill}} functions compared simulated and reconstructed patterns to evaluate model skill;
}
}
\author{
Edward Lavender
}
