% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_array}
\alias{sim_array}
\title{Simulate an acoustic array}
\usage{
sim_array(
  .bathy = rast_template(),
  .arrangement = "random",
  .n_receivers = 10L,
  ...,
  .receiver_start = NULL,
  .receiver_end = NULL,
  .receiver_range = NULL,
  .n_arrays = 1L,
  .plot = TRUE,
  .one_page = TRUE
)
}
\arguments{
\item{.bathy}{A \code{\link{SpatRaster}} that defines the region of interest.}

\item{.arrangement}{A character that defines the receiver arrangement (passed to the \code{method} argument of \code{\link[terra:sample]{terra::spatSample()}}).}

\item{.n_receivers}{An \code{integer} that defines the number of receivers (passed to the \code{size} argument of \code{\link[terra:sample]{terra::spatSample()}}).}

\item{...}{Additional arguments passed to \code{\link[terra:sample]{terra::spatSample()}}.}

\item{.receiver_start, .receiver_end, .receiver_range}{(optional) Additional columns to include in the output:
\itemize{
\item \code{.receiver_start} and \code{.receiver_end} specify the deployment time period;
\item \code{.receiver_range} defines the detection range;
}

Single inputs are expected to these arguments, which are constant across all receivers.}

\item{.n_arrays}{An \code{integer} that defines the number of array designs to simulate.}

\item{.plot}{A \code{logical} variable that defines whether or not to plot simulated arrays.}

\item{.one_page}{If \code{.plot = TRUE}, \code{.one_page} is a \code{logical} variable that defines whether or not to produce plots on a single page (see \code{\link[=par_mf]{par_mf()}}).}
}
\value{
The function returns a \code{list}. Each element of the list is a \code{data.table} with the following columns:
\itemize{
\item \code{receiver_id}---an \code{integer} vector of receiver IDs;
\item \code{receiver_easting}---a \code{numeric} vector that defines receiver x coordinates;
\item \code{receiver_northing}---a \code{numeric} vector that defines receiver y coordinates;
\item \code{receiver_start}---receiver start dates (if defined);
\item \code{receiver_end}---receiver end dates (if defined);
\item \code{receiver_range}---receiver detection ranges (if defined);
}
}
\description{
This function simulates acoustic receivers on a grid.
}
\details{
Receiver locations are simulated using \code{\link[terra:sample]{terra::spatSample()}}.
}
\examples{
#### Example (1): The default implementation
# The function returns a list, with one element for each simulated array
a <- sim_array()
summary(a)

#### Example (2): Customise receiver placement/number
a <- sim_array(.arrangement = "regular", .n_receivers = 100)

#### Example (3): Add additional columns for downstream functions
a <- sim_array(.receiver_start = as.Date("2013-01-01"),
               .receiver_end = as.Date("2014-01-01"),
               .receiver_range = 500)

#### Example (4): Control the plot(s)
sim_array(.plot = FALSE)
sim_array(.n_arrays = 5L, .plot = TRUE, .one_page = TRUE)
sim_array(.n_arrays = 5L, .plot = TRUE, .one_page = FALSE)

}
\author{
Edward Lavender
}
