% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-internal.R
\name{sim_detections_internals}
\alias{sim_detections_internals}
\alias{.sim_detections}
\alias{.sim_detections_call}
\title{Simulate detections}
\usage{
.sim_detections(
  .path,
  .array,
  .calc_distance = terra::distance,
  .lonlat = FALSE,
  .calc_detection_pr = calc_detection_pr,
  ...,
  .sim_obs = stats::rbinom,
  .return = c("array_id", "path_id", "timestep", "timestamp", "receiver_id", "dist",
    "pr")
)

.sim_detections_call(.path, .array, .args)
}
\description{
Simulate detections
}
\keyword{internal}
