% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_helpers}
\alias{sim_helpers}
\alias{rtruncgamma}
\alias{dtruncgamma}
\alias{rwn}
\alias{rlen}
\alias{rangrw}
\alias{rangcrw}
\alias{cstep}
\alias{dstep}
\title{Simulation helpers}
\usage{
rtruncgamma(.n = 1, .shape = 15, .scale = 15, .mobility = 500, ...)

dtruncgamma(.x = 1, .shape = 15, .scale = 15, .mobility = 500, ...)

rwn(.n = 1, .mu = 0, .rho = 0, .sd = 1, ...)

rlen(.n = 1, .prior = NULL, .t = NULL, ...)

rangrw(.n = 1, .prior = NULL, .t = NULL, ...)

rangcrw(.n = 1, .prior = NULL, .t = NULL, ...)

cstep(
  .xy_now,
  .xy_next = matrix(NA, nrow = nrow(.xy_now), ncol = 2L),
  .lonlat = FALSE,
  .length = rtruncgamma(nrow(.xy_now)),
  .angle = rwn(nrow(.xy_now))
)

dstep(.data_now, .data_past, ...)
}
\arguments{
\item{.n, .x, .shape, .scale, .mobility}{Arguments for \code{\link[=rtruncgamma]{rtruncgamma()}} and \code{\link[=dtruncgamma]{dtruncgamma()}}:
\itemize{
\item \code{.n} is an \code{integer} that defines the number of simulated outcome(s);
\item \code{.x} is a \code{numeric} vector that defines step length(s);
\item \code{.shape} is a \code{numeric} value that defines the shape parameter of a Gamma distribution (see \code{\link[stats:GammaDist]{stats::rgamma()}}).
\item \code{.scale} is a \code{numeric} value that defines the scale parameter of a Gamma distribution (see \code{\link[stats:GammaDist]{stats::rgamma()}}).
\item \code{.mobility} is a \code{numeric} value that defines the maximum length (see \code{\link[truncdist:rtrunc]{truncdist::rtrunc()}}).
}}

\item{...}{Arguments passed to/from functions.}

\item{.mu, .rho, .sd}{Arguments for \code{\link[=rwn]{rwn()}} for the simulation of turning angles, passed to the \code{mu}, \code{rho} and \code{sd} arguments of \code{\link[circular:wrappednormal]{circular::rwrappednormal()}}.}

\item{.prior, .t}{Arguments for \code{\link[=rlen]{rlen()}}, \code{\link[=rangrw]{rangrw()}} and \code{\link[=rangcrw]{rangcrw()}}, as used in top-level functions (e.g, \code{\link[=sim_path_walk]{sim_path_walk()}}):
\itemize{
\item \code{.n}---an \code{integer} that defines the number of simulated outcome(s);
\item \code{.prior}---a \code{numeric} vector that defines the simulated value(s) from the previous time step;
\item \code{.t}---an \code{integer} that defines the time step;
\item \code{...}---additional arguments, if needed;
}}

\item{.xy_now, .xy_next, .lonlat, .length, .angle}{Arguments for \code{\link[=cstep]{cstep()}}.
\itemize{
\item \code{.xy_now} is a two-column matrix of x and y coordinates;
\item (optional) \code{.xy_next} is a blank two-column matrix that is updated with future coordinates;
\item \code{.lonlat} is a \code{logical} variable that defines whether or not coordinates are in longitude/latitude;
\item \code{.length} is \code{numeric} vector of step lengths;
\item \code{.angle} is a \code{numeric} vector of turning angles;
}}

\item{.data_now, .data_past}{Arguments for \code{\link[=dstep]{dstep()}}.
\itemize{
\item \code{.data_now} is a \code{\link{data.table}} with coordinates in \code{x_now} and \code{y_now} columns;
\item \code{.data_past} is a \code{\link{data.table}} as above;
}}
}
\description{
These convenience functions support the generation of animal movement paths and observations in \emph{de novo} simulations (\verb{sim_*()} functions) and simulation-based reconstructions of movement paths (\code{\link[=pf_forward_*]{pf_forward_*()}} and \code{\link[=pf_backward_p]{pf_backward_p()}}).
\itemize{
\item \verb{r*()} functions simulate random variates;
\item \verb{c*()} functions calculate outcomes from random-variate inputs;
\item \verb{d*()} functions return densities;
}

\verb{r*()} and \verb{c*()} functions are used in \emph{de novo} simulations (via \verb{sim_*()} functions such as \code{\link[=sim_path_walk]{sim_path_walk()}}) and the forward simulation-based reconstruction of movement paths (via \code{\link[=pf_kick]{pf_kick()}} in \code{\link[=pf_forward_*]{pf_forward_*()}}).

\verb{d*()} functions are used in the simulation-based reconstruction of movement paths as part of the backward sampler via \code{\link[=pf_backward_p]{pf_backward_p()}}.
}
\section{Movement}{
\subsection{Core functions}{
\subsection{Step lengths}{
\itemize{
\item \code{\link[=rtruncgamma]{rtruncgamma()}} simulates step length(s) from a truncated gamma distribution with a \code{.mobility} parameter.
\item \code{\link[=dtruncgamma]{dtruncgamma()}} returns the densit(ies) of specified step length(s).
}
}

\subsection{Turning angles}{
\itemize{
\item \code{\link[=rwn]{rwn()}} simulates turning angle(s) from a wrapped normal distribution.
\item \code{dwn()} is not currently implemented.
}
}

}

\subsection{Wrappers}{

The following wrapper functions are provided in the form required by front-end functions (e.g., \code{\link[=sim_path_walk]{sim_path_walk()}}, \code{\link[=pf_kick]{pf_kick()}} and \code{\link[=pf_backward_p]{pf_backward_p()}}):
\itemize{
\item \code{\link[=rlen]{rlen()}} is a wrapper for \code{\link[=rtruncgamma]{rtruncgamma()}}. The corresponding function \code{dlen} is not currently implemented.
\item \code{\link[=rangrw]{rangrw()}}, \code{\link[=rangcrw]{rangcrw()}} are wrappers for \code{\link[=rwn]{rwn()}} for random walks and correlated random walks. The corresponding functions \code{dangrw()} and \code{dangcrw()} are not currently implemented.
}
}

\subsection{Extensions}{
\itemize{
\item \code{\link[=cstep]{cstep()}} steps from previous location(s) into future location(s), given simulated step length(s) and turning angle(s).
\item \code{\link[=dstep]{dstep()}} returns the density of steps between two locations. This is only suitable for random walks (it wraps \code{\link[=dtruncgamma]{dtruncgamma()}} but corresponding functions for turning angles are not yet implemented).
}
}
}

\seealso{
\itemize{
\item \verb{sim_*} functions implement \emph{de novo} simulation of movements and observations;
\item \code{\link[=pf_forward_*]{pf_forward_*()}} implements forward simulation-based reconstruction of movement paths;
\item \code{\link[=pf_backward]{pf_backward()}} implements backward simulation-based reconstruction of movement paths;
}
}
\author{
Edward Lavender
}
