% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill.R
\name{skill}
\alias{skill}
\alias{skill_mb}
\alias{skill_me}
\alias{skill_rmse}
\alias{skill_R}
\alias{skill_d}
\title{Skill: evaluation metrics}
\usage{
skill_mb(.obs, .mod, .summarise = "mean")

skill_me(.obs, .mod, .summarise = "mean")

skill_rmse(.obs, .mod)

skill_R(.obs, .mod)

skill_d(.obs, .mod)
}
\arguments{
\item{.obs}{The 'observed' (true) pattern of space use, as a \code{\link{SpatRaster}}.}

\item{.mod}{The 'modelled' (reconstructed) pattern of space use, as a \code{\link{SpatRaster}}.}

\item{.summarise}{A function, passed to \code{\link[terra:global]{terra::global()}}, used to summarise \code{\link{SpatRaster}} values.}
}
\value{
The functions return a number.
}
\description{
These function are designed to support comparisons of simulated and reconstructed patterns of space use.
}
\details{
\itemize{
\item \code{\link[=skill_mb]{skill_mb()}} computes mean bias.
\item \code{\link[=skill_me]{skill_me()}} computes mean error.
\item \code{\link[=skill_rmse]{skill_rmse()}} computes root mean squared error.
\item \code{\link[=skill_R]{skill_R()}} computes Spearman's rank correlation coefficient.
\item \code{\link[=skill_d]{skill_d()}} computes the index of agreement.
}
}
\examples{
# Define template SpatRasters
mod <- obs <- spatTemplate(.res = 1)
n <- terra::ncell(mod)
mod[] <- runif(n)
obs[] <- mod[] + rnorm(n)

# Calculate skill metrics
skill_mb(mod, obs)
skill_me(mod, obs)
skill_rmse(mod, obs)
skill_R(mod, obs)
skill_d(mod, obs)

}
\author{
Edward Lavender
}
