% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill.R
\name{skill}
\alias{skill}
\alias{skill_mb}
\alias{skill_me}
\alias{skill_rmse}
\alias{skill_R}
\alias{skill_d}
\title{Skill: evaluation metrics}
\usage{
skill_mb(.obs, .mod, .summarise = "mean")

skill_me(.obs, .mod, .summarise = "mean")

skill_rmse(.obs, .mod)

skill_R(.obs, .mod)

skill_d(.obs, .mod)
}
\arguments{
\item{.obs}{The 'observed' (true) pattern of space use, as a \code{\link{SpatRaster}}.}

\item{.mod}{The 'modelled' (reconstructed) pattern of space use, as a \code{\link{SpatRaster}}.}

\item{.summarise}{A function, passed to \code{\link[terra:global]{terra::global()}}, used to summarise \code{\link{SpatRaster}} values.}
}
\value{
The functions return a number.
}
\description{
These functions are standard model skill metrics. In \code{\link{patter}}, they support comparisons of simulated and reconstructed patterns of space use.
}
\details{
We follow the mathematical definitions in Lavender et al. (\href{https://www.doi.org/1038/s41598-022-20254-z}{2022}) Supplementary Information Sect. \href{https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41598-022-20254-z/MediaObjects/41598_2022_20254_MOESM2_ESM.pdf}{3.2.1}.
\itemize{
\item \code{\link[=skill_mb]{skill_mb()}} computes mean bias (if \code{.summarise = "mean"}).
\item \code{\link[=skill_me]{skill_me()}} computes mean error (if \code{.summarise = "mean"}).
\item \code{\link[=skill_rmse]{skill_rmse()}} computes root mean squared error.
\item \code{\link[=skill_R]{skill_R()}} computes Spearman's rank correlation coefficient.
\item \code{\link[=skill_d]{skill_d()}} computes the index of agreement.
}

These functions are not memory safe.
}
\examples{
# Define template SpatRasters
mod <- obs <- terra::rast()
n <- terra::ncell(mod)
mod[] <- runif(n)
obs[] <- mod[] + rnorm(n)

# Normalise template SpatRasters
mod   <- mod / terra::global(mod, "sum")[1, 1]
obs   <- obs / terra::global(obs, "sum")[1, 1]

# Visualise 'observed' versus 'modelled' distributions
pp <- par(mfrow = c(1, 2))
terra::plot(mod)
terra::plot(obs)
par(pp)

# Calculate skill metrics
skill_mb(mod, obs)
skill_me(mod, obs)
skill_rmse(mod, obs)
skill_R(mod, obs)
skill_d(mod, obs)
}
\references{
Lavender, E. et al. (2022). Benthic animal-borne sensors and citizen science combine to validate ocean modelling. Sci. Rep. 12: 16613. \url{https://www.doi.org/1038/s41598-022-20254-z}
}
\seealso{
\itemize{
\item To simulate observations, see \verb{sim_*()} functions (especially \code{\link[=sim_path_walk]{sim_path_walk()}}, \code{\link[=sim_array]{sim_array()}} and \code{\link[=sim_observations]{sim_observations()}});
\item To translate observations into coordinates for mapping patterns of space use, see:
\itemize{
\item \code{\link[=coa]{coa()}} to calculate centres of activity;
\item \code{\link[=pf_filter]{pf_filter()}} and associates to implement particle filtering algorithms;
}
\item To estimate utilisation distributions from simulated data and algorithm outputs, use \verb{map_*()} functions (see \code{\link[=map_pou]{map_pou()}}, \code{\link[=map_dens]{map_dens()}} and \code{\link[=map_hr]{map_hr()}});
}
}
\author{
Edward Lavender
}
