% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers-internal.R
\name{spat}
\alias{spat}
\alias{spatContainsNA}
\alias{spatCellCoordsDT}
\alias{spatIntersect}
\alias{spatIntersect.SpatRaster}
\alias{spatIntersect.SpatVector}
\alias{spatIsEmpty}
\title{Spatial helper: \verb{spat*} functions}
\usage{
spatContainsNA(.x)

spatCellCoordsDT(.x, .spatcell = .x)

spatIntersect(.x, .value = 1, .fun = NULL)

spatIntersect.SpatRaster(.x, .value, .fun)

spatIntersect.SpatVector(.x)

spatIsEmpty(.x)
}
\description{
Spatial helper: \verb{spat*} functions
}
\keyword{internal}
