% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers-internal.R
\name{spat}
\alias{spat}
\alias{spatContainsNA}
\alias{spatDT}
\alias{spatIntersect}
\alias{spatIntersect.SpatRaster}
\alias{spatIntersect.SpatVector}
\title{Spatial helper: \verb{spat*} functions}
\usage{
spatContainsNA(.x)

spatDT(.x, .bathy = .x, .weights = FALSE)

spatIntersect(.x, .value = 1, .fun = NULL)

spatIntersect.SpatRaster(.x, .value, .fun)

spatIntersect.SpatVector(.x)
}
\description{
Spatial helper: \verb{spat*} functions
}
\keyword{internal}
