% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spats.R
\name{spat}
\alias{spat}
\alias{spatIntersect}
\alias{spatContainsNA}
\alias{spatAllNA}
\alias{spatVmap}
\title{Spatial helper: \verb{spat*} functions}
\usage{
spatIntersect(.x)

spatContainsNA(.x)

spatAllNA(.x)

spatVmap(.map, .mobility, .plot = FALSE, ...)
}
\description{
Internal helpers for \code{terra} \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} and \code{\link[terra:SpatVector-class]{terra::SpatVector}} objects.
}
\author{
Edward Lavender
}
\keyword{internal}
