% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spats.R
\name{spat}
\alias{spat}
\alias{spatIntersect}
\alias{spatContainsNA}
\alias{spatAllNA}
\alias{spatMobilityBox}
\title{Spatial helper: \verb{spat*} functions}
\usage{
spatIntersect(.x)

spatContainsNA(.x)

spatAllNA(.x)

spatMobilityBox(.x, .mobility)
}
\description{
Internal helpers for \code{terra} \code{\link{SpatRaster}} and \code{\link{SpatVector}} objects.
}
\author{
Edward Lavender
}
\keyword{internal}
