% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers-internal.R
\name{spatial_helpers}
\alias{spatial_helpers}
\alias{.is_lonlat}
\alias{.st_invert}
\alias{cang_planar}
\alias{geoangle}
\alias{geomean}
\title{Spatial helper: internals}
\usage{
.is_lonlat(.data)

.st_invert(.x, .bbox = sf::st_bbox(.x))

cang_planar(.xy0, .xy1, .convention = c("180", "360"))

geoangle(.ang)

geomean(xy, w = NULL)
}
\description{
These functions are internal spatial helpers.
}
\author{
Edward Lavender
}
\keyword{internal}
