% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-helpers-internal.R
\name{spatial_helpers}
\alias{spatial_helpers}
\alias{.is_lonlat}
\alias{.st_invert}
\alias{geomean}
\title{Spatial helper: internals}
\usage{
.is_lonlat(.data)

.st_invert(.x, .bbox = sf::st_bbox(.x))

geomean(xy, w = NULL)
}
\description{
These functions are internal spatial helpers.
}
\author{
Edward Lavender
}
\keyword{internal}
