% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{utils-lists}
\alias{utils-lists}
\alias{list_compact}
\alias{list_merge}
\alias{list_args}
\title{Utilities: list helpers}
\usage{
list_compact(l)

list_merge(...)

list_args(.default = list(), .dots = list(), .user = list())
}
\description{
These are internal list helpers in \code{\link{patter}}.
}
\author{
Edward Lavender
}
\keyword{internal}
